Struct opencv::boxed_ref::BoxedRef

#[repr(transparent)]
pub struct BoxedRef<'r, T: Boxed> {
    pub(crate) reference: T,
    referenced_object: PhantomData<&'r T>,
}

Wrapper for the type implementing Boxed trait that allows to retain the lifetime of the referenced object.

This wrapper implements all traits that the underlying type does, but explicitly doesn’t implement Deref and DerefMut to avoid being able to mem::swap the reference out of the wrapper.

Fields
reference: Treferenced_object: PhantomData<&'r T>
Implementations

impl<T: Boxed + Clone> BoxedRef<'_, T>

pub fn clone_pointee(&self) -> T

Clones the pointee of this BoxedRef