Enum opencv::core::Code

#[repr(C)]
pub enum Code {

    StsOk = 0,
    StsBackTrace = -1,
    StsError = -2,
    StsInternal = -3,
    StsNoMem = -4,
    StsBadArg = -5,
    StsBadFunc = -6,
    StsNoConv = -7,
    StsAutoTrace = -8,
    HeaderIsNull = -9,
    BadImageSize = -10,
    BadOffset = -11,
    BadDataPtr = -12,
    BadStep = -13,
    BadModelOrChSeq = -14,
    BadNumChannels = -15,
    BadNumChannel1U = -16,
    BadDepth = -17,
    BadAlphaChannel = -18,
    BadOrder = -19,
    BadOrigin = -20,
    BadAlign = -21,
    BadCallBack = -22,
    BadTileSize = -23,
    BadCOI = -24,
    BadROISize = -25,
    MaskIsTiled = -26,
    StsNullPtr = -27,
    StsVecLengthErr = -28,
    StsFilterStructContentErr = -29,
    StsKernelStructContentErr = -30,
    StsFilterOffsetErr = -31,
    StsBadSize = -201,
    StsDivByZero = -202,
    StsInplaceNotSupported = -203,
    StsObjectNotFound = -204,
    StsUnmatchedFormats = -205,
    StsBadFlag = -206,
    StsBadPoint = -207,
    StsBadMask = -208,
    StsUnmatchedSizes = -209,
    StsUnsupportedFormat = -210,
    StsOutOfRange = -211,
    StsParseError = -212,
    StsNotImplemented = -213,
    StsBadMemBlock = -214,
    StsAssert = -215,
    GpuNotSupported = -216,
    GpuApiCallError = -217,
    OpenGlNotSupported = -218,
    OpenGlApiCallError = -219,
    OpenCLApiCallError = -220,
    OpenCLDoubleNotSupported = -221,
    OpenCLInitError = -222,
    OpenCLNoAMDBlasFft = -223,

}

error codes

Variants

StsOk = 0

everything is ok

StsBackTrace = -1

pseudo error for back trace

StsError = -2

unknown /unspecified error

StsInternal = -3

internal error (bad state)

StsNoMem = -4

insufficient memory

StsBadArg = -5

function arg/param is bad

StsBadFunc = -6

unsupported function

StsNoConv = -7

iteration didn’t converge

StsAutoTrace = -8

tracing

HeaderIsNull = -9

image header is NULL

BadImageSize = -10

image size is invalid

BadOffset = -11

offset is invalid

BadDataPtr = -12

BadStep = -13

image step is wrong, this may happen for a non-continuous matrix.

BadModelOrChSeq = -14

BadNumChannels = -15

bad number of channels, for example, some functions accept only single channel matrices.

BadNumChannel1U = -16

BadDepth = -17

input image depth is not supported by the function

BadAlphaChannel = -18

BadOrder = -19

number of dimensions is out of range

BadOrigin = -20

incorrect input origin

BadAlign = -21

incorrect input align

BadCallBack = -22

BadTileSize = -23

BadCOI = -24

input COI is not supported

BadROISize = -25

incorrect input roi

MaskIsTiled = -26

StsNullPtr = -27

null pointer

StsVecLengthErr = -28

incorrect vector length

StsFilterStructContentErr = -29

incorrect filter structure content

StsKernelStructContentErr = -30

incorrect transform kernel content

StsFilterOffsetErr = -31

incorrect filter offset value

StsBadSize = -201

the input/output structure size is incorrect

StsDivByZero = -202

division by zero

StsInplaceNotSupported = -203

in-place operation is not supported

StsObjectNotFound = -204

request can’t be completed

StsUnmatchedFormats = -205

formats of input/output arrays differ

StsBadFlag = -206

flag is wrong or not supported

StsBadPoint = -207

bad CvPoint

StsBadMask = -208

bad format of mask (neither 8uC1 nor 8sC1)

StsUnmatchedSizes = -209

sizes of input/output structures do not match

StsUnsupportedFormat = -210

the data format/type is not supported by the function

StsOutOfRange = -211

some of parameters are out of range

StsParseError = -212

invalid syntax/structure of the parsed file

StsNotImplemented = -213

the requested function/feature is not implemented

StsBadMemBlock = -214

an allocated block has been corrupted

StsAssert = -215

assertion failed

GpuNotSupported = -216

no CUDA support

GpuApiCallError = -217

GPU API call error

OpenGlNotSupported = -218

no OpenGL support

OpenGlApiCallError = -219

OpenGL API call error

OpenCLApiCallError = -220

OpenCL API call error

OpenCLDoubleNotSupported = -221

OpenCLInitError = -222

OpenCL initialization error

OpenCLNoAMDBlasFft = -223