Function opencv::core::idft_def

pub fn idft_def(
    src: &impl ToInputArray,
    dst: &mut impl ToOutputArray
) -> Result<()>

Calculates the inverse Discrete Fourier Transform of a 1D or 2D array.

idft(src, dst, flags) is equivalent to dft(src, dst, flags | #DFT_INVERSE) .

Note: None of dft and idft scales the result by default. So, you should pass DFT_SCALE to one of dft or idft explicitly to make these transforms mutually inverse.

See also

dft, dct, idct, mulSpectrums, getOptimalDFTSize

Parameters

src: input floating-point real or complex array.

dst: output array whose size and type depend on the flags.

flags: operation flags (see dft and #DftFlags).

nonzeroRows: number of dst rows to process; the rest of the rows have undefined content (see the convolution sample in dft description.

Note

This alternative version of idft function uses the following default values for its arguments:

flags: 0

nonzero_rows: 0