Function opencv::core::map_gl_buffer

pub fn map_gl_buffer(
    buffer: &impl BufferTraitConst,
    access_flags: AccessFlag,
) -> Result<UMat>

Maps Buffer object to process on CL side (convert to UMat).

Function creates CL buffer from GL one, and then constructs UMat that can be used to process buffer data with OpenCV functions. Note that in current implementation UMat constructed this way doesn’t own corresponding GL buffer object, so it is the user responsibility to close down CL/GL buffers relationships by explicitly calling unmapGLBuffer() function.

Parameters

buffer: - source Buffer object.

accessFlags: - data access flags (ACCESS_READ|ACCESS_WRITE).

Returns

Returns UMat object

C++ default parameters

access_flags: ACCESS_READ|ACCESS_WRITE