opencv::core
Function min_max_loc_def

pub fn min_max_loc_def(
    src: &impl ToInputArray,
    min_val: Option<&mut f64>,
) -> Result<()>

Finds the global minimum and maximum in an array.

The function cv::minMaxLoc finds the minimum and maximum element values and their positions. The extremums are searched across the whole array or, if mask is not an empty array, in the specified array region.

The function do not work with multi-channel arrays. If you need to find minimum or maximum elements across all the channels, use Mat::reshape first to reinterpret the array as single-channel. Or you may extract the particular channel using either extractImageCOI , or mixChannels , or split .

Parameters

src: input single-channel array.

minVal: pointer to the returned minimum value; NULL is used if not required.

maxVal: pointer to the returned maximum value; NULL is used if not required.

minLoc: pointer to the returned minimum location (in 2D case); NULL is used if not required.

maxLoc: pointer to the returned maximum location (in 2D case); NULL is used if not required.

mask: optional mask used to select a sub-array.

See also

max, min, reduceArgMin, reduceArgMax, compare, inRange, extractImageCOI, mixChannels, split, Mat::reshape

Note

This alternative version of min_max_loc function uses the following default values for its arguments:

max_val: 0

min_loc: 0

max_loc: 0

mask: noArray()