Function opencv::core::rand_shuffle

pub fn rand_shuffle(
    dst: &mut impl ToInputOutputArray,
    iter_factor: f64,
    rng: &mut impl RNGTrait
) -> Result<()>

Shuffles the array elements randomly.

The function cv::randShuffle shuffles the specified 1D array by randomly choosing pairs of elements and swapping them. The number of such swap operations will be dst.rows*dst.cols*iterFactor .

Parameters

dst: input/output numerical 1D array.

iterFactor: scale factor that determines the number of random swap operations (see the details below).

rng: optional random number generator used for shuffling; if it is zero, theRNG () is used instead.

See also

RNG, sort

C++ default parameters

iter_factor: 1.

rng: 0