opencv::core
Function randn

pub fn randn(
    dst: &mut impl ToInputOutputArray,
    mean: &impl ToInputArray,
    stddev: &impl ToInputArray,
) -> Result<()>

Fills the array with normally distributed random numbers.

The function cv::randn fills the matrix dst with normally distributed random numbers with the specified mean vector and the standard deviation matrix. The generated random numbers are clipped to fit the value range of the output array data type.

Parameters

dst: output array of random numbers; the array must be pre-allocated and have 1 to 4 channels.

mean: mean value (expectation) of the generated random numbers.

stddev: standard deviation of the generated random numbers; it can be either a vector (in which case a diagonal standard deviation matrix is assumed) or a square matrix.

See also

RNG, randu