Function opencv::core::rotate

pub fn rotate(
    src: &impl ToInputArray,
    dst: &mut impl ToOutputArray,
    rotate_code: i32
) -> Result<()>

Rotates a 2D array in multiples of 90 degrees. The function cv::rotate rotates the array in one of three different ways:

Rotate by 90 degrees clockwise (rotateCode = ROTATE_90_CLOCKWISE).

Rotate by 180 degrees clockwise (rotateCode = ROTATE_180).

Rotate by 270 degrees clockwise (rotateCode = ROTATE_90_COUNTERCLOCKWISE).

Parameters

src: input array.

dst: output array of the same type as src. The size is the same with ROTATE_180, and the rows and cols are switched for ROTATE_90_CLOCKWISE and ROTATE_90_COUNTERCLOCKWISE.

rotateCode: an enum to specify how to rotate the array; see the enum [rotate_flags]

See also

transpose , repeat , completeSymm, flip, RotateFlags