Struct opencv::core::Buffer

pub struct Buffer {
    ptr: *mut c_void,
}

Smart pointer for OpenGL buffer object with reference counting.

Buffer Objects are OpenGL objects that store an array of unformatted memory allocated by the OpenGL context. These can be used to store vertex data, pixel data retrieved from images or the framebuffer, and a variety of other things.

ogl::Buffer has interface similar with Mat interface and represents 2D array memory.

ogl::Buffer supports memory transfers between host and device and also can be mapped to CUDA memory.

Fields
ptr: *mut c_void
Implementations

impl Buffer

pub fn default() -> Result<Buffer>

The constructors.

Creates empty ogl::Buffer object, creates ogl::Buffer object from existed buffer ( abufId parameter), allocates memory for ogl::Buffer object or copies from host/device memory.

pub fn new( arows: i32, acols: i32, atype: i32, abuf_id: u32, auto_release: bool, ) -> Result<Buffer>

The constructors.

Creates empty ogl::Buffer object, creates ogl::Buffer object from existed buffer ( abufId parameter), allocates memory for ogl::Buffer object or copies from host/device memory.

Overloaded parameters

Parameters

arows: Number of rows in a 2D array.

acols: Number of columns in a 2D array.

atype: Array type ( CV_8UC1, …, CV_64FC4 ). See Mat for details.

abufId: Buffer object name.

autoRelease: Auto release mode (if true, release will be called in object’s destructor).

C++ default parameters

auto_release: false

pub fn new_def( arows: i32, acols: i32, atype: i32, abuf_id: u32, ) -> Result<Buffer>

@overload

Parameters

arows: Number of rows in a 2D array.

acols: Number of columns in a 2D array.

atype: Array type ( CV_8UC1, …, CV_64FC4 ). See Mat for details.

abufId: Buffer object name.

autoRelease: Auto release mode (if true, release will be called in object’s destructor).

Note

This alternative version of [new] function uses the following default values for its arguments:

auto_release: false

pub fn new_1( asize: Size, atype: i32, abuf_id: u32, auto_release: bool, ) -> Result<Buffer>

The constructors.

Creates empty ogl::Buffer object, creates ogl::Buffer object from existed buffer ( abufId parameter), allocates memory for ogl::Buffer object or copies from host/device memory.

Overloaded parameters

Parameters

asize: 2D array size.

atype: Array type ( CV_8UC1, …, CV_64FC4 ). See Mat for details.

abufId: Buffer object name.

autoRelease: Auto release mode (if true, release will be called in object’s destructor).

C++ default parameters

auto_release: false

pub fn new_def_1(asize: Size, atype: i32, abuf_id: u32) -> Result<Buffer>

@overload

Parameters

asize: 2D array size.

atype: Array type ( CV_8UC1, …, CV_64FC4 ). See Mat for details.

abufId: Buffer object name.

autoRelease: Auto release mode (if true, release will be called in object’s destructor).

Note

This alternative version of [new] function uses the following default values for its arguments:

auto_release: false

pub fn new_2( arows: i32, acols: i32, atype: i32, target: Buffer_Target, auto_release: bool, ) -> Result<Buffer>

The constructors.

Creates empty ogl::Buffer object, creates ogl::Buffer object from existed buffer ( abufId parameter), allocates memory for ogl::Buffer object or copies from host/device memory.

Overloaded parameters

Parameters

arows: Number of rows in a 2D array.

acols: Number of columns in a 2D array.

atype: Array type ( CV_8UC1, …, CV_64FC4 ). See Mat for details.

target: Buffer usage. See cv::ogl::Buffer::Target .

autoRelease: Auto release mode (if true, release will be called in object’s destructor).

C++ default parameters

target: ARRAY_BUFFER

auto_release: false

pub fn new_def_2(arows: i32, acols: i32, atype: i32) -> Result<Buffer>

@overload

Parameters

arows: Number of rows in a 2D array.

acols: Number of columns in a 2D array.

atype: Array type ( CV_8UC1, …, CV_64FC4 ). See Mat for details.

target: Buffer usage. See cv::ogl::Buffer::Target .

autoRelease: Auto release mode (if true, release will be called in object’s destructor).

Note

This alternative version of [new] function uses the following default values for its arguments:

target: ARRAY_BUFFER

auto_release: false

pub fn new_3( asize: Size, atype: i32, target: Buffer_Target, auto_release: bool, ) -> Result<Buffer>

The constructors.

Creates empty ogl::Buffer object, creates ogl::Buffer object from existed buffer ( abufId parameter), allocates memory for ogl::Buffer object or copies from host/device memory.

Overloaded parameters

Parameters

asize: 2D array size.

atype: Array type ( CV_8UC1, …, CV_64FC4 ). See Mat for details.

target: Buffer usage. See cv::ogl::Buffer::Target .

autoRelease: Auto release mode (if true, release will be called in object’s destructor).

C++ default parameters

target: ARRAY_BUFFER

auto_release: false

pub fn new_def_3(asize: Size, atype: i32) -> Result<Buffer>

@overload

Parameters

asize: 2D array size.

atype: Array type ( CV_8UC1, …, CV_64FC4 ). See Mat for details.

target: Buffer usage. See cv::ogl::Buffer::Target .

autoRelease: Auto release mode (if true, release will be called in object’s destructor).

Note

This alternative version of [new] function uses the following default values for its arguments:

target: ARRAY_BUFFER

auto_release: false

pub fn new_4( arr: &impl ToInputArray, target: Buffer_Target, auto_release: bool, ) -> Result<Buffer>

The constructors.

Creates empty ogl::Buffer object, creates ogl::Buffer object from existed buffer ( abufId parameter), allocates memory for ogl::Buffer object or copies from host/device memory.

Overloaded parameters

Parameters

arr: Input array (host or device memory, it can be Mat , cuda::GpuMat or std::vector ).

target: Buffer usage. See cv::ogl::Buffer::Target .

autoRelease: Auto release mode (if true, release will be called in object’s destructor).

C++ default parameters

target: ARRAY_BUFFER

auto_release: false

pub fn new_def_4(arr: &impl ToInputArray) -> Result<Buffer>

@overload

Parameters

arr: Input array (host or device memory, it can be Mat , cuda::GpuMat or std::vector ).

target: Buffer usage. See cv::ogl::Buffer::Target .

autoRelease: Auto release mode (if true, release will be called in object’s destructor).

Note

This alternative version of [new] function uses the following default values for its arguments:

target: ARRAY_BUFFER

auto_release: false

pub fn unbind(target: Buffer_Target) -> Result<()>

Unbind any buffers from the specified binding point.

Parameters

target: Binding point. See cv::ogl::Buffer::Target .