Struct opencv::core::FileStorage

pub struct FileStorage {
    ptr: *mut c_void,
}

XML/YAML/JSON file storage class that encapsulates all the information necessary for writing or reading data to/from a file.

Fields
ptr: *mut c_void
Implementations

impl FileStorage

pub fn default() -> Result<FileStorage>

The constructors.

The full constructor opens the file. Alternatively you can use the default constructor and then call FileStorage::open.

pub fn new(filename: &str, flags: i32, encoding: &str) -> Result<FileStorage>

The constructors.

The full constructor opens the file. Alternatively you can use the default constructor and then call FileStorage::open.

Overloaded parameters

@copydoc open()

C++ default parameters

encoding: String()

pub fn new_def(filename: &str, flags: i32) -> Result<FileStorage>

@overload @copydoc open()

Note

This alternative version of [new] function uses the following default values for its arguments:

encoding: String()

pub fn get_default_object_name(filename: &str) -> Result<String>

Returns the normalized object name for the specified name of a file.

Parameters

filename: Name of a file

Returns

The normalized object name.