Struct opencv::core::KernelArg

pub struct KernelArg {
    ptr: *mut c_void,
}

Fields
ptr: *mut c_void
Implementations

impl KernelArg

pub unsafe fn new( _flags: i32, _m: &mut impl UMatTrait, wscale: i32, iwscale: i32, _obj: *const c_void, _sz: size_t, ) -> Result<KernelArg>

C++ default parameters

wscale: 1

iwscale: 1

_obj: 0

_sz: 0

pub fn new_def(_flags: i32, _m: &mut impl UMatTrait) -> Result<KernelArg>

Note

This alternative version of [new] function uses the following default values for its arguments:

wscale: 1

iwscale: 1

_obj: 0

_sz: 0

pub fn default() -> KernelArg

pub fn local(local_mem_size: size_t) -> Result<KernelArg>

pub fn ptr_write_only(m: &impl UMatTraitConst) -> Result<KernelArg>

pub fn ptr_read_only(m: &impl UMatTraitConst) -> Result<KernelArg>

pub fn ptr_read_write(m: &impl UMatTraitConst) -> Result<KernelArg>

pub fn read_write( m: &impl UMatTraitConst, wscale: i32, iwscale: i32, ) -> Result<KernelArg>

C++ default parameters

wscale: 1

iwscale: 1

pub fn read_write_def(m: &impl UMatTraitConst) -> Result<KernelArg>

Note

This alternative version of KernelArg::read_write function uses the following default values for its arguments:

wscale: 1

iwscale: 1

pub fn read_write_no_size( m: &impl UMatTraitConst, wscale: i32, iwscale: i32, ) -> Result<KernelArg>

C++ default parameters

wscale: 1

iwscale: 1

pub fn read_write_no_size_def(m: &impl UMatTraitConst) -> Result<KernelArg>

Note

This alternative version of KernelArg::read_write_no_size function uses the following default values for its arguments:

wscale: 1

iwscale: 1

pub fn read_only( m: &impl UMatTraitConst, wscale: i32, iwscale: i32, ) -> Result<KernelArg>

C++ default parameters

wscale: 1

iwscale: 1

pub fn read_only_def(m: &impl UMatTraitConst) -> Result<KernelArg>

Note

This alternative version of KernelArg::read_only function uses the following default values for its arguments:

wscale: 1

iwscale: 1

pub fn write_only( m: &impl UMatTraitConst, wscale: i32, iwscale: i32, ) -> Result<KernelArg>

C++ default parameters

wscale: 1

iwscale: 1

pub fn write_only_def(m: &impl UMatTraitConst) -> Result<KernelArg>

Note

This alternative version of KernelArg::write_only function uses the following default values for its arguments:

wscale: 1

iwscale: 1

pub fn read_only_no_size( m: &impl UMatTraitConst, wscale: i32, iwscale: i32, ) -> Result<KernelArg>

C++ default parameters

wscale: 1

iwscale: 1

pub fn read_only_no_size_def(m: &impl UMatTraitConst) -> Result<KernelArg>

Note

This alternative version of KernelArg::read_only_no_size function uses the following default values for its arguments:

wscale: 1

iwscale: 1

pub fn write_only_no_size( m: &impl UMatTraitConst, wscale: i32, iwscale: i32, ) -> Result<KernelArg>

C++ default parameters

wscale: 1

iwscale: 1

pub fn write_only_no_size_def(m: &impl UMatTraitConst) -> Result<KernelArg>

Note

This alternative version of KernelArg::write_only_no_size function uses the following default values for its arguments:

wscale: 1

iwscale: 1

pub fn constant(m: &impl MatTraitConst) -> Result<KernelArg>