Struct opencv::core::KeyPoint

pub struct KeyPoint {
    ptr: *mut c_void,
}

Data structure for salient point detectors.

The class instance stores a keypoint, i.e. a point feature found by one of many available keypoint detectors, such as Harris corner detector, #FAST, %StarDetector, %SURF, %SIFT etc.

The keypoint is characterized by the 2D position, scale (proportional to the diameter of the neighborhood that needs to be taken into account), orientation and some other parameters. The keypoint neighborhood is then analyzed by another algorithm that builds a descriptor (usually represented as a feature vector). The keypoints representing the same object in different images can then be matched using %KDTree or another method.

Fields
ptr: *mut c_void
Implementations

impl KeyPoint

pub fn default() -> Result<KeyPoint>

the default constructor

pub fn new_point( pt: Point2f, size: f32, angle: f32, response: f32, octave: i32, class_id: i32 ) -> Result<KeyPoint>

Parameters

pt: x & y coordinates of the keypoint

size: keypoint diameter

angle: keypoint orientation

response: keypoint detector response on the keypoint (that is, strength of the keypoint)

octave: pyramid octave in which the keypoint has been detected

class_id: object id

C++ default parameters

angle: -1

response: 0

octave: 0

class_id: -1

pub fn new_point_def(pt: Point2f, size: f32) -> Result<KeyPoint>

Parameters

pt: x & y coordinates of the keypoint

size: keypoint diameter

angle: keypoint orientation

response: keypoint detector response on the keypoint (that is, strength of the keypoint)

octave: pyramid octave in which the keypoint has been detected

class_id: object id

Note

This alternative version of [new_point] function uses the following default values for its arguments:

angle: -1

response: 0

octave: 0

class_id: -1

pub fn new_coords( x: f32, y: f32, size: f32, angle: f32, response: f32, octave: i32, class_id: i32 ) -> Result<KeyPoint>

Parameters

x: x-coordinate of the keypoint

y: y-coordinate of the keypoint

size: keypoint diameter

angle: keypoint orientation

response: keypoint detector response on the keypoint (that is, strength of the keypoint)

octave: pyramid octave in which the keypoint has been detected

class_id: object id

C++ default parameters

angle: -1

response: 0

octave: 0

class_id: -1

pub fn new_coords_def(x: f32, y: f32, size: f32) -> Result<KeyPoint>

Parameters

x: x-coordinate of the keypoint

y: y-coordinate of the keypoint

size: keypoint diameter

angle: keypoint orientation

response: keypoint detector response on the keypoint (that is, strength of the keypoint)

octave: pyramid octave in which the keypoint has been detected

class_id: object id

Note

This alternative version of [new_coords] function uses the following default values for its arguments:

angle: -1

response: 0

octave: 0

class_id: -1

pub fn convert( keypoints: &Vector<KeyPoint>, points2f: &mut Vector<Point2f>, keypoint_indexes: &Vector<i32> ) -> Result<()>

This method converts vector of keypoints to vector of points or the reverse, where each keypoint is assigned the same size and the same orientation.

Parameters

keypoints: Keypoints obtained from any feature detection algorithm like SIFT/SURF/ORB

points2f: Array of (x,y) coordinates of each keypoint

keypointIndexes: Array of indexes of keypoints to be converted to points. (Acts like a mask to convert only specified keypoints)

C++ default parameters

keypoint_indexes: std::vector()

pub fn convert_def( keypoints: &Vector<KeyPoint>, points2f: &mut Vector<Point2f> ) -> Result<()>

This method converts vector of keypoints to vector of points or the reverse, where each keypoint is assigned the same size and the same orientation.

Parameters

keypoints: Keypoints obtained from any feature detection algorithm like SIFT/SURF/ORB

points2f: Array of (x,y) coordinates of each keypoint

keypointIndexes: Array of indexes of keypoints to be converted to points. (Acts like a mask to convert only specified keypoints)

Note

This alternative version of KeyPoint::convert function uses the following default values for its arguments:

keypoint_indexes: std::vector()

pub fn convert_to( points2f: &Vector<Point2f>, keypoints: &mut Vector<KeyPoint>, size: f32, response: f32, octave: i32, class_id: i32 ) -> Result<()>

This method converts vector of keypoints to vector of points or the reverse, where each keypoint is assigned the same size and the same orientation.

Parameters

keypoints: Keypoints obtained from any feature detection algorithm like SIFT/SURF/ORB

points2f: Array of (x,y) coordinates of each keypoint

keypointIndexes: Array of indexes of keypoints to be converted to points. (Acts like a mask to convert only specified keypoints)

Overloaded parameters

points2f: Array of (x,y) coordinates of each keypoint

keypoints: Keypoints obtained from any feature detection algorithm like SIFT/SURF/ORB

size: keypoint diameter

response: keypoint detector response on the keypoint (that is, strength of the keypoint)

octave: pyramid octave in which the keypoint has been detected

class_id: object id

C++ default parameters

size: 1

response: 1

octave: 0

class_id: -1

pub fn convert_to_def( points2f: &Vector<Point2f>, keypoints: &mut Vector<KeyPoint> ) -> Result<()>

@overload

Parameters

points2f: Array of (x,y) coordinates of each keypoint

keypoints: Keypoints obtained from any feature detection algorithm like SIFT/SURF/ORB

size: keypoint diameter

response: keypoint detector response on the keypoint (that is, strength of the keypoint)

octave: pyramid octave in which the keypoint has been detected

class_id: object id

Note

This alternative version of KeyPoint::convert_to function uses the following default values for its arguments:

size: 1

response: 1

octave: 0

class_id: -1

pub fn overlap( kp1: &impl KeyPointTraitConst, kp2: &impl KeyPointTraitConst ) -> Result<f32>

This method computes overlap for pair of keypoints. Overlap is the ratio between area of keypoint regions’ intersection and area of keypoint regions’ union (considering keypoint region as circle). If they don’t overlap, we get zero. If they coincide at same location with same size, we get 1.

Parameters

kp1: First keypoint

kp2: Second keypoint