Struct opencv::core::MatConstIterator

pub struct MatConstIterator {
    ptr: *mut c_void,
}

/////////////////////////////// MatConstIterator //////////////////////////////////

Fields
ptr: *mut c_void
Implementations

impl MatConstIterator

pub fn default() -> Result<MatConstIterator>

default constructor

pub fn over(_m: &Mat) -> Result<MatConstIterator>

constructor that sets the iterator to the beginning of the matrix

pub fn with_rows_cols( _m: &Mat, _row: i32, _col: i32 ) -> Result<MatConstIterator>

constructor that sets the iterator to the specified element of the matrix

C++ default parameters

_col: 0

pub fn with_rows_cols_def(_m: &Mat, _row: i32) -> Result<MatConstIterator>

constructor that sets the iterator to the specified element of the matrix

Note

This alternative version of [with_rows_cols] function uses the following default values for its arguments:

_col: 0

pub fn with_start(_m: &Mat, _pt: Point) -> Result<MatConstIterator>

constructor that sets the iterator to the specified element of the matrix

pub fn copy(it: &MatConstIterator) -> Result<MatConstIterator>

copy constructor