Struct opencv::core::Mat_

pub struct Mat_<T> {
    inner: Mat,
    _type: PhantomData<T>,
}

docs.opencv.org

This struct is freely convertible into and from Mat using into and try_from methods. You might want to convert Mat to Mat_ before calling typed methods (like at, data_typed) when more performance is required because this way you will skip the data type checks (still WIP, not all methods are covered).

Fields
inner: Mat_type: PhantomData<T>
Implementations

impl<T: DataType> Mat_<T>

pub fn into_untyped(self) -> Mat

pub fn as_untyped(&self) -> &Mat

pub fn as_raw_Mat_(&self) -> *const c_void

pub fn as_raw_mut_Mat_(&mut self) -> *mut c_void

pub fn at(&self, i0: i32) -> Result<&T>

pub fn at_mut(&mut self, i0: i32) -> Result<&mut T>

pub fn data_typed(&self) -> Result<&[T]>

pub fn data_typed_mut(&mut self) -> Result<&mut [T]>