Struct opencv::core::OpenCLExecutionContext

pub struct OpenCLExecutionContext {
    ptr: *mut c_void,
}

Fields
ptr: *mut c_void
Implementations

impl OpenCLExecutionContext

pub fn default() -> OpenCLExecutionContext

pub fn copy( unnamed: &impl OpenCLExecutionContextTraitConst, ) -> OpenCLExecutionContext

pub fn copy_mut(unnamed: OpenCLExecutionContext) -> OpenCLExecutionContext

pub fn get_current() -> Result<OpenCLExecutionContext>

Get OpenCL execution context of current thread.

Initialize OpenCL execution context if it is empty

create new

reuse context of the main thread (threadID = 0)

pub fn get_current_ref() -> Result<OpenCLExecutionContext>

Get OpenCL execution context of current thread (can be empty)

pub unsafe fn create( platform_name: &str, platform_id: *mut c_void, context: *mut c_void, device_id: *mut c_void, ) -> Result<OpenCLExecutionContext>

Creates OpenCL execution context OpenCV will check if available OpenCL platform has platformName name, then assign context to OpenCV. The deviceID device will be used as target device and a new command queue will be created.

Note: On success, ownership of one reference of the context and device is taken. The caller should additionally call clRetainContext and/or clRetainDevice to increase the reference count if it wishes to continue using them.

Parameters

platformName: name of OpenCL platform to attach, this string is used to check if platform is available to OpenCV at runtime

platformID: ID of platform attached context was created for (cl_platform_id)

context: OpenCL context to be attached to OpenCV (cl_context)

deviceID: OpenCL device (cl_device_id)

pub fn create_1( context: &impl ContextTraitConst, device: &impl DeviceTraitConst, queue: &impl QueueTraitConst, ) -> Result<OpenCLExecutionContext>

Creates OpenCL execution context

Parameters

context: non-empty OpenCL context

device: non-empty OpenCL device (must be a part of context)

queue: non-empty OpenCL queue for provided context and device

pub fn create_2( context: &impl ContextTraitConst, device: &impl DeviceTraitConst, ) -> Result<OpenCLExecutionContext>

Creates OpenCL execution context

Parameters

context: non-empty OpenCL context

device: non-empty OpenCL device (must be a part of context)

queue: non-empty OpenCL queue for provided context and device

Overloaded parameters