Struct opencv::core::TargetArchs

pub struct TargetArchs {
    ptr: *mut c_void,
}

Class providing a set of static methods to check what NVIDIA* card architecture the CUDA module was built for.

According to the CUDA C Programming Guide Version 3.2: “PTX code produced for some specific compute capability can always be compiled to binary code of greater or equal compute capability”.

Fields
ptr: *mut c_void
Implementations

impl TargetArchs

pub fn built_with(feature_set: FeatureSet) -> Result<bool>

The following method checks whether the module was built with the support of the given feature:

Parameters

feature_set: Features to be checked. See :ocvcuda::FeatureSet.

pub fn has(major: i32, minor: i32) -> Result<bool>

There is a set of methods to check whether the module contains intermediate (PTX) or binary CUDA code for the given architecture(s):

Parameters

major: Major compute capability version.

minor: Minor compute capability version.

pub fn has_ptx(major: i32, minor: i32) -> Result<bool>

pub fn has_bin(major: i32, minor: i32) -> Result<bool>

pub fn has_equal_or_less_ptx(major: i32, minor: i32) -> Result<bool>

pub fn has_equal_or_greater(major: i32, minor: i32) -> Result<bool>

pub fn has_equal_or_greater_ptx(major: i32, minor: i32) -> Result<bool>

pub fn has_equal_or_greater_bin(major: i32, minor: i32) -> Result<bool>