Struct opencv::core::UMat

pub struct UMat {
    ptr: *mut c_void,
}

@todo document

Fields
ptr: *mut c_void
Implementations

impl UMat

pub fn new(usage_flags: UMatUsageFlags) -> UMat

default constructor

C++ default parameters

usage_flags: USAGE_DEFAULT

pub fn new_def() -> UMat

default constructor

Note

This alternative version of [new] function uses the following default values for its arguments:

usage_flags: USAGE_DEFAULT

pub unsafe fn new_rows_cols( rows: i32, cols: i32, typ: i32, usage_flags: UMatUsageFlags ) -> Result<UMat>

constructs 2D matrix of the specified size and type

C++ default parameters

usage_flags: USAGE_DEFAULT

pub fn new_rows_cols_def(rows: i32, cols: i32, typ: i32) -> Result<UMat>

constructs 2D matrix of the specified size and type

Note

This alternative version of [new_rows_cols] function uses the following default values for its arguments:

usage_flags: USAGE_DEFAULT

pub unsafe fn new_size( size: Size, typ: i32, usage_flags: UMatUsageFlags ) -> Result<UMat>

C++ default parameters

usage_flags: USAGE_DEFAULT

pub fn new_size_def(size: Size, typ: i32) -> Result<UMat>

Note

This alternative version of [new_size] function uses the following default values for its arguments:

usage_flags: USAGE_DEFAULT

pub fn new_rows_cols_with_default( rows: i32, cols: i32, typ: i32, s: Scalar, usage_flags: UMatUsageFlags ) -> Result<UMat>

constructs 2D matrix and fills it with the specified value _s.

C++ default parameters

usage_flags: USAGE_DEFAULT

pub fn new_rows_cols_with_default_def( rows: i32, cols: i32, typ: i32, s: Scalar ) -> Result<UMat>

constructs 2D matrix and fills it with the specified value _s.

Note

This alternative version of [new_rows_cols_with_default] function uses the following default values for its arguments:

usage_flags: USAGE_DEFAULT

pub fn new_size_with_default( size: Size, typ: i32, s: Scalar, usage_flags: UMatUsageFlags ) -> Result<UMat>

C++ default parameters

usage_flags: USAGE_DEFAULT

pub fn new_size_with_default_def( size: Size, typ: i32, s: Scalar ) -> Result<UMat>

Note

This alternative version of [new_size_with_default] function uses the following default values for its arguments:

usage_flags: USAGE_DEFAULT

pub unsafe fn new_nd( sizes: &[i32], typ: i32, usage_flags: UMatUsageFlags ) -> Result<UMat>

constructs n-dimensional matrix

C++ default parameters

usage_flags: USAGE_DEFAULT

pub fn new_nd_def(sizes: &[i32], typ: i32) -> Result<UMat>

constructs n-dimensional matrix

Note

This alternative version of [new_nd] function uses the following default values for its arguments:

usage_flags: USAGE_DEFAULT

pub fn new_nd_with_default( sizes: &[i32], typ: i32, s: Scalar, usage_flags: UMatUsageFlags ) -> Result<UMat>

C++ default parameters

usage_flags: USAGE_DEFAULT

pub fn new_nd_with_default_def( sizes: &[i32], typ: i32, s: Scalar ) -> Result<UMat>

Note

This alternative version of [new_nd_with_default] function uses the following default values for its arguments:

usage_flags: USAGE_DEFAULT

pub fn copy(m: &impl UMatTraitConst) -> Result<UMat>

copy constructor

pub fn rowscols<'boxed>( m: &'boxed impl UMatTraitConst, row_range: &impl RangeTraitConst, col_range: &impl RangeTraitConst ) -> Result<BoxedRef<'boxed, UMat>>

creates a matrix header for a part of the bigger matrix

C++ default parameters

col_range: Range::all()

pub fn rowscols_def_mut<'boxed>( m: &'boxed mut impl UMatTrait, row_range: &impl RangeTraitConst ) -> Result<BoxedRefMut<'boxed, UMat>>

creates a matrix header for a part of the bigger matrix

Note

This alternative version of [rowscols] function uses the following default values for its arguments:

col_range: Range::all()

pub fn rowscols_def<'boxed>( m: &'boxed impl UMatTraitConst, row_range: &impl RangeTraitConst ) -> Result<BoxedRef<'boxed, UMat>>

creates a matrix header for a part of the bigger matrix

Note

This alternative version of [rowscols] function uses the following default values for its arguments:

col_range: Range::all()

pub fn rowscols_mut<'boxed>( m: &'boxed mut impl UMatTrait, row_range: &impl RangeTraitConst, col_range: &impl RangeTraitConst ) -> Result<BoxedRefMut<'boxed, UMat>>

creates a matrix header for a part of the bigger matrix

C++ default parameters

col_range: Range::all()

pub fn roi(m: &impl UMatTraitConst, roi: Rect) -> Result<BoxedRef<'_, UMat>>

pub fn roi_mut( m: &mut impl UMatTrait, roi: Rect ) -> Result<BoxedRefMut<'_, UMat>>

pub fn ranges<'boxed>( m: &'boxed impl UMatTraitConst, ranges: &Vector<Range> ) -> Result<BoxedRef<'boxed, UMat>>

pub fn ranges_mut<'boxed>( m: &'boxed mut impl UMatTrait, ranges: &Vector<Range> ) -> Result<BoxedRefMut<'boxed, UMat>>

pub fn diag_flags( d: &impl UMatTraitConst, usage_flags: UMatUsageFlags ) -> Result<UMat>

constructs a square diagonal matrix which main diagonal is vector “d”

pub fn diag(d: &impl UMatTraitConst) -> Result<UMat>

pub fn zeros_flags( rows: i32, cols: i32, typ: i32, usage_flags: UMatUsageFlags ) -> Result<UMat>

Matlab-style matrix initialization

pub fn zeros_size_flags( size: Size, typ: i32, usage_flags: UMatUsageFlags ) -> Result<UMat>

pub fn zeros_nd_flags( sz: &[i32], typ: i32, usage_flags: UMatUsageFlags ) -> Result<UMat>

pub fn zeros(rows: i32, cols: i32, typ: i32) -> Result<UMat>

pub fn zeros_size(size: Size, typ: i32) -> Result<UMat>

pub fn zeros_nd(sz: &[i32], typ: i32) -> Result<UMat>

pub fn ones_flags( rows: i32, cols: i32, typ: i32, usage_flags: UMatUsageFlags ) -> Result<UMat>

pub fn ones_size_flags( size: Size, typ: i32, usage_flags: UMatUsageFlags ) -> Result<UMat>

pub fn ones_nd_flags( sz: &[i32], typ: i32, usage_flags: UMatUsageFlags ) -> Result<UMat>

pub fn ones(rows: i32, cols: i32, typ: i32) -> Result<UMat>

pub fn ones_size(size: Size, typ: i32) -> Result<UMat>

pub fn ones_nd(sz: &[i32], typ: i32) -> Result<UMat>

pub fn eye_flags( rows: i32, cols: i32, typ: i32, usage_flags: UMatUsageFlags ) -> Result<UMat>

pub fn eye_size_flags( size: Size, typ: i32, usage_flags: UMatUsageFlags ) -> Result<UMat>

pub fn eye(rows: i32, cols: i32, typ: i32) -> Result<UMat>

pub fn eye_size(size: Size, typ: i32) -> Result<UMat>

pub fn copy_mut(m: UMat) -> Result<UMat>