Struct opencv::core::_OutputArray

pub struct _OutputArray {
    ptr: *mut c_void,
}

This type is very similar to InputArray except that it is used for input/output and output function parameters.

Just like with InputArray, OpenCV users should not care about OutputArray, they just pass Mat, vector<T> etc. to the functions. The same limitation as for InputArray: Do not explicitly create OutputArray instances applies here too.

If you want to make your function polymorphic (i.e. accept different arrays as output parameters), it is also not very difficult. Take the sample above as the reference. Note that _OutputArray::create() needs to be called before _OutputArray::getMat(). This way you guarantee that the output array is properly allocated.

Optional output parameters. If you do not need certain output array to be computed and returned to you, pass cv::noArray(), just like you would in the case of optional input array. At the implementation level, use _OutputArray::needed() to check if certain output array needs to be computed or not.

There are several synonyms for OutputArray that are used to assist automatic Python/Java/… wrapper generators:

   typedef OutputArray OutputArrayOfArrays;
   typedef OutputArray InputOutputArray;
   typedef OutputArray InputOutputArrayOfArrays;

Fields
ptr: *mut c_void
Implementations

impl _OutputArray

pub fn default() -> Result<_OutputArray>

/////////////////////////////////////////////////////////////////////////////////////

pub unsafe fn new(_flags: i32, _obj: *mut c_void) -> Result<_OutputArray>

pub fn from_mat_mut( m: &mut impl MatTrait ) -> Result<BoxedRefMut<'_, _OutputArray>>

pub fn from_mat_vec_mut( vec: &mut Vector<Mat> ) -> Result<BoxedRefMut<'_, _OutputArray>>

pub fn from_gpumat_mut( d_mat: &mut impl GpuMatTrait ) -> Result<BoxedRefMut<'_, _OutputArray>>

pub fn from_gpumat_vec_mut( d_mat: &mut Vector<GpuMat> ) -> Result<BoxedRefMut<'_, _OutputArray>>

pub fn from_opengl_mut( buf: &mut impl BufferTrait ) -> Result<BoxedRefMut<'_, _OutputArray>>

pub fn from_hostmem_mut( cuda_mem: &mut impl HostMemTrait ) -> Result<BoxedRefMut<'_, _OutputArray>>

pub fn from_umat_mut( m: &mut impl UMatTrait ) -> Result<BoxedRefMut<'_, _OutputArray>>

pub fn from_umat_vec_mut( vec: &mut Vector<UMat> ) -> Result<BoxedRefMut<'_, _OutputArray>>

pub fn from_mat(m: &impl MatTraitConst) -> Result<BoxedRef<'_, _OutputArray>>

pub fn from_mat_vec(vec: &Vector<Mat>) -> Result<BoxedRef<'_, _OutputArray>>

pub fn from_gpumat( d_mat: &impl GpuMatTraitConst ) -> Result<BoxedRef<'_, _OutputArray>>

pub fn from_opengl( buf: &impl BufferTraitConst ) -> Result<BoxedRef<'_, _OutputArray>>

pub fn from_hostmem( cuda_mem: &impl HostMemTraitConst ) -> Result<BoxedRef<'_, _OutputArray>>

pub fn from_umat(m: &impl UMatTraitConst) -> Result<BoxedRef<'_, _OutputArray>>

pub fn from_umat_vec(vec: &Vector<UMat>) -> Result<BoxedRef<'_, _OutputArray>>