Trait opencv::core::FileNodeIteratorTrait

pub trait FileNodeIteratorTrait: FileNodeIteratorTraitConst {
    // Required method
    fn as_raw_mut_FileNodeIterator(&mut self) -> *mut c_void;

    // Provided methods
    fn set(&mut self, it: &impl FileNodeIteratorTraitConst) -> Result<()> { ... }
    fn incr(&mut self) -> Result<FileNodeIterator> { ... }
    fn read_raw(
        &mut self,
        fmt: &str,
        vec: &mut [u8]
    ) -> Result<FileNodeIterator> { ... }
}

Mutable methods for core::FileNodeIterator

Required Methods

fn as_raw_mut_FileNodeIterator(&mut self) -> *mut c_void

Provided Methods

fn set(&mut self, it: &impl FileNodeIteratorTraitConst) -> Result<()>

fn incr(&mut self) -> Result<FileNodeIterator>

moves iterator to the next node

fn read_raw(&mut self, fmt: &str, vec: &mut [u8]) -> Result<FileNodeIterator>

Reads node elements to the buffer with the specified format.

Usually it is more convenient to use operator >> instead of this method.

Parameters

fmt: Specification of each array element. See [format_spec] “format specification”

vec: Pointer to the destination array.

len: Number of bytes to read (buffer size limit). If it is greater than number of remaining elements then all of them will be read.

C++ default parameters

len: (size_t)INT_MAX

Object Safety

This trait is not object safe.