Trait opencv::core::GpuMatTraitConst

pub trait GpuMatTraitConst {

    // Required method
    fn as_raw_GpuMat(&self) -> *const c_void;

    // Provided methods
    fn flags(&self) -> i32 { ... }
    fn rows(&self) -> i32 { ... }
    fn cols(&self) -> i32 { ... }
    fn step(&self) -> size_t { ... }
    fn data(&self) -> *const u8 { ... }
    fn refcount(&self) -> *const i32 { ... }
    fn datastart(&self) -> *const u8 { ... }
    fn dataend(&self) -> *const u8 { ... }
    fn download(&self, dst: &mut impl ToOutputArray) -> Result<()> { ... }
    fn download_async(
        &self,
        dst: &mut impl ToOutputArray,
        stream: &mut Stream
    ) -> Result<()> { ... }
    fn try_clone(&self) -> Result<GpuMat> { ... }
    fn copy_to(&self, dst: &mut impl ToOutputArray) -> Result<()> { ... }
    fn copy_to_1(
        &self,
        dst: &mut impl ToOutputArray,
        stream: &mut Stream
    ) -> Result<()> { ... }
    fn copy_to_2(
        &self,
        dst: &mut impl ToOutputArray,
        mask: &impl ToInputArray
    ) -> Result<()> { ... }
    fn copy_to_3(
        &self,
        dst: &mut impl ToOutputArray,
        mask: &impl ToInputArray,
        stream: &mut Stream
    ) -> Result<()> { ... }
    fn convert_to(&self, dst: &mut impl ToOutputArray, rtype: i32) -> Result<()> { ... }
    fn convert_to_1(
        &self,
        dst: &mut impl ToOutputArray,
        rtype: i32,
        stream: &mut Stream
    ) -> Result<()> { ... }
    fn convert_to_2(
        &self,
        dst: &mut impl ToOutputArray,
        rtype: i32,
        alpha: f64,
        beta: f64
    ) -> Result<()> { ... }
    fn convert_to_def(
        &self,
        dst: &mut impl ToOutputArray,
        rtype: i32,
        alpha: f64
    ) -> Result<()> { ... }
    fn convert_to_3(
        &self,
        dst: &mut impl ToOutputArray,
        rtype: i32,
        alpha: f64,
        stream: &mut Stream
    ) -> Result<()> { ... }
    fn convert_to_4(
        &self,
        dst: &mut impl ToOutputArray,
        rtype: i32,
        alpha: f64,
        beta: f64,
        stream: &mut Stream
    ) -> Result<()> { ... }
    fn assign_to(&self, m: &mut GpuMat, typ: i32) -> Result<()> { ... }
    fn assign_to_def(&self, m: &mut GpuMat) -> Result<()> { ... }
    fn ptr(&self, y: i32) -> Result<*const u8> { ... }
    fn ptr_def(&self) -> Result<*const u8> { ... }
    fn row(&self, y: i32) -> Result<GpuMat> { ... }
    fn col(&self, x: i32) -> Result<GpuMat> { ... }
    fn row_range(&self, startrow: i32, endrow: i32) -> Result<GpuMat> { ... }
    fn row_range_1(&self, r: Range) -> Result<GpuMat> { ... }
    fn col_range(&self, startcol: i32, endcol: i32) -> Result<GpuMat> { ... }
    fn col_range_1(&self, r: Range) -> Result<GpuMat> { ... }
    fn apply(&self, row_range: Range, col_range: Range) -> Result<GpuMat> { ... }
    fn apply_1(&self, roi: Rect) -> Result<GpuMat> { ... }
    fn reshape(&self, cn: i32, rows: i32) -> Result<GpuMat> { ... }
    fn reshape_def(&self, cn: i32) -> Result<GpuMat> { ... }
    fn locate_roi(&self, whole_size: &mut Size, ofs: &mut Point) -> Result<()> { ... }
    fn is_continuous(&self) -> Result<bool> { ... }
    fn elem_size(&self) -> Result<size_t> { ... }
    fn elem_size1(&self) -> Result<size_t> { ... }
    fn typ(&self) -> Result<i32> { ... }
    fn depth(&self) -> Result<i32> { ... }
    fn channels(&self) -> Result<i32> { ... }
    fn step1(&self) -> Result<size_t> { ... }
    fn size(&self) -> Result<Size> { ... }
    fn empty(&self) -> Result<bool> { ... }
    fn cuda_ptr(&self) -> Result<*mut c_void> { ... }

}

Constant methods for core::GpuMat

Required Methods

fn as_raw_GpuMat(&self) -> *const c_void

Provided Methods

fn flags(&self) -> i32

! includes several bit-fields:

the magic signature

continuity flag

depth

number of channels

fn rows(&self) -> i32

the number of rows and columns

fn cols(&self) -> i32

the number of rows and columns

fn step(&self) -> size_t

a distance between successive rows in bytes; includes the gap if any

fn data(&self) -> *const u8

pointer to the data

fn refcount(&self) -> *const i32

pointer to the reference counter; when GpuMat points to user-allocated data, the pointer is NULL

fn datastart(&self) -> *const u8

helper fields used in locateROI and adjustROI

fn dataend(&self) -> *const u8

fn download(&self, dst: &mut impl ToOutputArray) -> Result<()>

Performs data download from GpuMat (Blocking call)

This function copies data from device memory to host memory. As being a blocking call, it is guaranteed that the copy operation is finished when this function returns.

fn download_async( &self, dst: &mut impl ToOutputArray, stream: &mut Stream ) -> Result<()>

Performs data download from GpuMat (Non-Blocking call)

This function copies data from device memory to host memory. As being a non-blocking call, this function may return even if the copy operation is not finished.

The copy operation may be overlapped with operations in other non-default streams if \p stream is not the default stream and \p dst is HostMem allocated with HostMem::PAGE_LOCKED option.

fn try_clone(&self) -> Result<GpuMat>

returns deep copy of the GpuMat, i.e. the data is copied

fn copy_to(&self, dst: &mut impl ToOutputArray) -> Result<()>

copies the GpuMat content to device memory (Blocking call)

fn copy_to_1( &self, dst: &mut impl ToOutputArray, stream: &mut Stream ) -> Result<()>

copies the GpuMat content to device memory (Non-Blocking call)

fn copy_to_2( &self, dst: &mut impl ToOutputArray, mask: &impl ToInputArray ) -> Result<()>

copies those GpuMat elements to “m” that are marked with non-zero mask elements (Blocking call)

fn copy_to_3( &self, dst: &mut impl ToOutputArray, mask: &impl ToInputArray, stream: &mut Stream ) -> Result<()>

copies those GpuMat elements to “m” that are marked with non-zero mask elements (Non-Blocking call)

fn convert_to(&self, dst: &mut impl ToOutputArray, rtype: i32) -> Result<()>

converts GpuMat to another datatype (Blocking call)

fn convert_to_1( &self, dst: &mut impl ToOutputArray, rtype: i32, stream: &mut Stream ) -> Result<()>

converts GpuMat to another datatype (Non-Blocking call)

fn convert_to_2( &self, dst: &mut impl ToOutputArray, rtype: i32, alpha: f64, beta: f64 ) -> Result<()>

converts GpuMat to another datatype with scaling (Blocking call)

C++ default parameters

beta: 0.0

fn convert_to_def( &self, dst: &mut impl ToOutputArray, rtype: i32, alpha: f64 ) -> Result<()>

converts GpuMat to another datatype with scaling (Blocking call)

Note

This alternative version of GpuMatTraitConst::convert_to function uses the following default values for its arguments:

beta: 0.0

fn convert_to_3( &self, dst: &mut impl ToOutputArray, rtype: i32, alpha: f64, stream: &mut Stream ) -> Result<()>

converts GpuMat to another datatype with scaling (Non-Blocking call)

fn convert_to_4( &self, dst: &mut impl ToOutputArray, rtype: i32, alpha: f64, beta: f64, stream: &mut Stream ) -> Result<()>

converts GpuMat to another datatype with scaling (Non-Blocking call)

fn assign_to(&self, m: &mut GpuMat, typ: i32) -> Result<()>

C++ default parameters

typ: -1

fn assign_to_def(&self, m: &mut GpuMat) -> Result<()>

Note

This alternative version of GpuMatTraitConst::assign_to function uses the following default values for its arguments:

typ: -1

fn ptr(&self, y: i32) -> Result<*const u8>

C++ default parameters

y: 0

fn ptr_def(&self) -> Result<*const u8>

Note

This alternative version of GpuMatTraitConst::ptr function uses the following default values for its arguments:

y: 0

fn row(&self, y: i32) -> Result<GpuMat>

returns a new GpuMat header for the specified row

fn col(&self, x: i32) -> Result<GpuMat>

returns a new GpuMat header for the specified column

fn row_range(&self, startrow: i32, endrow: i32) -> Result<GpuMat>

… for the specified row span

fn row_range_1(&self, r: Range) -> Result<GpuMat>

fn col_range(&self, startcol: i32, endcol: i32) -> Result<GpuMat>

… for the specified column span

fn col_range_1(&self, r: Range) -> Result<GpuMat>

fn apply(&self, row_range: Range, col_range: Range) -> Result<GpuMat>

extracts a rectangular sub-GpuMat (this is a generalized form of row, rowRange etc.)

fn apply_1(&self, roi: Rect) -> Result<GpuMat>

fn reshape(&self, cn: i32, rows: i32) -> Result<GpuMat>

creates alternative GpuMat header for the same data, with different number of channels and/or different number of rows

C++ default parameters

rows: 0

fn reshape_def(&self, cn: i32) -> Result<GpuMat>

creates alternative GpuMat header for the same data, with different number of channels and/or different number of rows

Note

This alternative version of GpuMatTraitConst::reshape function uses the following default values for its arguments:

rows: 0

fn locate_roi(&self, whole_size: &mut Size, ofs: &mut Point) -> Result<()>

locates GpuMat header within a parent GpuMat

fn is_continuous(&self) -> Result<bool>

returns true iff the GpuMat data is continuous (i.e. when there are no gaps between successive rows)

fn elem_size(&self) -> Result<size_t>

returns element size in bytes

fn elem_size1(&self) -> Result<size_t>

returns the size of element channel in bytes

fn typ(&self) -> Result<i32>

returns element type

fn depth(&self) -> Result<i32>

returns element type

fn channels(&self) -> Result<i32>

returns number of channels

fn step1(&self) -> Result<size_t>

returns step/elemSize1()

fn size(&self) -> Result<Size>

returns GpuMat size : width == number of columns, height == number of rows

fn empty(&self) -> Result<bool>

returns true if GpuMat data is NULL

fn cuda_ptr(&self) -> Result<*mut c_void>

Object Safety

This trait is not object safe.