Trait opencv::core::LDATrait

pub trait LDATrait: LDATraitConst {
    // Required method
    fn as_raw_mut_LDA(&mut self) -> *mut c_void;

    // Provided methods
    fn load(&mut self, filename: &str) -> Result<()> { ... }
    fn load_1(&mut self, node: &FileStorage) -> Result<()> { ... }
    fn compute(
        &mut self,
        src: &impl ToInputArray,
        labels: &impl ToInputArray
    ) -> Result<()> { ... }
    fn project(&mut self, src: &impl ToInputArray) -> Result<Mat> { ... }
    fn reconstruct(&mut self, src: &impl ToInputArray) -> Result<Mat> { ... }
}

Mutable methods for core::LDA

Required Methods

fn as_raw_mut_LDA(&mut self) -> *mut c_void

Provided Methods

fn load(&mut self, filename: &str) -> Result<()>

Deserializes this object from a given filename.

fn load_1(&mut self, node: &FileStorage) -> Result<()>

Deserializes this object from a given cv::FileStorage.

fn compute( &mut self, src: &impl ToInputArray, labels: &impl ToInputArray ) -> Result<()>

Compute the discriminants for data in src (row aligned) and labels.

fn project(&mut self, src: &impl ToInputArray) -> Result<Mat>

Projects samples into the LDA subspace. src may be one or more row aligned samples.

fn reconstruct(&mut self, src: &impl ToInputArray) -> Result<Mat>

Reconstructs projections from the LDA subspace. src may be one or more row aligned projections.

Object Safety

This trait is not object safe.