Trait opencv::core::MatTraitManual

pub trait MatTraitManual: MatTraitConstManual + MatTrait {
    // Provided methods
    unsafe fn at_unchecked_mut<T: DataType>(
        &mut self,
        i0: i32
    ) -> Result<&mut T> { ... }
    unsafe fn at_2d_unchecked_mut<T: DataType>(
        &mut self,
        row: i32,
        col: i32
    ) -> Result<&mut T> { ... }
    unsafe fn at_pt_unchecked_mut<T: DataType>(
        &mut self,
        pt: Point
    ) -> Result<&mut T> { ... }
    unsafe fn at_3d_unchecked_mut<T: DataType>(
        &mut self,
        i0: i32,
        i1: i32,
        i2: i32
    ) -> Result<&mut T> { ... }
    unsafe fn at_nd_unchecked_mut<T: DataType>(
        &mut self,
        idx: &[i32]
    ) -> Result<&mut T> { ... }
    fn at_row_mut<T: DataType>(&mut self, row: i32) -> Result<&mut [T]> { ... }
    unsafe fn at_row_unchecked_mut<T: DataType>(
        &mut self,
        row: i32
    ) -> Result<&mut [T]> { ... }
    fn data_bytes_mut(&mut self) -> Result<&mut [u8]> { ... }
    fn data_typed_mut<T: DataType>(&mut self) -> Result<&mut [T]> { ... }
    unsafe fn data_typed_unchecked_mut<T: DataType>(
        &mut self
    ) -> Result<&mut [T]> { ... }
}

Provided Methods

unsafe fn at_unchecked_mut<T: DataType>(&mut self, i0: i32) -> Result<&mut T>

Like Mat::at_mut() but performs no bounds or type checks

Safety

Caller must ensure that index is within Mat bounds

unsafe fn at_2d_unchecked_mut<T: DataType>( &mut self, row: i32, col: i32 ) -> Result<&mut T>

Like Mat::at_2d_mut() but performs no bounds or type checks

Safety

Caller must ensure that indices are within Mat bounds

unsafe fn at_pt_unchecked_mut<T: DataType>( &mut self, pt: Point ) -> Result<&mut T>

Like Mat::at_pt_mut() but performs no bounds or type checks

Safety

Caller must ensure that point is within Mat bounds

unsafe fn at_3d_unchecked_mut<T: DataType>( &mut self, i0: i32, i1: i32, i2: i32 ) -> Result<&mut T>

Like Mat::at_3d_mut() but performs no bounds or type checks

Safety

Caller must ensure that indices are within Mat bounds

unsafe fn at_nd_unchecked_mut<T: DataType>( &mut self, idx: &[i32] ) -> Result<&mut T>

Like Mat::at_nd_mut() but performs no bounds or type checks

Safety

Caller must ensure that indices are within Mat bounds

fn at_row_mut<T: DataType>(&mut self, row: i32) -> Result<&mut [T]>

Return a complete writeable row

unsafe fn at_row_unchecked_mut<T: DataType>( &mut self, row: i32 ) -> Result<&mut [T]>

Like Mat::at_row_mut() but performs no bounds or type checks

Safety

Caller must ensure that index is within Mat bounds

fn data_bytes_mut(&mut self) -> Result<&mut [u8]>

Returns underlying data array as mutable byte slice, Mat must be continuous.

fn data_typed_mut<T: DataType>(&mut self) -> Result<&mut [T]>

unsafe fn data_typed_unchecked_mut<T: DataType>(&mut self) -> Result<&mut [T]>

Safety

Caller must ensure that the T type argument corresponds to the data stored in the Mat

Object Safety

This trait is not object safe.