Trait opencv::core::SVDTrait

pub trait SVDTrait: SVDTraitConst {
    // Required method
    fn as_raw_mut_SVD(&mut self) -> *mut c_void;

    // Provided methods
    fn set_u(&mut self, val: Mat) { ... }
    fn set_w(&mut self, val: Mat) { ... }
    fn set_vt(&mut self, val: Mat) { ... }
    fn apply(&mut self, src: &impl ToInputArray, flags: i32) -> Result<SVD> { ... }
    fn apply_def(&mut self, src: &impl ToInputArray) -> Result<SVD> { ... }
}

Mutable methods for core::SVD

Required Methods

fn as_raw_mut_SVD(&mut self) -> *mut c_void

Provided Methods

fn set_u(&mut self, val: Mat)

fn set_w(&mut self, val: Mat)

fn set_vt(&mut self, val: Mat)

fn apply(&mut self, src: &impl ToInputArray, flags: i32) -> Result<SVD>

the operator that performs SVD. The previously allocated u, w and vt are released.

The operator performs the singular value decomposition of the supplied matrix. The u,vt , and the vector of singular values w are stored in the structure. The same SVD structure can be reused many times with different matrices. Each time, if needed, the previous u,vt , and w are reclaimed and the new matrices are created, which is all handled by Mat::create.

Parameters

src: decomposed matrix. The depth has to be CV_32F or CV_64F.

flags: operation flags (SVD::Flags)

C++ default parameters

flags: 0

fn apply_def(&mut self, src: &impl ToInputArray) -> Result<SVD>

the operator that performs SVD. The previously allocated u, w and vt are released.

The operator performs the singular value decomposition of the supplied matrix. The u,vt , and the vector of singular values w are stored in the structure. The same SVD structure can be reused many times with different matrices. Each time, if needed, the previous u,vt , and w are reclaimed and the new matrices are created, which is all handled by Mat::create.

Parameters

src: decomposed matrix. The depth has to be CV_32F or CV_64F.

flags: operation flags (SVD::Flags)

Note

This alternative version of SVDTrait::apply function uses the following default values for its arguments:

flags: 0

Object Safety

This trait is not object safe.