opencv
Module dnn

Deep Neural Network module

This module contains: - API for new layers creation, layers are building bricks of neural networks; - set of built-in most-useful Layers; - API to construct and modify comprehensive neural networks from layers; - functionality for loading serialized networks models from different frameworks.

Functionality of this module is designed only for forward pass computations (i.e. network testing). A network training is in principle not supported.

Modules

prelude

Structs

AbsLayer

AccumLayer

AcosLayer

AcoshLayer

ActivationLayer

ActivationLayerInt8

ArgLayer

ArgMax/ArgMin layer

AsinLayer

AsinhLayer

AtanLayer

AtanhLayer

BNLLLayer

BackendNode

Derivatives of this class encapsulates functions of certain backends.

BackendWrapper

Derivatives of this class wraps cv::Mat for different backends and targets.

BaseConvolutionLayer

BatchNormLayer

BatchNormLayerInt8

BlankLayer

Partial List of Implemented Layers

CeilLayer

CeluLayer

ChannelsPReLULayer

ClassificationModel

This class represents high-level API for classification models.

CompareLayer

ConcatLayer

ConstLayer

Constant layer produces the same data blob at an every forward pass.

ConvolutionLayer

ConvolutionLayerInt8

CorrelationLayer

CosLayer

CoshLayer

CropAndResizeLayer

CropLayer

CumSumLayer

DataAugmentationLayer

DeconvolutionLayer

DequantizeLayer

DetectionModel

This class represents high-level API for object detection networks.

DetectionOutputLayer

Detection output layer.

Dict

This class implements name-value dictionary, values are instances of DictValue.

DictValue

This struct stores the scalar value (or array) of one of the following type: double, cv::String or int64. @todo Maybe int64 is useless because double type exactly stores at least 2^52 integers.

ELULayer

EltwiseLayer

Element wise operation on inputs

EltwiseLayerInt8

ErfLayer

ExpLayer

FlattenLayer

FloorLayer

FlowWarpLayer

GRULayer

GRU recurrent one-layer

HardSigmoidLayer

HardSwishLayer

InnerProductLayer

InnerProductLayerInt8

InterpLayer

Bilinear resize layer from https://github.com/cdmh/deeplab-public-ver2

KeypointsModel

This class represents high-level API for keypoints models

LRNLayer

LSTMLayer

LSTM recurrent layer

Layer

This interface class allows to build new Layers - are building blocks of networks.

LayerFactory

%Layer factory allows to create instances of registered layers.

LayerParams

This class provides all data needed to initialize layer.

LogLayer

MVNLayer

MaxUnpoolLayer

MishLayer

Model

This class is presented high-level API for neural networks.

Net

This class allows to create and manipulate comprehensive artificial neural networks.

NormalizeBBoxLayer

inline formula - normalization layer.

NotLayer

PaddingLayer

Adds extra values for specific axes.

PermuteLayer

PoolingLayer

PoolingLayerInt8

PowerLayer

PriorBoxLayer

ProposalLayer

QuantizeLayer

RNNLayer

Classical recurrent layer

ReLU6Layer

ReLULayer

ReciprocalLayer

ReduceLayer

ReduceLayerInt8

RegionLayer

ReorgLayer

RequantizeLayer

ReshapeLayer

ResizeLayer

Resize input 4-dimensional blob by nearest neighbor or bilinear strategy.

RoundLayer

ScaleLayer

ScaleLayerInt8

SegmentationModel

This class represents high-level API for segmentation models

SeluLayer

ShiftLayer

ShiftLayerInt8

ShrinkLayer

ShuffleChannelLayer

Permute channels of 4-dimensional input blob.

SigmoidLayer

SignLayer

SinLayer

SinhLayer

SliceLayer

Slice layer has several modes:

SoftmaxLayer

SoftmaxLayerInt8

SoftplusLayer

SoftsignLayer

SplitLayer

SqrtLayer

SwishLayer

TanHLayer

TanLayer

TextDetectionModel

Base class for text detection networks

TextDetectionModel_DB

This class represents high-level API for text detection DL networks compatible with DB model.

TextDetectionModel_EAST

This class represents high-level API for text detection DL networks compatible with EAST model.

TextRecognitionModel

This class represents high-level API for text recognition networks.

ThresholdedReluLayer

_Range

Enums

Backend

Enum of computation backends supported by layers.

SoftNMSMethod

Enum of Soft NMS methods.

Target

Enum of target devices for computations.

Constants

CV_DNN_BACKEND_INFERENCE_ENGINE_NGRAPH

CV_DNN_BACKEND_INFERENCE_ENGINE_NN_BUILDER_API

CV_DNN_INFERENCE_ENGINE_CPU_TYPE_ARM_COMPUTE

CV_DNN_INFERENCE_ENGINE_CPU_TYPE_X86

CV_DNN_INFERENCE_ENGINE_VPU_TYPE_MYRIAD_2

CV_DNN_INFERENCE_ENGINE_VPU_TYPE_MYRIAD_X

CV_DNN_INFERENCE_ENGINE_VPU_TYPE_UNSPECIFIED

DNN_BACKEND_CUDA

DNN_BACKEND_DEFAULT

DNN_BACKEND_DEFAULT equals to DNN_BACKEND_INFERENCE_ENGINE if OpenCV is built with Intel’s Inference Engine library or DNN_BACKEND_OPENCV otherwise.

DNN_BACKEND_HALIDE

DNN_BACKEND_DEFAULT equals to DNN_BACKEND_INFERENCE_ENGINE if OpenCV is built with Intel’s Inference Engine library or DNN_BACKEND_OPENCV otherwise.

DNN_BACKEND_INFERENCE_ENGINE

Intel’s Inference Engine computational backend

DNN_BACKEND_OPENCV

DNN_BACKEND_TIMVX

DNN_BACKEND_VKCOM

DNN_BACKEND_WEBNN

DNN_TARGET_CPU

DNN_TARGET_CUDA

DNN_TARGET_CUDA_FP16

DNN_TARGET_FPGA

FPGA device with CPU fallbacks using Inference Engine’s Heterogeneous plugin.

DNN_TARGET_HDDL

DNN_TARGET_MYRIAD

DNN_TARGET_NPU

DNN_TARGET_OPENCL

DNN_TARGET_OPENCL_FP16

DNN_TARGET_VULKAN

OPENCV_DNN_API_VERSION

SoftNMSMethod_SOFTNMS_GAUSSIAN

SoftNMSMethod_SOFTNMS_LINEAR

Traits

AbsLayerTrait

Mutable methods for crate::dnn::AbsLayer

AbsLayerTraitConst

Constant methods for crate::dnn::AbsLayer

AccumLayerTrait

Mutable methods for crate::dnn::AccumLayer

AccumLayerTraitConst

Constant methods for crate::dnn::AccumLayer

AcosLayerTrait

Mutable methods for crate::dnn::AcosLayer

AcosLayerTraitConst

Constant methods for crate::dnn::AcosLayer

AcoshLayerTrait

Mutable methods for crate::dnn::AcoshLayer

AcoshLayerTraitConst

Constant methods for crate::dnn::AcoshLayer

ActivationLayerInt8Trait

Mutable methods for crate::dnn::ActivationLayerInt8

ActivationLayerInt8TraitConst

Constant methods for crate::dnn::ActivationLayerInt8

ActivationLayerTrait

Mutable methods for crate::dnn::ActivationLayer

ActivationLayerTraitConst

Constant methods for crate::dnn::ActivationLayer

ArgLayerTrait

Mutable methods for crate::dnn::ArgLayer

ArgLayerTraitConst

Constant methods for crate::dnn::ArgLayer

AsinLayerTrait

Mutable methods for crate::dnn::AsinLayer

AsinLayerTraitConst

Constant methods for crate::dnn::AsinLayer

AsinhLayerTrait

Mutable methods for crate::dnn::AsinhLayer

AsinhLayerTraitConst

Constant methods for crate::dnn::AsinhLayer

AtanLayerTrait

Mutable methods for crate::dnn::AtanLayer

AtanLayerTraitConst

Constant methods for crate::dnn::AtanLayer

AtanhLayerTrait

Mutable methods for crate::dnn::AtanhLayer

AtanhLayerTraitConst

Constant methods for crate::dnn::AtanhLayer

BNLLLayerTrait

Mutable methods for crate::dnn::BNLLLayer

BNLLLayerTraitConst

Constant methods for crate::dnn::BNLLLayer

BackendNodeTrait

Mutable methods for crate::dnn::BackendNode

BackendNodeTraitConst

Constant methods for crate::dnn::BackendNode

BackendWrapperTrait

Mutable methods for crate::dnn::BackendWrapper

BackendWrapperTraitConst

Constant methods for crate::dnn::BackendWrapper

BaseConvolutionLayerTrait

Mutable methods for crate::dnn::BaseConvolutionLayer

BaseConvolutionLayerTraitConst

Constant methods for crate::dnn::BaseConvolutionLayer

BatchNormLayerInt8Trait

Mutable methods for crate::dnn::BatchNormLayerInt8

BatchNormLayerInt8TraitConst

Constant methods for crate::dnn::BatchNormLayerInt8

BatchNormLayerTrait

Mutable methods for crate::dnn::BatchNormLayer

BatchNormLayerTraitConst

Constant methods for crate::dnn::BatchNormLayer

BlankLayerTrait

Mutable methods for crate::dnn::BlankLayer

BlankLayerTraitConst

Constant methods for crate::dnn::BlankLayer

CeilLayerTrait

Mutable methods for crate::dnn::CeilLayer

CeilLayerTraitConst

Constant methods for crate::dnn::CeilLayer

CeluLayerTrait

Mutable methods for crate::dnn::CeluLayer

CeluLayerTraitConst

Constant methods for crate::dnn::CeluLayer

ChannelsPReLULayerTrait

Mutable methods for crate::dnn::ChannelsPReLULayer

ChannelsPReLULayerTraitConst

Constant methods for crate::dnn::ChannelsPReLULayer

ClassificationModelTrait

Mutable methods for crate::dnn::ClassificationModel

ClassificationModelTraitConst

Constant methods for crate::dnn::ClassificationModel

CompareLayerTrait

Mutable methods for crate::dnn::CompareLayer

CompareLayerTraitConst

Constant methods for crate::dnn::CompareLayer

ConcatLayerTrait

Mutable methods for crate::dnn::ConcatLayer

ConcatLayerTraitConst

Constant methods for crate::dnn::ConcatLayer

ConstLayerTrait

Mutable methods for crate::dnn::ConstLayer

ConstLayerTraitConst

Constant methods for crate::dnn::ConstLayer

ConvolutionLayerInt8Trait

Mutable methods for crate::dnn::ConvolutionLayerInt8

ConvolutionLayerInt8TraitConst

Constant methods for crate::dnn::ConvolutionLayerInt8

ConvolutionLayerTrait

Mutable methods for crate::dnn::ConvolutionLayer

ConvolutionLayerTraitConst

Constant methods for crate::dnn::ConvolutionLayer

CorrelationLayerTrait

Mutable methods for crate::dnn::CorrelationLayer

CorrelationLayerTraitConst

Constant methods for crate::dnn::CorrelationLayer

CosLayerTrait

Mutable methods for crate::dnn::CosLayer

CosLayerTraitConst

Constant methods for crate::dnn::CosLayer

CoshLayerTrait

Mutable methods for crate::dnn::CoshLayer

CoshLayerTraitConst

Constant methods for crate::dnn::CoshLayer

CropAndResizeLayerTrait

Mutable methods for crate::dnn::CropAndResizeLayer

CropAndResizeLayerTraitConst

Constant methods for crate::dnn::CropAndResizeLayer

CropLayerTrait

Mutable methods for crate::dnn::CropLayer

CropLayerTraitConst

Constant methods for crate::dnn::CropLayer

CumSumLayerTrait

Mutable methods for crate::dnn::CumSumLayer

CumSumLayerTraitConst

Constant methods for crate::dnn::CumSumLayer

DataAugmentationLayerTrait

Mutable methods for crate::dnn::DataAugmentationLayer

DataAugmentationLayerTraitConst

Constant methods for crate::dnn::DataAugmentationLayer

DeconvolutionLayerTrait

Mutable methods for crate::dnn::DeconvolutionLayer

DeconvolutionLayerTraitConst

Constant methods for crate::dnn::DeconvolutionLayer

DequantizeLayerTrait

Mutable methods for crate::dnn::DequantizeLayer

DequantizeLayerTraitConst

Constant methods for crate::dnn::DequantizeLayer

DetectionModelTrait

Mutable methods for crate::dnn::DetectionModel

DetectionModelTraitConst

Constant methods for crate::dnn::DetectionModel

DetectionOutputLayerTrait

Mutable methods for crate::dnn::DetectionOutputLayer

DetectionOutputLayerTraitConst

Constant methods for crate::dnn::DetectionOutputLayer

DictTrait

Mutable methods for crate::dnn::Dict

DictTraitConst

Constant methods for crate::dnn::Dict

DictValueTrait

Mutable methods for crate::dnn::DictValue

DictValueTraitConst

Constant methods for crate::dnn::DictValue

ELULayerTrait

Mutable methods for crate::dnn::ELULayer

ELULayerTraitConst

Constant methods for crate::dnn::ELULayer

EltwiseLayerInt8Trait

Mutable methods for crate::dnn::EltwiseLayerInt8

EltwiseLayerInt8TraitConst

Constant methods for crate::dnn::EltwiseLayerInt8

EltwiseLayerTrait

Mutable methods for crate::dnn::EltwiseLayer

EltwiseLayerTraitConst

Constant methods for crate::dnn::EltwiseLayer

ErfLayerTrait

Mutable methods for crate::dnn::ErfLayer

ErfLayerTraitConst

Constant methods for crate::dnn::ErfLayer

ExpLayerTrait

Mutable methods for crate::dnn::ExpLayer

ExpLayerTraitConst

Constant methods for crate::dnn::ExpLayer

FlattenLayerTrait

Mutable methods for crate::dnn::FlattenLayer

FlattenLayerTraitConst

Constant methods for crate::dnn::FlattenLayer

FloorLayerTrait

Mutable methods for crate::dnn::FloorLayer

FloorLayerTraitConst

Constant methods for crate::dnn::FloorLayer

FlowWarpLayerTrait

Mutable methods for crate::dnn::FlowWarpLayer

FlowWarpLayerTraitConst

Constant methods for crate::dnn::FlowWarpLayer

GRULayerTrait

Mutable methods for crate::dnn::GRULayer

GRULayerTraitConst

Constant methods for crate::dnn::GRULayer

HardSigmoidLayerTrait

Mutable methods for crate::dnn::HardSigmoidLayer

HardSigmoidLayerTraitConst

Constant methods for crate::dnn::HardSigmoidLayer

HardSwishLayerTrait

Mutable methods for crate::dnn::HardSwishLayer

HardSwishLayerTraitConst

Constant methods for crate::dnn::HardSwishLayer

InnerProductLayerInt8Trait

Mutable methods for crate::dnn::InnerProductLayerInt8

InnerProductLayerInt8TraitConst

Constant methods for crate::dnn::InnerProductLayerInt8

InnerProductLayerTrait

Mutable methods for crate::dnn::InnerProductLayer

InnerProductLayerTraitConst

Constant methods for crate::dnn::InnerProductLayer

InterpLayerTrait

Mutable methods for crate::dnn::InterpLayer

InterpLayerTraitConst

Constant methods for crate::dnn::InterpLayer

KeypointsModelTrait

Mutable methods for crate::dnn::KeypointsModel

KeypointsModelTraitConst

Constant methods for crate::dnn::KeypointsModel

LRNLayerTrait

Mutable methods for crate::dnn::LRNLayer

LRNLayerTraitConst

Constant methods for crate::dnn::LRNLayer

LSTMLayerTrait

Mutable methods for crate::dnn::LSTMLayer

LSTMLayerTraitConst

Constant methods for crate::dnn::LSTMLayer

LayerFactoryTrait

Mutable methods for crate::dnn::LayerFactory

LayerFactoryTraitConst

Constant methods for crate::dnn::LayerFactory

LayerParamsTrait

Mutable methods for crate::dnn::LayerParams

LayerParamsTraitConst

Constant methods for crate::dnn::LayerParams

LayerTrait

Mutable methods for crate::dnn::Layer

LayerTraitConst

Constant methods for crate::dnn::Layer

LogLayerTrait

Mutable methods for crate::dnn::LogLayer

LogLayerTraitConst

Constant methods for crate::dnn::LogLayer

MVNLayerTrait

Mutable methods for crate::dnn::MVNLayer

MVNLayerTraitConst

Constant methods for crate::dnn::MVNLayer

MaxUnpoolLayerTrait

Mutable methods for crate::dnn::MaxUnpoolLayer

MaxUnpoolLayerTraitConst

Constant methods for crate::dnn::MaxUnpoolLayer

MishLayerTrait

Mutable methods for crate::dnn::MishLayer

MishLayerTraitConst

Constant methods for crate::dnn::MishLayer

ModelTrait

Mutable methods for crate::dnn::Model

ModelTraitConst

Constant methods for crate::dnn::Model

NetTrait

Mutable methods for crate::dnn::Net

NetTraitConst

Constant methods for crate::dnn::Net

NormalizeBBoxLayerTrait

Mutable methods for crate::dnn::NormalizeBBoxLayer

NormalizeBBoxLayerTraitConst

Constant methods for crate::dnn::NormalizeBBoxLayer

NotLayerTrait

Mutable methods for crate::dnn::NotLayer

NotLayerTraitConst

Constant methods for crate::dnn::NotLayer

PaddingLayerTrait

Mutable methods for crate::dnn::PaddingLayer

PaddingLayerTraitConst

Constant methods for crate::dnn::PaddingLayer

PermuteLayerTrait

Mutable methods for crate::dnn::PermuteLayer

PermuteLayerTraitConst

Constant methods for crate::dnn::PermuteLayer

PoolingLayerInt8Trait

Mutable methods for crate::dnn::PoolingLayerInt8

PoolingLayerInt8TraitConst

Constant methods for crate::dnn::PoolingLayerInt8

PoolingLayerTrait

Mutable methods for crate::dnn::PoolingLayer

PoolingLayerTraitConst

Constant methods for crate::dnn::PoolingLayer

PowerLayerTrait

Mutable methods for crate::dnn::PowerLayer

PowerLayerTraitConst

Constant methods for crate::dnn::PowerLayer

PriorBoxLayerTrait

Mutable methods for crate::dnn::PriorBoxLayer

PriorBoxLayerTraitConst

Constant methods for crate::dnn::PriorBoxLayer

ProposalLayerTrait

Mutable methods for crate::dnn::ProposalLayer

ProposalLayerTraitConst

Constant methods for crate::dnn::ProposalLayer

QuantizeLayerTrait

Mutable methods for crate::dnn::QuantizeLayer

QuantizeLayerTraitConst

Constant methods for crate::dnn::QuantizeLayer

RNNLayerTrait

Mutable methods for crate::dnn::RNNLayer

RNNLayerTraitConst

Constant methods for crate::dnn::RNNLayer

ReLU6LayerTrait

Mutable methods for crate::dnn::ReLU6Layer

ReLU6LayerTraitConst

Constant methods for crate::dnn::ReLU6Layer

ReLULayerTrait

Mutable methods for crate::dnn::ReLULayer

ReLULayerTraitConst

Constant methods for crate::dnn::ReLULayer

ReciprocalLayerTrait

Mutable methods for crate::dnn::ReciprocalLayer

ReciprocalLayerTraitConst

Constant methods for crate::dnn::ReciprocalLayer

ReduceLayerInt8Trait

Mutable methods for crate::dnn::ReduceLayerInt8

ReduceLayerInt8TraitConst

Constant methods for crate::dnn::ReduceLayerInt8

ReduceLayerTrait

Mutable methods for crate::dnn::ReduceLayer

ReduceLayerTraitConst

Constant methods for crate::dnn::ReduceLayer

RegionLayerTrait

Mutable methods for crate::dnn::RegionLayer

RegionLayerTraitConst

Constant methods for crate::dnn::RegionLayer

ReorgLayerTrait

Mutable methods for crate::dnn::ReorgLayer

ReorgLayerTraitConst

Constant methods for crate::dnn::ReorgLayer

RequantizeLayerTrait

Mutable methods for crate::dnn::RequantizeLayer

RequantizeLayerTraitConst

Constant methods for crate::dnn::RequantizeLayer

ReshapeLayerTrait

Mutable methods for crate::dnn::ReshapeLayer

ReshapeLayerTraitConst

Constant methods for crate::dnn::ReshapeLayer

ResizeLayerTrait

Mutable methods for crate::dnn::ResizeLayer

ResizeLayerTraitConst

Constant methods for crate::dnn::ResizeLayer

RoundLayerTrait

Mutable methods for crate::dnn::RoundLayer

RoundLayerTraitConst

Constant methods for crate::dnn::RoundLayer

ScaleLayerInt8Trait

Mutable methods for crate::dnn::ScaleLayerInt8

ScaleLayerInt8TraitConst

Constant methods for crate::dnn::ScaleLayerInt8

ScaleLayerTrait

Mutable methods for crate::dnn::ScaleLayer

ScaleLayerTraitConst

Constant methods for crate::dnn::ScaleLayer

SegmentationModelTrait

Mutable methods for crate::dnn::SegmentationModel

SegmentationModelTraitConst

Constant methods for crate::dnn::SegmentationModel

SeluLayerTrait

Mutable methods for crate::dnn::SeluLayer

SeluLayerTraitConst

Constant methods for crate::dnn::SeluLayer

ShiftLayerInt8Trait

Mutable methods for crate::dnn::ShiftLayerInt8

ShiftLayerInt8TraitConst

Constant methods for crate::dnn::ShiftLayerInt8

ShiftLayerTrait

Mutable methods for crate::dnn::ShiftLayer

ShiftLayerTraitConst

Constant methods for crate::dnn::ShiftLayer

ShrinkLayerTrait

Mutable methods for crate::dnn::ShrinkLayer

ShrinkLayerTraitConst

Constant methods for crate::dnn::ShrinkLayer

ShuffleChannelLayerTrait

Mutable methods for crate::dnn::ShuffleChannelLayer

ShuffleChannelLayerTraitConst

Constant methods for crate::dnn::ShuffleChannelLayer

SigmoidLayerTrait

Mutable methods for crate::dnn::SigmoidLayer

SigmoidLayerTraitConst

Constant methods for crate::dnn::SigmoidLayer

SignLayerTrait

Mutable methods for crate::dnn::SignLayer

SignLayerTraitConst

Constant methods for crate::dnn::SignLayer

SinLayerTrait

Mutable methods for crate::dnn::SinLayer

SinLayerTraitConst

Constant methods for crate::dnn::SinLayer

SinhLayerTrait

Mutable methods for crate::dnn::SinhLayer

SinhLayerTraitConst

Constant methods for crate::dnn::SinhLayer

SliceLayerTrait

Mutable methods for crate::dnn::SliceLayer

SliceLayerTraitConst

Constant methods for crate::dnn::SliceLayer

SoftmaxLayerInt8Trait

Mutable methods for crate::dnn::SoftmaxLayerInt8

SoftmaxLayerInt8TraitConst

Constant methods for crate::dnn::SoftmaxLayerInt8

SoftmaxLayerTrait

Mutable methods for crate::dnn::SoftmaxLayer

SoftmaxLayerTraitConst

Constant methods for crate::dnn::SoftmaxLayer

SoftplusLayerTrait

Mutable methods for crate::dnn::SoftplusLayer

SoftplusLayerTraitConst

Constant methods for crate::dnn::SoftplusLayer

SoftsignLayerTrait

Mutable methods for crate::dnn::SoftsignLayer

SoftsignLayerTraitConst

Constant methods for crate::dnn::SoftsignLayer

SplitLayerTrait

Mutable methods for crate::dnn::SplitLayer

SplitLayerTraitConst

Constant methods for crate::dnn::SplitLayer

SqrtLayerTrait

Mutable methods for crate::dnn::SqrtLayer

SqrtLayerTraitConst

Constant methods for crate::dnn::SqrtLayer

SwishLayerTrait

Mutable methods for crate::dnn::SwishLayer

SwishLayerTraitConst

Constant methods for crate::dnn::SwishLayer

TanHLayerTrait

Mutable methods for crate::dnn::TanHLayer

TanHLayerTraitConst

Constant methods for crate::dnn::TanHLayer

TanLayerTrait

Mutable methods for crate::dnn::TanLayer

TanLayerTraitConst

Constant methods for crate::dnn::TanLayer

TextDetectionModelTrait

Mutable methods for crate::dnn::TextDetectionModel

TextDetectionModelTraitConst

Constant methods for crate::dnn::TextDetectionModel

TextDetectionModel_DBTrait

Mutable methods for crate::dnn::TextDetectionModel_DB

TextDetectionModel_DBTraitConst

Constant methods for crate::dnn::TextDetectionModel_DB

TextDetectionModel_EASTTrait

Mutable methods for crate::dnn::TextDetectionModel_EAST

TextDetectionModel_EASTTraitConst

Constant methods for crate::dnn::TextDetectionModel_EAST

TextRecognitionModelTrait

Mutable methods for crate::dnn::TextRecognitionModel

TextRecognitionModelTraitConst

Constant methods for crate::dnn::TextRecognitionModel

ThresholdedReluLayerTrait

Mutable methods for crate::dnn::ThresholdedReluLayer

ThresholdedReluLayerTraitConst

Constant methods for crate::dnn::ThresholdedReluLayer

_RangeTrait

Mutable methods for crate::dnn::_Range

_RangeTraitConst

Constant methods for crate::dnn::_Range

Functions

blob_from_image

Creates 4-dimensional blob from image. Optionally resizes and crops @p image from center, subtract @p mean values, scales values by @p scalefactor, swap Blue and Red channels.

blob_from_image_def

Creates 4-dimensional blob from image. Optionally resizes and crops @p image from center, subtract @p mean values, scales values by @p scalefactor, swap Blue and Red channels.

blob_from_image_to

Creates 4-dimensional blob from image. @details This is an overloaded member function, provided for convenience. It differs from the above function only in what argument(s) it accepts.

blob_from_image_to_def

Creates 4-dimensional blob from image. @details This is an overloaded member function, provided for convenience. It differs from the above function only in what argument(s) it accepts.

blob_from_images

Creates 4-dimensional blob from series of images. Optionally resizes and crops @p images from center, subtract @p mean values, scales values by @p scalefactor, swap Blue and Red channels.

blob_from_images_def

Creates 4-dimensional blob from series of images. Optionally resizes and crops @p images from center, subtract @p mean values, scales values by @p scalefactor, swap Blue and Red channels.

blob_from_images_to

Creates 4-dimensional blob from series of images. @details This is an overloaded member function, provided for convenience. It differs from the above function only in what argument(s) it accepts.

blob_from_images_to_def

Creates 4-dimensional blob from series of images. @details This is an overloaded member function, provided for convenience. It differs from the above function only in what argument(s) it accepts.

concat

enable_model_diagnostics

Enables detailed logging of the DNN model loading with CV DNN API.

get_available_backends

get_available_targets

get_inference_engine_backend_type

Returns Inference Engine internal backend API.

get_inference_engine_cpu_type

Returns Inference Engine CPU type.

get_inference_engine_vpu_type

Returns Inference Engine VPU type.

get_plane

images_from_blob

Parse a 4D blob and output the images it contains as 2D arrays through a simpler data structure (std::vectorcv::Mat).

nms_boxes

Performs non maximum suppression given boxes and corresponding scores.

nms_boxes_def

Performs non maximum suppression given boxes and corresponding scores.

nms_boxes_f64

C++ default parameters

nms_boxes_f64_def

Note

nms_boxes_rotated

C++ default parameters

nms_boxes_rotated_def

Note

read_net

Read deep learning network represented in one of the supported formats.

read_net_1

Read deep learning network represented in one of the supported formats. @details This is an overloaded member function, provided for convenience. It differs from the above function only in what argument(s) it accepts.

read_net_1_def

Read deep learning network represented in one of the supported formats. @details This is an overloaded member function, provided for convenience. It differs from the above function only in what argument(s) it accepts.

read_net_def

Read deep learning network represented in one of the supported formats.

read_net_from_caffe

Reads a network model stored in Caffe framework’s format.

read_net_from_caffe_buffer

Reads a network model stored in Caffe model in memory.

read_net_from_caffe_buffer_def

Reads a network model stored in Caffe model in memory.

read_net_from_caffe_def

Reads a network model stored in Caffe framework’s format.

read_net_from_caffe_str

Reads a network model stored in Caffe model in memory. @details This is an overloaded member function, provided for convenience. It differs from the above function only in what argument(s) it accepts.

read_net_from_caffe_str_def

Reads a network model stored in Caffe model in memory. @details This is an overloaded member function, provided for convenience. It differs from the above function only in what argument(s) it accepts.

read_net_from_darknet

Reads a network model stored in Darknet model files.

read_net_from_darknet_buffer

Reads a network model stored in Darknet model files.

read_net_from_darknet_buffer_def

Reads a network model stored in Darknet model files.

read_net_from_darknet_def

Reads a network model stored in Darknet model files.

read_net_from_darknet_str

Reads a network model stored in Darknet model files.

read_net_from_darknet_str_def

Reads a network model stored in Darknet model files.

read_net_from_model_optimizer

Load a network from Intel’s Model Optimizer intermediate representation.

read_net_from_model_optimizer_1

Load a network from Intel’s Model Optimizer intermediate representation.

read_net_from_model_optimizer_2

Load a network from Intel’s Model Optimizer intermediate representation.

read_net_from_onnx

Reads a network model ONNX.

read_net_from_onnx_buffer

Reads a network model from ONNX in-memory buffer.

read_net_from_onnx_str

Reads a network model from ONNX in-memory buffer.

read_net_from_tensorflow

Reads a network model stored in TensorFlow framework’s format.

read_net_from_tensorflow_buffer

Reads a network model stored in TensorFlow framework’s format.

read_net_from_tensorflow_buffer_def

Reads a network model stored in TensorFlow framework’s format.

read_net_from_tensorflow_def

Reads a network model stored in TensorFlow framework’s format.

read_net_from_tensorflow_str

Reads a network model stored in TensorFlow framework’s format. @details This is an overloaded member function, provided for convenience. It differs from the above function only in what argument(s) it accepts.

read_net_from_tensorflow_str_def

Reads a network model stored in TensorFlow framework’s format. @details This is an overloaded member function, provided for convenience. It differs from the above function only in what argument(s) it accepts.

read_net_from_torch

Reads a network model stored in Torch7 framework’s format.

read_net_from_torch_def

Reads a network model stored in Torch7 framework’s format.

read_tensor_from_onnx

Creates blob from .pb file.

read_torch_blob

Loads blob which was serialized as torch.Tensor object of Torch7 framework. @warning This function has the same limitations as readNetFromTorch().

read_torch_blob_def

Loads blob which was serialized as torch.Tensor object of Torch7 framework. @warning This function has the same limitations as readNetFromTorch().

release_hddl_plugin

Release a HDDL plugin.

reset_myriad_device

Release a Myriad device (binded by OpenCV).

set_inference_engine_backend_type

Specify Inference Engine internal backend API.

shape

shape_1

shape_2

shape_3

shape_4

C++ default parameters

shape_4_def

Note

shrink_caffe_model

Convert all weights of Caffe network to half precision floating point.

shrink_caffe_model_def

Convert all weights of Caffe network to half precision floating point.

slice

slice_1

slice_2

slice_3

soft_nms_boxes

Performs soft non maximum suppression given boxes and corresponding scores. Reference: https://arxiv.org/abs/1704.04503

soft_nms_boxes_def

Performs soft non maximum suppression given boxes and corresponding scores. Reference: https://arxiv.org/abs/1704.04503

total

C++ default parameters

total_def

Note

write_text_graph

Create a text representation for a binary network stored in protocol buffer format.

Type Aliases

LayerFactory_Constructor

Each Layer class must provide this function to the factory

MatShape

Net_LayerIdDeprecated

Container for strings and integers.