opencv::dnn
Function blob_from_images_to

pub fn blob_from_images_to(
    images: &impl ToInputArray,
    blob: &mut impl ToOutputArray,
    scalefactor: f64,
    size: Size,
    mean: Scalar,
    swap_rb: bool,
    crop: bool,
    ddepth: i32,
) -> Result<()>

Creates 4-dimensional blob from series of images. @details This is an overloaded member function, provided for convenience. It differs from the above function only in what argument(s) it accepts.

C++ default parameters

scalefactor: 1.0

size: Size()

mean: Scalar()

swap_rb: false

crop: false

ddepth: CV_32F