Function opencv::dnn::images_from_blob

pub fn images_from_blob(
    blob_: &impl MatTraitConst,
    images_: &mut impl ToOutputArray
) -> Result<()>

Parse a 4D blob and output the images it contains as 2D arrays through a simpler data structure (std::vectorcv::Mat).

Parameters

blob_: 4 dimensional array (images, channels, height, width) in floating point precision (CV_32F) from which you would like to extract the images.

images_:[out] array of 2D Mat containing the images extracted from the blob in floating point precision (CV_32F). They are non normalized neither mean added. The number of returned images equals the first dimension of the blob (batch size). Every image has a number of channels equals to the second dimension of the blob (depth).