opencv::dnn
Function read_net_from_caffe_str

pub fn read_net_from_caffe_str(
    buffer_proto: &str,
    len_proto: size_t,
    buffer_model: &str,
    len_model: size_t,
) -> Result<Net>

Reads a network model stored in Caffe model in memory. @details This is an overloaded member function, provided for convenience. It differs from the above function only in what argument(s) it accepts.

Parameters

bufferProto: buffer containing the content of the .prototxt file

lenProto: length of bufferProto

bufferModel: buffer containing the content of the .caffemodel file

lenModel: length of bufferModel

Returns

Net object.

C++ default parameters

buffer_model: NULL

len_model: 0