Function opencv::dnn::read_net_from_caffe_str_def

pub fn read_net_from_caffe_str_def(
    buffer_proto: &str,
    len_proto: size_t
) -> Result<Net>

Reads a network model stored in Caffe model in memory. @details This is an overloaded member function, provided for convenience. It differs from the above function only in what argument(s) it accepts.

Parameters

bufferProto: buffer containing the content of the .prototxt file

lenProto: length of bufferProto

bufferModel: buffer containing the content of the .caffemodel file

lenModel: length of bufferModel

Returns

Net object.

Note

This alternative version of read_net_from_caffe_str function uses the following default values for its arguments:

buffer_model: NULL

len_model: 0