Struct opencv::dnn::ClassificationModel

pub struct ClassificationModel {
    ptr: *mut c_void,
}

This class represents high-level API for classification models.

ClassificationModel allows to set params for preprocessing input image. ClassificationModel creates net from file with trained weights and config, sets preprocessing input, runs forward pass and return top-1 prediction.

Fields
ptr: *mut c_void
Implementations

impl ClassificationModel

pub fn new(model: &str, config: &str) -> Result<ClassificationModel>

Create classification model from network represented in one of the supported formats. An order of @p model and @p config arguments does not matter.

Parameters

model: Binary file contains trained weights.

config: Text file contains network configuration.

C++ default parameters

config: “”

pub fn new_def(model: &str) -> Result<ClassificationModel>

Create classification model from network represented in one of the supported formats. An order of @p model and @p config arguments does not matter.

Parameters

model: Binary file contains trained weights.

config: Text file contains network configuration.

Note

This alternative version of [new] function uses the following default values for its arguments:

config: “”

pub fn new_1(network: &impl NetTraitConst) -> Result<ClassificationModel>

Create model from deep learning network.

Parameters

network: Net object.