Struct opencv::dnn::DetectionModel

pub struct DetectionModel {
    ptr: *mut c_void,
}

This class represents high-level API for object detection networks.

DetectionModel allows to set params for preprocessing input image. DetectionModel creates net from file with trained weights and config, sets preprocessing input, runs forward pass and return result detections. For DetectionModel SSD, Faster R-CNN, YOLO topologies are supported.

Fields
ptr: *mut c_void
Implementations

impl DetectionModel

pub fn new(model: &str, config: &str) -> Result<DetectionModel>

Create detection model from network represented in one of the supported formats. An order of @p model and @p config arguments does not matter.

Parameters

model: Binary file contains trained weights.

config: Text file contains network configuration.

C++ default parameters

config: “”

pub fn new_def(model: &str) -> Result<DetectionModel>

Create detection model from network represented in one of the supported formats. An order of @p model and @p config arguments does not matter.

Parameters

model: Binary file contains trained weights.

config: Text file contains network configuration.

Note

This alternative version of [new] function uses the following default values for its arguments:

config: “”

pub fn new_1(network: &impl NetTraitConst) -> Result<DetectionModel>

Create model from deep learning network.

Parameters

network: Net object.

pub fn default() -> Result<DetectionModel>