opencv::dnn
Struct Layer

pub struct Layer {
    ptr: *mut c_void,
}

This interface class allows to build new Layers - are building blocks of networks.

Each class, derived from Layer, must implement allocate() methods to declare own outputs and forward() to compute outputs. Also before using the new layer into networks you must register your layer by using one of [dnnLayerFactory] “LayerFactory” macros.

Fields
ptr: *mut c_void
Implementations

impl Layer

pub fn default() -> Result<Layer>

pub fn new(params: &impl LayerParamsTraitConst) -> Result<Layer>