Struct opencv::dnn::Net

pub struct Net {
    ptr: *mut c_void,
}

This class allows to create and manipulate comprehensive artificial neural networks.

Neural network is presented as directed acyclic graph (DAG), where vertices are Layer instances, and edges specify relationships between layers inputs and outputs.

Each network layer has unique integer id and unique string name inside its network. LayerId can store either layer name or layer id.

This class supports reference counting of its instances, i. e. copies point to the same instance.

Fields
ptr: *mut c_void
Implementations

impl Net

pub fn default() -> Result<Net>

pub fn read_from_model_optimizer(xml: &str, bin: &str) -> Result<Net>

Create a network from Intel’s Model Optimizer intermediate representation (IR).

Parameters

xml: XML configuration file with network’s topology.

bin: Binary file with trained weights. Networks imported from Intel’s Model Optimizer are launched in Intel’s Inference Engine backend.

pub fn read_from_model_optimizer_1( buffer_model_config: &Vector<u8>, buffer_weights: &Vector<u8>, ) -> Result<Net>

Create a network from Intel’s Model Optimizer in-memory buffers with intermediate representation (IR).

Parameters

bufferModelConfig: buffer with model’s configuration.

bufferWeights: buffer with model’s trained weights.

Returns

Net object.

pub fn read_from_model_optimizer_2( buffer_model_config_ptr: &[u8], buffer_weights_ptr: &[u8], ) -> Result<Net>

Create a network from Intel’s Model Optimizer in-memory buffers with intermediate representation (IR).

Parameters

bufferModelConfigPtr: buffer pointer of model’s configuration.

bufferModelConfigSize: buffer size of model’s configuration.

bufferWeightsPtr: buffer pointer of model’s trained weights.

bufferWeightsSize: buffer size of model’s trained weights.

Returns

Net object.