Struct opencv::dnn::TextDetectionModel_DB

pub struct TextDetectionModel_DB {
    ptr: *mut c_void,
}

This class represents high-level API for text detection DL networks compatible with DB model.

Related publications: liao2020real Paper: https://arxiv.org/abs/1911.08947 For more information about the hyper-parameters setting, please refer to https://github.com/MhLiao/DB

Configurable parameters:

(float) binaryThreshold - The threshold of the binary map. It is usually set to 0.3.

(float) polygonThreshold - The threshold of text polygons. It is usually set to 0.5, 0.6, and 0.7. Default is 0.5f

(double) unclipRatio - The unclip ratio of the detected text region, which determines the output size. It is usually set to 2.0.

(int) maxCandidates - The max number of the output results.

Fields
ptr: *mut c_void
Implementations

impl TextDetectionModel_DB

pub fn default() -> Result<TextDetectionModel_DB>

pub fn new(network: &impl NetTraitConst) -> Result<TextDetectionModel_DB>

Create text detection algorithm from deep learning network.

Parameters

network: Net object.

pub fn new_1(model: &str, config: &str) -> Result<TextDetectionModel_DB>

Create text detection model from network represented in one of the supported formats. An order of @p model and @p config arguments does not matter.

Parameters

model: Binary file contains trained weights.

config: Text file contains network configuration.

C++ default parameters

config: “”

pub fn new_def(model: &str) -> Result<TextDetectionModel_DB>

Create text detection model from network represented in one of the supported formats. An order of @p model and @p config arguments does not matter.

Parameters

model: Binary file contains trained weights.

config: Text file contains network configuration.

Note

This alternative version of [new] function uses the following default values for its arguments:

config: “”