Struct opencv::dnn::TextDetectionModel_EAST

pub struct TextDetectionModel_EAST {
    ptr: *mut c_void,
}

This class represents high-level API for text detection DL networks compatible with EAST model.

Configurable parameters:

(float) confThreshold - used to filter boxes by confidences, default: 0.5f

(float) nmsThreshold - used in non maximum suppression, default: 0.0f

Fields
ptr: *mut c_void
Implementations

impl TextDetectionModel_EAST

pub fn default() -> Result<TextDetectionModel_EAST>

pub fn new(network: &Net) -> Result<TextDetectionModel_EAST>

Create text detection algorithm from deep learning network

Parameters

network: Net object

pub fn from_file(model: &str, config: &str) -> Result<TextDetectionModel_EAST>

Create text detection model from network represented in one of the supported formats. An order of @p model and @p config arguments does not matter.

Parameters

model: Binary file contains trained weights.

config: Text file contains network configuration.

C++ default parameters

config: “”

pub fn from_file_def(model: &str) -> Result<TextDetectionModel_EAST>

Create text detection model from network represented in one of the supported formats. An order of @p model and @p config arguments does not matter.

Parameters

model: Binary file contains trained weights.

config: Text file contains network configuration.

Note

This alternative version of [from_file] function uses the following default values for its arguments:

config: “”