Trait opencv::dnn::DictValueTraitConst

pub trait DictValueTraitConst {

    // Required method
    fn as_raw_DictValue(&self) -> *const c_void;

    // Provided methods
    fn get_str(&self, idx: i32) -> Result<String> { ... }
    fn get_str_def(&self) -> Result<String> { ... }
    fn get_f64(&self, idx: i32) -> Result<f64> { ... }
    fn get_f64_def(&self) -> Result<f64> { ... }
    fn get_i32(&self, idx: i32) -> Result<i32> { ... }
    fn get_i32_def(&self) -> Result<i32> { ... }
    fn get_i64(&self, idx: i32) -> Result<i64> { ... }
    fn get_i64_def(&self) -> Result<i64> { ... }
    fn size(&self) -> Result<i32> { ... }
    fn is_int(&self) -> Result<bool> { ... }
    fn is_string(&self) -> Result<bool> { ... }
    fn is_real(&self) -> Result<bool> { ... }
    fn get_int_value(&self, idx: i32) -> Result<i32> { ... }
    fn get_int_value_def(&self) -> Result<i32> { ... }
    fn get_real_value(&self, idx: i32) -> Result<f64> { ... }
    fn get_real_value_def(&self) -> Result<f64> { ... }
    fn get_string_value(&self, idx: i32) -> Result<String> { ... }
    fn get_string_value_def(&self) -> Result<String> { ... }

}

Constant methods for crate::dnn::DictValue

Required Methods

fn as_raw_DictValue(&self) -> *const c_void

Provided Methods

fn get_str(&self, idx: i32) -> Result<String>

C++ default parameters

idx: -1

fn get_str_def(&self) -> Result<String>

fn get_f64(&self, idx: i32) -> Result<f64>

C++ default parameters

idx: -1

fn get_f64_def(&self) -> Result<f64>

fn get_i32(&self, idx: i32) -> Result<i32>

C++ default parameters

idx: -1

fn get_i32_def(&self) -> Result<i32>

fn get_i64(&self, idx: i32) -> Result<i64>

C++ default parameters

idx: -1

fn get_i64_def(&self) -> Result<i64>

fn size(&self) -> Result<i32>

fn is_int(&self) -> Result<bool>

fn is_string(&self) -> Result<bool>

fn is_real(&self) -> Result<bool>

fn get_int_value(&self, idx: i32) -> Result<i32>

C++ default parameters

idx: -1

fn get_int_value_def(&self) -> Result<i32>

Note

This alternative version of DictValueTraitConst::get_int_value function uses the following default values for its arguments:

idx: -1

fn get_real_value(&self, idx: i32) -> Result<f64>

C++ default parameters

idx: -1

fn get_real_value_def(&self) -> Result<f64>

Note

This alternative version of DictValueTraitConst::get_real_value function uses the following default values for its arguments:

idx: -1

fn get_string_value(&self, idx: i32) -> Result<String>

C++ default parameters

idx: -1

fn get_string_value_def(&self) -> Result<String>

Note

This alternative version of DictValueTraitConst::get_string_value function uses the following default values for its arguments:

idx: -1