Trait opencv::dnn::ModelTraitConst

pub trait ModelTraitConst {
    // Required method
    fn as_raw_Model(&self) -> *const c_void;

    // Provided methods
    fn predict(
        &self,
        frame: &impl ToInputArray,
        outs: &mut impl ToOutputArray
    ) -> Result<()> { ... }
    fn get_network_(&self) -> Result<Net> { ... }
}

Constant methods for crate::dnn::Model

Required Methods

fn as_raw_Model(&self) -> *const c_void

Provided Methods

fn predict( &self, frame: &impl ToInputArray, outs: &mut impl ToOutputArray ) -> Result<()>

Given the @p input frame, create input blob, run net and return the output @p blobs.

Parameters

frame: The input image.

outs:[out] Allocated output blobs, which will store results of the computation.

fn get_network_(&self) -> Result<Net>

Object Safety

This trait is not object safe.