Trait opencv::hub_prelude::FormatterTrait

pub trait FormatterTrait: FormatterTraitConst {
    // Required method
    fn as_raw_mut_Formatter(&mut self) -> *mut c_void;

    // Provided methods
    fn set16f_precision(&mut self, p: i32) -> Result<()> { ... }
    fn set16f_precision_def(&mut self) -> Result<()> { ... }
    fn set32f_precision(&mut self, p: i32) -> Result<()> { ... }
    fn set32f_precision_def(&mut self) -> Result<()> { ... }
    fn set64f_precision(&mut self, p: i32) -> Result<()> { ... }
    fn set64f_precision_def(&mut self) -> Result<()> { ... }
    fn set_multiline(&mut self, ml: bool) -> Result<()> { ... }
    fn set_multiline_def(&mut self) -> Result<()> { ... }
}

Mutable methods for core::Formatter

Required Methods

fn as_raw_mut_Formatter(&mut self) -> *mut c_void

Provided Methods

fn set16f_precision(&mut self, p: i32) -> Result<()>

C++ default parameters

p: 4

fn set16f_precision_def(&mut self) -> Result<()>

Note

This alternative version of FormatterTrait::set16f_precision function uses the following default values for its arguments:

p: 4

fn set32f_precision(&mut self, p: i32) -> Result<()>

C++ default parameters

p: 8

fn set32f_precision_def(&mut self) -> Result<()>

Note

This alternative version of FormatterTrait::set32f_precision function uses the following default values for its arguments:

p: 8

fn set64f_precision(&mut self, p: i32) -> Result<()>

C++ default parameters

p: 16

fn set64f_precision_def(&mut self) -> Result<()>

Note

This alternative version of FormatterTrait::set64f_precision function uses the following default values for its arguments:

p: 16

fn set_multiline(&mut self, ml: bool) -> Result<()>

C++ default parameters

ml: true

fn set_multiline_def(&mut self) -> Result<()>

Note

This alternative version of FormatterTrait::set_multiline function uses the following default values for its arguments:

ml: true

Object Safety

This trait is not object safe.