Trait opencv::hub_prelude::KeypointsModelTrait

pub trait KeypointsModelTrait: KeypointsModelTraitConst + ModelTrait {
    // Required method
    fn as_raw_mut_KeypointsModel(&mut self) -> *mut c_void;

    // Provided methods
    fn estimate(
        &mut self,
        frame: &impl ToInputArray,
        thresh: f32,
    ) -> Result<Vector<Point2f>> { ... }
    fn estimate_def(
        &mut self,
        frame: &impl ToInputArray,
    ) -> Result<Vector<Point2f>> { ... }
}

Mutable methods for crate::dnn::KeypointsModel

Required Methods

fn as_raw_mut_KeypointsModel(&mut self) -> *mut c_void

Provided Methods

fn estimate( &mut self, frame: &impl ToInputArray, thresh: f32, ) -> Result<Vector<Point2f>>

Given the @p input frame, create input blob, run net

Parameters

frame: The input image.

thresh: minimum confidence threshold to select a keypoint

Returns

a vector holding the x and y coordinates of each detected keypoint

C++ default parameters

thresh: 0.5

fn estimate_def(&mut self, frame: &impl ToInputArray) -> Result<Vector<Point2f>>

Given the @p input frame, create input blob, run net

Parameters

frame: The input image.

thresh: minimum confidence threshold to select a keypoint

Returns

a vector holding the x and y coordinates of each detected keypoint

Note

This alternative version of KeypointsModelTrait::estimate function uses the following default values for its arguments:

thresh: 0.5

Object Safety

This trait is not object safe.