Trait opencv::hub_prelude::MinProblemSolverTraitConst

pub trait MinProblemSolverTraitConst: AlgorithmTraitConst {
    // Required method
    fn as_raw_MinProblemSolver(&self) -> *const c_void;

    // Provided methods
    fn get_function(&self) -> Result<Ptr<MinProblemSolver_Function>> { ... }
    fn get_term_criteria(&self) -> Result<TermCriteria> { ... }
}

Constant methods for core::MinProblemSolver

Required Methods

fn as_raw_MinProblemSolver(&self) -> *const c_void

Provided Methods

fn get_function(&self) -> Result<Ptr<MinProblemSolver_Function>>

Getter for the optimized function.

The optimized function is represented by Function interface, which requires derivatives to implement the calc(double*) and getDim() methods to evaluate the function.

Returns

Smart-pointer to an object that implements Function interface - it represents the function that is being optimized. It can be empty, if no function was given so far.

fn get_term_criteria(&self) -> Result<TermCriteria>

Getter for the previously set terminal criteria for this algorithm.

Returns

Deep copy of the terminal criteria used at the moment.

Object Safety

This trait is not object safe.