Trait opencv::hub_prelude::RNGTrait

pub trait RNGTrait: RNGTraitConst {

    // Required method
    fn as_raw_mut_RNG(&mut self) -> *mut c_void;

    // Provided methods
    fn set_state(&mut self, val: u64) { ... }
    fn next(&mut self) -> Result<u32> { ... }
    fn to_u8(&mut self) -> Result<u8> { ... }
    fn to_i8(&mut self) -> Result<i8> { ... }
    fn to_u16(&mut self) -> Result<u16> { ... }
    fn to_i16(&mut self) -> Result<i16> { ... }
    fn to_u32(&mut self) -> Result<u32> { ... }
    fn to_i32(&mut self) -> Result<i32> { ... }
    fn to_f32(&mut self) -> Result<f32> { ... }
    fn to_f64(&mut self) -> Result<f64> { ... }
    fn apply(&mut self) -> Result<u32> { ... }
    fn apply_range(&mut self, n: u32) -> Result<u32> { ... }
    fn uniform(&mut self, a: i32, b: i32) -> Result<i32> { ... }
    fn uniform_f32(&mut self, a: f32, b: f32) -> Result<f32> { ... }
    fn uniform_f64(&mut self, a: f64, b: f64) -> Result<f64> { ... }
    fn fill(
        &mut self,
        mat: &mut impl ToInputOutputArray,
        dist_type: i32,
        a: &impl ToInputArray,
        b: &impl ToInputArray,
        saturate_range: bool
    ) -> Result<()> { ... }
    fn fill_def(
        &mut self,
        mat: &mut impl ToInputOutputArray,
        dist_type: i32,
        a: &impl ToInputArray,
        b: &impl ToInputArray
    ) -> Result<()> { ... }
    fn gaussian(&mut self, sigma: f64) -> Result<f64> { ... }

}

Mutable methods for core::RNG

Required Methods

fn as_raw_mut_RNG(&mut self) -> *mut c_void

Provided Methods

fn set_state(&mut self, val: u64)

fn next(&mut self) -> Result<u32>

The method updates the state using the MWC algorithm and returns the next 32-bit random number.

fn to_u8(&mut self) -> Result<u8>

Each of the methods updates the state using the MWC algorithm and returns the next random number of the specified type. In case of integer types, the returned number is from the available value range for the specified type. In case of floating-point types, the returned value is from [0,1) range.

fn to_i8(&mut self) -> Result<i8>

This is an overloaded member function, provided for convenience. It differs from the above function only in what argument(s) it accepts.

fn to_u16(&mut self) -> Result<u16>

This is an overloaded member function, provided for convenience. It differs from the above function only in what argument(s) it accepts.

fn to_i16(&mut self) -> Result<i16>

This is an overloaded member function, provided for convenience. It differs from the above function only in what argument(s) it accepts.

fn to_u32(&mut self) -> Result<u32>

This is an overloaded member function, provided for convenience. It differs from the above function only in what argument(s) it accepts.

fn to_i32(&mut self) -> Result<i32>

This is an overloaded member function, provided for convenience. It differs from the above function only in what argument(s) it accepts.

fn to_f32(&mut self) -> Result<f32>

This is an overloaded member function, provided for convenience. It differs from the above function only in what argument(s) it accepts.

fn to_f64(&mut self) -> Result<f64>

This is an overloaded member function, provided for convenience. It differs from the above function only in what argument(s) it accepts.

fn apply(&mut self) -> Result<u32>

returns a random integer sampled uniformly from [0, N).

The methods transform the state using the MWC algorithm and return the next random number. The first form is equivalent to RNG::next . The second form returns the random number modulo N , which means that the result is in the range [0, N) .

fn apply_range(&mut self, n: u32) -> Result<u32>

returns a random integer sampled uniformly from [0, N).

The methods transform the state using the MWC algorithm and return the next random number. The first form is equivalent to RNG::next . The second form returns the random number modulo N , which means that the result is in the range [0, N) .

Overloaded parameters

Parameters

N: upper non-inclusive boundary of the returned random number.

fn uniform(&mut self, a: i32, b: i32) -> Result<i32>

returns uniformly distributed integer random number from [a,b) range

The methods transform the state using the MWC algorithm and return the next uniformly-distributed random number of the specified type, deduced from the input parameter type, from the range [a, b) . There is a nuance illustrated by the following sample:

RNG rng;
 
// always produces 0
double a = rng.uniform(0, 1);
 
// produces double from [0, 1)
double a1 = rng.uniform((double)0, (double)1);
 
// produces float from [0, 1)
float b = rng.uniform(0.f, 1.f);
 
// produces double from [0, 1)
double c = rng.uniform(0., 1.);
 
// may cause compiler error because of ambiguity:
//  RNG::uniform(0, (int)0.999999)? or RNG::uniform((double)0, 0.99999)?
double d = rng.uniform(0, 0.999999);

The compiler does not take into account the type of the variable to which you assign the result of RNG::uniform . The only thing that matters to the compiler is the type of a and b parameters. So, if you want a floating-point random number, but the range boundaries are integer numbers, either put dots in the end, if they are constants, or use explicit type cast operators, as in the a1 initialization above.

Parameters

a: lower inclusive boundary of the returned random number.

b: upper non-inclusive boundary of the returned random number.

fn uniform_f32(&mut self, a: f32, b: f32) -> Result<f32>

returns uniformly distributed integer random number from [a,b) range

The methods transform the state using the MWC algorithm and return the next uniformly-distributed random number of the specified type, deduced from the input parameter type, from the range [a, b) . There is a nuance illustrated by the following sample:

RNG rng;
 
// always produces 0
double a = rng.uniform(0, 1);
 
// produces double from [0, 1)
double a1 = rng.uniform((double)0, (double)1);
 
// produces float from [0, 1)
float b = rng.uniform(0.f, 1.f);
 
// produces double from [0, 1)
double c = rng.uniform(0., 1.);
 
// may cause compiler error because of ambiguity:
//  RNG::uniform(0, (int)0.999999)? or RNG::uniform((double)0, 0.99999)?
double d = rng.uniform(0, 0.999999);

The compiler does not take into account the type of the variable to which you assign the result of RNG::uniform . The only thing that matters to the compiler is the type of a and b parameters. So, if you want a floating-point random number, but the range boundaries are integer numbers, either put dots in the end, if they are constants, or use explicit type cast operators, as in the a1 initialization above.

Parameters

a: lower inclusive boundary of the returned random number.

b: upper non-inclusive boundary of the returned random number.

Overloaded parameters

fn uniform_f64(&mut self, a: f64, b: f64) -> Result<f64>

returns uniformly distributed integer random number from [a,b) range

The methods transform the state using the MWC algorithm and return the next uniformly-distributed random number of the specified type, deduced from the input parameter type, from the range [a, b) . There is a nuance illustrated by the following sample:

RNG rng;
 
// always produces 0
double a = rng.uniform(0, 1);
 
// produces double from [0, 1)
double a1 = rng.uniform((double)0, (double)1);
 
// produces float from [0, 1)
float b = rng.uniform(0.f, 1.f);
 
// produces double from [0, 1)
double c = rng.uniform(0., 1.);
 
// may cause compiler error because of ambiguity:
//  RNG::uniform(0, (int)0.999999)? or RNG::uniform((double)0, 0.99999)?
double d = rng.uniform(0, 0.999999);

The compiler does not take into account the type of the variable to which you assign the result of RNG::uniform . The only thing that matters to the compiler is the type of a and b parameters. So, if you want a floating-point random number, but the range boundaries are integer numbers, either put dots in the end, if they are constants, or use explicit type cast operators, as in the a1 initialization above.

Parameters

a: lower inclusive boundary of the returned random number.

b: upper non-inclusive boundary of the returned random number.

Overloaded parameters

fn fill( &mut self, mat: &mut impl ToInputOutputArray, dist_type: i32, a: &impl ToInputArray, b: &impl ToInputArray, saturate_range: bool ) -> Result<()>

Fills arrays with random numbers.

Parameters

mat: 2D or N-dimensional matrix; currently matrices with more than 4 channels are not supported by the methods, use Mat::reshape as a possible workaround.

distType: distribution type, RNG::UNIFORM or RNG::NORMAL.

a: first distribution parameter; in case of the uniform distribution, this is an inclusive lower boundary, in case of the normal distribution, this is a mean value.

b: second distribution parameter; in case of the uniform distribution, this is a non-inclusive upper boundary, in case of the normal distribution, this is a standard deviation (diagonal of the standard deviation matrix or the full standard deviation matrix).

saturateRange: pre-saturation flag; for uniform distribution only; if true, the method will first convert a and b to the acceptable value range (according to the mat datatype) and then will generate uniformly distributed random numbers within the range [saturate(a), saturate(b)), if saturateRange=false, the method will generate uniformly distributed random numbers in the original range [a, b) and then will saturate them, it means, for example, that theRNG().fill(mat_8u, RNG::UNIFORM, -DBL_MAX, DBL_MAX) will likely produce array mostly filled with 0’s and 255’s, since the range (0, 255) is significantly smaller than [-DBL_MAX, DBL_MAX).

Each of the methods fills the matrix with the random values from the specified distribution. As the new numbers are generated, the RNG state is updated accordingly. In case of multiple-channel images, every channel is filled independently, which means that RNG cannot generate samples from the multi-dimensional Gaussian distribution with non-diagonal covariance matrix directly. To do that, the method generates samples from multi-dimensional standard Gaussian distribution with zero mean and identity covariation matrix, and then transforms them using transform to get samples from the specified Gaussian distribution.

C++ default parameters

saturate_range: false

fn fill_def( &mut self, mat: &mut impl ToInputOutputArray, dist_type: i32, a: &impl ToInputArray, b: &impl ToInputArray ) -> Result<()>

Fills arrays with random numbers.

Parameters

mat: 2D or N-dimensional matrix; currently matrices with more than 4 channels are not supported by the methods, use Mat::reshape as a possible workaround.

distType: distribution type, RNG::UNIFORM or RNG::NORMAL.

a: first distribution parameter; in case of the uniform distribution, this is an inclusive lower boundary, in case of the normal distribution, this is a mean value.

b: second distribution parameter; in case of the uniform distribution, this is a non-inclusive upper boundary, in case of the normal distribution, this is a standard deviation (diagonal of the standard deviation matrix or the full standard deviation matrix).

saturateRange: pre-saturation flag; for uniform distribution only; if true, the method will first convert a and b to the acceptable value range (according to the mat datatype) and then will generate uniformly distributed random numbers within the range [saturate(a), saturate(b)), if saturateRange=false, the method will generate uniformly distributed random numbers in the original range [a, b) and then will saturate them, it means, for example, that theRNG().fill(mat_8u, RNG::UNIFORM, -DBL_MAX, DBL_MAX) will likely produce array mostly filled with 0’s and 255’s, since the range (0, 255) is significantly smaller than [-DBL_MAX, DBL_MAX).

Each of the methods fills the matrix with the random values from the specified distribution. As the new numbers are generated, the RNG state is updated accordingly. In case of multiple-channel images, every channel is filled independently, which means that RNG cannot generate samples from the multi-dimensional Gaussian distribution with non-diagonal covariance matrix directly. To do that, the method generates samples from multi-dimensional standard Gaussian distribution with zero mean and identity covariation matrix, and then transforms them using transform to get samples from the specified Gaussian distribution.

Note

This alternative version of RNGTrait::fill function uses the following default values for its arguments:

saturate_range: false

fn gaussian(&mut self, sigma: f64) -> Result<f64>

Returns the next random number sampled from the Gaussian distribution

Parameters

sigma: standard deviation of the distribution.

The method transforms the state using the MWC algorithm and returns the next random number from the Gaussian distribution N(0,sigma) . That is, the mean value of the returned random numbers is zero and the standard deviation is the specified sigma .

Object Safety

This trait is not object safe.