Trait opencv::hub_prelude::UMatTrait

pub trait UMatTrait: UMatTraitConst {

    // Required method
    fn as_raw_mut_UMat(&mut self) -> *mut c_void;

    // Provided methods
    fn set_flags(&mut self, val: i32) { ... }
    fn set_dims(&mut self, val: i32) { ... }
    fn set_rows(&mut self, val: i32) { ... }
    fn set_cols(&mut self, val: i32) { ... }
    fn set_usage_flags(&mut self, val: UMatUsageFlags) { ... }
    fn u(&mut self) -> UMatData { ... }
    fn set_u(&mut self, val: &impl UMatDataTraitConst) { ... }
    fn set_offset(&mut self, val: size_t) { ... }
    fn set_size(&mut self, val: MatSize) { ... }
    fn set(&mut self, m: &impl UMatTraitConst) -> Result<()> { ... }
    fn row_mut(&mut self, y: i32) -> Result<BoxedRefMut<'_, UMat>> { ... }
    fn col_mut(&mut self, x: i32) -> Result<BoxedRefMut<'_, UMat>> { ... }
    fn row_bounds_mut(
        &mut self,
        startrow: i32,
        endrow: i32
    ) -> Result<BoxedRefMut<'_, UMat>> { ... }
    fn row_range_mut(
        &mut self,
        r: &impl RangeTraitConst
    ) -> Result<BoxedRefMut<'_, UMat>> { ... }
    fn col_bounds_mut(
        &mut self,
        startcol: i32,
        endcol: i32
    ) -> Result<BoxedRefMut<'_, UMat>> { ... }
    fn col_range_mut(
        &mut self,
        r: &impl RangeTraitConst
    ) -> Result<BoxedRefMut<'_, UMat>> { ... }
    fn diag_def_mut(&mut self) -> Result<BoxedRefMut<'_, UMat>> { ... }
    fn diag_mut(&mut self, d: i32) -> Result<BoxedRefMut<'_, UMat>> { ... }
    fn set_scalar(&mut self, s: Scalar) -> Result<()> { ... }
    fn set_to(
        &mut self,
        value: &impl ToInputArray,
        mask: &impl ToInputArray
    ) -> Result<UMat> { ... }
    fn set_to_def(&mut self, value: &impl ToInputArray) -> Result<UMat> { ... }
    fn reshape_def_mut(&mut self, cn: i32) -> Result<BoxedRefMut<'_, UMat>> { ... }
    fn reshape_mut(
        &mut self,
        cn: i32,
        rows: i32
    ) -> Result<BoxedRefMut<'_, UMat>> { ... }
    fn reshape_nd_mut(
        &mut self,
        cn: i32,
        newsz: &[i32]
    ) -> Result<BoxedRefMut<'_, UMat>> { ... }
    unsafe fn create_rows_cols(
        &mut self,
        rows: i32,
        cols: i32,
        typ: i32,
        usage_flags: UMatUsageFlags
    ) -> Result<()> { ... }
    fn create_rows_cols_def(
        &mut self,
        rows: i32,
        cols: i32,
        typ: i32
    ) -> Result<()> { ... }
    unsafe fn create_size(
        &mut self,
        size: Size,
        typ: i32,
        usage_flags: UMatUsageFlags
    ) -> Result<()> { ... }
    fn create_size_def(&mut self, size: Size, typ: i32) -> Result<()> { ... }
    unsafe fn create_nd(
        &mut self,
        sizes: &[i32],
        typ: i32,
        usage_flags: UMatUsageFlags
    ) -> Result<()> { ... }
    fn create_nd_def(&mut self, sizes: &[i32], typ: i32) -> Result<()> { ... }
    unsafe fn create_nd_vec(
        &mut self,
        sizes: &Vector<i32>,
        typ: i32,
        usage_flags: UMatUsageFlags
    ) -> Result<()> { ... }
    fn create_nd_vec_def(&mut self, sizes: &Vector<i32>, typ: i32) -> Result<()> { ... }
    unsafe fn addref(&mut self) -> Result<()> { ... }
    unsafe fn release(&mut self) -> Result<()> { ... }
    fn deallocate(&mut self) -> Result<()> { ... }
    fn adjust_roi(
        &mut self,
        dtop: i32,
        dbottom: i32,
        dleft: i32,
        dright: i32
    ) -> Result<UMat> { ... }
    fn rowscols_mut(
        &mut self,
        row_range: impl RangeTrait,
        col_range: impl RangeTrait
    ) -> Result<BoxedRefMut<'_, UMat>> { ... }
    fn roi_mut(&mut self, roi: Rect) -> Result<BoxedRefMut<'_, UMat>> { ... }
    fn ranges_mut(
        &mut self,
        ranges: &Vector<Range>
    ) -> Result<BoxedRefMut<'_, UMat>> { ... }
    fn set_1(&mut self, m: UMat) -> Result<()> { ... }
    fn update_continuity_flag(&mut self) -> Result<()> { ... }

}

Mutable methods for core::UMat

Required Methods

fn as_raw_mut_UMat(&mut self) -> *mut c_void

Provided Methods

fn set_flags(&mut self, val: i32)

! includes several bit-fields:

the magic signature

continuity flag

depth

number of channels

fn set_dims(&mut self, val: i32)

the matrix dimensionality, >= 2

fn set_rows(&mut self, val: i32)

number of rows in the matrix; -1 when the matrix has more than 2 dimensions

fn set_cols(&mut self, val: i32)

number of columns in the matrix; -1 when the matrix has more than 2 dimensions

fn set_usage_flags(&mut self, val: UMatUsageFlags)

usage flags for allocator; recommend do not set directly, instead set during construct/create/getUMat

fn u(&mut self) -> UMatData

black-box container of UMat data

fn set_u(&mut self, val: &impl UMatDataTraitConst)

black-box container of UMat data

fn set_offset(&mut self, val: size_t)

offset of the submatrix (or 0)

fn set_size(&mut self, val: MatSize)

dimensional size of the matrix; accessible in various formats

fn set(&mut self, m: &impl UMatTraitConst) -> Result<()>

assignment operators

fn row_mut(&mut self, y: i32) -> Result<BoxedRefMut<'_, UMat>>

returns a new matrix header for the specified row

fn col_mut(&mut self, x: i32) -> Result<BoxedRefMut<'_, UMat>>

returns a new matrix header for the specified column

fn row_bounds_mut( &mut self, startrow: i32, endrow: i32 ) -> Result<BoxedRefMut<'_, UMat>>

… for the specified row span

fn row_range_mut( &mut self, r: &impl RangeTraitConst ) -> Result<BoxedRefMut<'_, UMat>>

fn col_bounds_mut( &mut self, startcol: i32, endcol: i32 ) -> Result<BoxedRefMut<'_, UMat>>

… for the specified column span

fn col_range_mut( &mut self, r: &impl RangeTraitConst ) -> Result<BoxedRefMut<'_, UMat>>

fn diag_def_mut(&mut self) -> Result<BoxedRefMut<'_, UMat>>

… for the specified diagonal (d=0 - the main diagonal, >0 - a diagonal from the upper half, <0 - a diagonal from the lower half)

Note

This alternative version of UMatTraitConst::diag function uses the following default values for its arguments:

d: 0

fn diag_mut(&mut self, d: i32) -> Result<BoxedRefMut<'_, UMat>>

… for the specified diagonal (d=0 - the main diagonal, >0 - a diagonal from the upper half, <0 - a diagonal from the lower half)

C++ default parameters

d: 0

fn set_scalar(&mut self, s: Scalar) -> Result<()>

sets every matrix element to s

fn set_to( &mut self, value: &impl ToInputArray, mask: &impl ToInputArray ) -> Result<UMat>

sets some of the matrix elements to s, according to the mask

C++ default parameters

mask: noArray()

fn set_to_def(&mut self, value: &impl ToInputArray) -> Result<UMat>

sets some of the matrix elements to s, according to the mask

Note

This alternative version of UMatTrait::set_to function uses the following default values for its arguments:

mask: noArray()

fn reshape_def_mut(&mut self, cn: i32) -> Result<BoxedRefMut<'_, UMat>>

creates alternative matrix header for the same data, with different

Note

This alternative version of UMatTraitConst::reshape function uses the following default values for its arguments:

rows: 0

fn reshape_mut(&mut self, cn: i32, rows: i32) -> Result<BoxedRefMut<'_, UMat>>

creates alternative matrix header for the same data, with different

C++ default parameters

rows: 0

fn reshape_nd_mut( &mut self, cn: i32, newsz: &[i32] ) -> Result<BoxedRefMut<'_, UMat>>

unsafe fn create_rows_cols( &mut self, rows: i32, cols: i32, typ: i32, usage_flags: UMatUsageFlags ) -> Result<()>

allocates new matrix data unless the matrix already has specified size and type.

C++ default parameters

usage_flags: USAGE_DEFAULT

fn create_rows_cols_def(&mut self, rows: i32, cols: i32, typ: i32) -> Result<()>

allocates new matrix data unless the matrix already has specified size and type.

Note

This alternative version of UMatTrait::create_rows_cols function uses the following default values for its arguments:

usage_flags: USAGE_DEFAULT

unsafe fn create_size( &mut self, size: Size, typ: i32, usage_flags: UMatUsageFlags ) -> Result<()>

C++ default parameters

usage_flags: USAGE_DEFAULT

fn create_size_def(&mut self, size: Size, typ: i32) -> Result<()>

Note

This alternative version of UMatTrait::create_size function uses the following default values for its arguments:

usage_flags: USAGE_DEFAULT

unsafe fn create_nd( &mut self, sizes: &[i32], typ: i32, usage_flags: UMatUsageFlags ) -> Result<()>

C++ default parameters

usage_flags: USAGE_DEFAULT

fn create_nd_def(&mut self, sizes: &[i32], typ: i32) -> Result<()>

Note

This alternative version of UMatTrait::create_nd function uses the following default values for its arguments:

usage_flags: USAGE_DEFAULT

unsafe fn create_nd_vec( &mut self, sizes: &Vector<i32>, typ: i32, usage_flags: UMatUsageFlags ) -> Result<()>

C++ default parameters

usage_flags: USAGE_DEFAULT

fn create_nd_vec_def(&mut self, sizes: &Vector<i32>, typ: i32) -> Result<()>

Note

This alternative version of UMatTrait::create_nd_vec function uses the following default values for its arguments:

usage_flags: USAGE_DEFAULT

unsafe fn addref(&mut self) -> Result<()>

increases the reference counter; use with care to avoid memleaks

unsafe fn release(&mut self) -> Result<()>

decreases reference counter;

fn deallocate(&mut self) -> Result<()>

deallocates the matrix data

fn adjust_roi( &mut self, dtop: i32, dbottom: i32, dleft: i32, dright: i32 ) -> Result<UMat>

moves/resizes the current matrix ROI inside the parent matrix.

fn rowscols_mut( &mut self, row_range: impl RangeTrait, col_range: impl RangeTrait ) -> Result<BoxedRefMut<'_, UMat>>

extracts a rectangular sub-matrix

fn roi_mut(&mut self, roi: Rect) -> Result<BoxedRefMut<'_, UMat>>

fn ranges_mut( &mut self, ranges: &Vector<Range> ) -> Result<BoxedRefMut<'_, UMat>>

fn set_1(&mut self, m: UMat) -> Result<()>

fn update_continuity_flag(&mut self) -> Result<()>

internal use method: updates the continuity flag

Object Safety

This trait is not object safe.