Module opencv::imgproc

Structs

FilterEngine

Moments
raster image moments

Subdiv2D

Constants

ADAPTIVE_THRESH_GAUSSIAN_C

ADAPTIVE_THRESH_MEAN_C

BORDER_CONSTANT

BORDER_DEFAULT

BORDER_ISOLATED

BORDER_REFLECT

BORDER_REFLECT101

BORDER_REFLECT_101

BORDER_REPLICATE

BORDER_TRANSPARENT

BORDER_WRAP

CHAIN_APPROX_NONE

CHAIN_APPROX_SIMPLE

CHAIN_APPROX_TC89_KCOS

CHAIN_APPROX_TC89_L1

COLOR_BGR2BGR555

COLOR_BGR2BGR565

COLOR_BGR2BGRA

COLOR_BGR2GRAY

COLOR_BGR2HLS

COLOR_BGR2HLS_FULL

COLOR_BGR2HSV

COLOR_BGR2HSV_FULL

COLOR_BGR2Lab

COLOR_BGR2Luv

COLOR_BGR2RGB

COLOR_BGR2RGBA

COLOR_BGR2XYZ

COLOR_BGR2YCrCb

COLOR_BGR2YUV

COLOR_BGR2YUV_I420

COLOR_BGR2YUV_IYUV

COLOR_BGR2YUV_YV12

COLOR_BGR5552BGR

COLOR_BGR5552BGRA

COLOR_BGR5552GRAY

COLOR_BGR5552RGB

COLOR_BGR5552RGBA

COLOR_BGR5652BGR

COLOR_BGR5652BGRA

COLOR_BGR5652GRAY

COLOR_BGR5652RGB

COLOR_BGR5652RGBA

COLOR_BGRA2BGR

COLOR_BGRA2BGR555

COLOR_BGRA2BGR565

COLOR_BGRA2GRAY

COLOR_BGRA2RGB

COLOR_BGRA2RGBA

COLOR_BGRA2YUV_I420

COLOR_BGRA2YUV_IYUV

COLOR_BGRA2YUV_YV12

COLOR_BayerBG2BGR

COLOR_BayerBG2BGR_VNG

COLOR_BayerBG2GRAY

COLOR_BayerBG2RGB

COLOR_BayerBG2RGB_VNG

COLOR_BayerGB2BGR

COLOR_BayerGB2BGR_VNG

COLOR_BayerGB2GRAY

COLOR_BayerGB2RGB

COLOR_BayerGB2RGB_VNG

COLOR_BayerGR2BGR

COLOR_BayerGR2BGR_VNG

COLOR_BayerGR2GRAY

COLOR_BayerGR2RGB

COLOR_BayerGR2RGB_VNG

COLOR_BayerRG2BGR

COLOR_BayerRG2BGR_VNG

COLOR_BayerRG2GRAY

COLOR_BayerRG2RGB

COLOR_BayerRG2RGB_VNG

COLOR_COLORCVT_MAX

COLOR_GRAY2BGR

COLOR_GRAY2BGR555

COLOR_GRAY2BGR565

COLOR_GRAY2BGRA

COLOR_GRAY2RGB

COLOR_GRAY2RGBA

COLOR_HLS2BGR

COLOR_HLS2BGR_FULL

COLOR_HLS2RGB

COLOR_HLS2RGB_FULL

COLOR_HSV2BGR

COLOR_HSV2BGR_FULL

COLOR_HSV2RGB

COLOR_HSV2RGB_FULL

COLOR_LBGR2Lab

COLOR_LBGR2Luv

COLOR_LRGB2Lab

COLOR_LRGB2Luv

COLOR_Lab2BGR

COLOR_Lab2LBGR

COLOR_Lab2LRGB

COLOR_Lab2RGB

COLOR_Luv2BGR

COLOR_Luv2LBGR

COLOR_Luv2LRGB

COLOR_Luv2RGB

COLOR_RGB2BGR

COLOR_RGB2BGR555

COLOR_RGB2BGR565

COLOR_RGB2BGRA

COLOR_RGB2GRAY

COLOR_RGB2HLS

COLOR_RGB2HLS_FULL

COLOR_RGB2HSV

COLOR_RGB2HSV_FULL

COLOR_RGB2Lab

COLOR_RGB2Luv

COLOR_RGB2RGBA

COLOR_RGB2XYZ

COLOR_RGB2YCrCb

COLOR_RGB2YUV

COLOR_RGB2YUV_I420

COLOR_RGB2YUV_IYUV

COLOR_RGB2YUV_YV12

COLOR_RGBA2BGR

COLOR_RGBA2BGR555

COLOR_RGBA2BGR565

COLOR_RGBA2BGRA

COLOR_RGBA2GRAY

COLOR_RGBA2RGB

COLOR_RGBA2YUV_I420

COLOR_RGBA2YUV_IYUV

COLOR_RGBA2YUV_YV12

COLOR_RGBA2mRGBA

COLOR_XYZ2BGR

COLOR_XYZ2RGB

COLOR_YCrCb2BGR

COLOR_YCrCb2RGB

COLOR_YUV2BGR

COLOR_YUV2BGRA_I420

COLOR_YUV2BGRA_IYUV

COLOR_YUV2BGRA_NV12

COLOR_YUV2BGRA_NV21

COLOR_YUV2BGRA_UYNV

COLOR_YUV2BGRA_UYVY

COLOR_YUV2BGRA_Y422

COLOR_YUV2BGRA_YUNV

COLOR_YUV2BGRA_YUY2

COLOR_YUV2BGRA_YUYV

COLOR_YUV2BGRA_YV12

COLOR_YUV2BGRA_YVYU

COLOR_YUV2BGR_I420

COLOR_YUV2BGR_IYUV

COLOR_YUV2BGR_NV12

COLOR_YUV2BGR_NV21

COLOR_YUV2BGR_UYNV

COLOR_YUV2BGR_UYVY

COLOR_YUV2BGR_Y422

COLOR_YUV2BGR_YUNV

COLOR_YUV2BGR_YUY2

COLOR_YUV2BGR_YUYV

COLOR_YUV2BGR_YV12

COLOR_YUV2BGR_YVYU

COLOR_YUV2GRAY_420

COLOR_YUV2GRAY_I420

COLOR_YUV2GRAY_IYUV

COLOR_YUV2GRAY_NV12

COLOR_YUV2GRAY_NV21

COLOR_YUV2GRAY_UYNV

COLOR_YUV2GRAY_UYVY

COLOR_YUV2GRAY_Y422

COLOR_YUV2GRAY_YUNV

COLOR_YUV2GRAY_YUY2

COLOR_YUV2GRAY_YUYV

COLOR_YUV2GRAY_YV12

COLOR_YUV2GRAY_YVYU

COLOR_YUV2RGB

COLOR_YUV2RGBA_I420

COLOR_YUV2RGBA_IYUV

COLOR_YUV2RGBA_NV12

COLOR_YUV2RGBA_NV21

COLOR_YUV2RGBA_UYNV

COLOR_YUV2RGBA_UYVY

COLOR_YUV2RGBA_Y422

COLOR_YUV2RGBA_YUNV

COLOR_YUV2RGBA_YUY2

COLOR_YUV2RGBA_YUYV

COLOR_YUV2RGBA_YV12

COLOR_YUV2RGBA_YVYU

COLOR_YUV2RGB_I420

COLOR_YUV2RGB_IYUV

COLOR_YUV2RGB_NV12

COLOR_YUV2RGB_NV21

COLOR_YUV2RGB_UYNV

COLOR_YUV2RGB_UYVY

COLOR_YUV2RGB_Y422

COLOR_YUV2RGB_YUNV

COLOR_YUV2RGB_YUY2

COLOR_YUV2RGB_YUYV

COLOR_YUV2RGB_YV12

COLOR_YUV2RGB_YVYU

COLOR_YUV420p2BGR

COLOR_YUV420p2BGRA

COLOR_YUV420p2GRAY

COLOR_YUV420p2RGB

COLOR_YUV420p2RGBA

COLOR_YUV420sp2BGR

COLOR_YUV420sp2BGRA

COLOR_YUV420sp2GRAY

COLOR_YUV420sp2RGB

COLOR_YUV420sp2RGBA

COLOR_mRGBA2RGBA

CV_ADAPTIVE_THRESH_GAUSSIAN_C

CV_ADAPTIVE_THRESH_MEAN_C

CV_BGR2BGR555

CV_BGR2BGR565

CV_BGR2BGRA

CV_BGR2GRAY

CV_BGR2HLS

CV_BGR2HLS_FULL

CV_BGR2HSV

CV_BGR2HSV_FULL

CV_BGR2Lab

CV_BGR2Luv

CV_BGR2RGB

CV_BGR2RGBA

CV_BGR2XYZ

CV_BGR2YCrCb

CV_BGR2YUV

CV_BGR2YUV_I420

CV_BGR2YUV_IYUV

CV_BGR2YUV_YV12

CV_BGR5552BGR

CV_BGR5552BGRA

CV_BGR5552GRAY

CV_BGR5552RGB

CV_BGR5552RGBA

CV_BGR5652BGR

CV_BGR5652BGRA

CV_BGR5652GRAY

CV_BGR5652RGB

CV_BGR5652RGBA

CV_BGRA2BGR

CV_BGRA2BGR555

CV_BGRA2BGR565

CV_BGRA2GRAY

CV_BGRA2RGB

CV_BGRA2RGBA

CV_BGRA2YUV_I420

CV_BGRA2YUV_IYUV

CV_BGRA2YUV_YV12

CV_BILATERAL

CV_BLUR

CV_BLUR_NO_SCALE

CV_BayerBG2BGR

CV_BayerBG2BGR_VNG

CV_BayerBG2GRAY

CV_BayerBG2RGB

CV_BayerBG2RGB_VNG

CV_BayerGB2BGR

CV_BayerGB2BGR_VNG

CV_BayerGB2GRAY

CV_BayerGB2RGB

CV_BayerGB2RGB_VNG

CV_BayerGR2BGR

CV_BayerGR2BGR_VNG

CV_BayerGR2GRAY

CV_BayerGR2RGB

CV_BayerGR2RGB_VNG

CV_BayerRG2BGR

CV_BayerRG2BGR_VNG

CV_BayerRG2GRAY

CV_BayerRG2RGB

CV_BayerRG2RGB_VNG

CV_CANNY_L2_GRADIENT

CV_CHAIN_APPROX_NONE

CV_CHAIN_APPROX_SIMPLE

CV_CHAIN_APPROX_TC89_KCOS

CV_CHAIN_APPROX_TC89_L1

CV_CHAIN_CODE

CV_CLOCKWISE

CV_COLORCVT_MAX

CV_COMP_BHATTACHARYYA

CV_COMP_CHISQR

CV_COMP_CORREL

CV_COMP_HELLINGER

CV_COMP_INTERSECT

CV_CONTOURS_MATCH_I1

CV_CONTOURS_MATCH_I2

CV_CONTOURS_MATCH_I3

CV_COUNTER_CLOCKWISE

CV_DIST_C

CV_DIST_FAIR

CV_DIST_HUBER

CV_DIST_L1

CV_DIST_L12

CV_DIST_L2

CV_DIST_LABEL_CCOMP

CV_DIST_LABEL_PIXEL

CV_DIST_MASK_3

CV_DIST_MASK_5

CV_DIST_MASK_PRECISE

CV_DIST_USER

CV_DIST_WELSCH

CV_FLOODFILL_FIXED_RANGE

CV_FLOODFILL_MASK_ONLY

CV_GAUSSIAN

CV_GAUSSIAN_5x5

CV_GRAY2BGR

CV_GRAY2BGR555

CV_GRAY2BGR565

CV_GRAY2BGRA

CV_GRAY2RGB

CV_GRAY2RGBA

CV_HLS2BGR

CV_HLS2BGR_FULL

CV_HLS2RGB

CV_HLS2RGB_FULL

CV_HOUGH_GRADIENT

CV_HOUGH_MULTI_SCALE

CV_HOUGH_PROBABILISTIC

CV_HOUGH_STANDARD

CV_HSV2BGR

CV_HSV2BGR_FULL

CV_HSV2RGB

CV_HSV2RGB_FULL

CV_INTER_AREA

CV_INTER_CUBIC

CV_INTER_LANCZOS4

CV_INTER_LINEAR

CV_INTER_NN

CV_LBGR2Lab

CV_LBGR2Luv

CV_LINK_RUNS

CV_LRGB2Lab

CV_LRGB2Luv

CV_Lab2BGR

CV_Lab2LBGR

CV_Lab2LRGB

CV_Lab2RGB

CV_Luv2BGR

CV_Luv2LBGR

CV_Luv2LRGB

CV_Luv2RGB

CV_MAX_SOBEL_KSIZE

CV_MEDIAN

CV_MOP_BLACKHAT

CV_MOP_CLOSE

CV_MOP_DILATE

CV_MOP_ERODE

CV_MOP_GRADIENT

CV_MOP_OPEN

CV_MOP_TOPHAT

CV_POLY_APPROX_DP

CV_RETR_CCOMP

CV_RETR_EXTERNAL

CV_RETR_FLOODFILL

CV_RETR_LIST

CV_RETR_TREE

CV_RGB2BGR

CV_RGB2BGR555

CV_RGB2BGR565

CV_RGB2BGRA

CV_RGB2GRAY

CV_RGB2HLS

CV_RGB2HLS_FULL

CV_RGB2HSV

CV_RGB2HSV_FULL

CV_RGB2Lab

CV_RGB2Luv

CV_RGB2RGBA

CV_RGB2XYZ

CV_RGB2YCrCb

CV_RGB2YUV

CV_RGB2YUV_I420

CV_RGB2YUV_IYUV

CV_RGB2YUV_YV12

CV_RGBA2BGR

CV_RGBA2BGR555

CV_RGBA2BGR565

CV_RGBA2BGRA

CV_RGBA2GRAY

CV_RGBA2RGB

CV_RGBA2YUV_I420

CV_RGBA2YUV_IYUV

CV_RGBA2YUV_YV12

CV_RGBA2mRGBA

CV_SCHARR

CV_SHAPE_CROSS

CV_SHAPE_CUSTOM

CV_SHAPE_ELLIPSE

CV_SHAPE_RECT

CV_SUBDIV2D_VIRTUAL_POINT_FLAG

CV_THRESH_BINARY

CV_THRESH_BINARY_INV

CV_THRESH_MASK

CV_THRESH_OTSU

CV_THRESH_TOZERO

CV_THRESH_TOZERO_INV

CV_THRESH_TRUNC

CV_TM_CCOEFF

CV_TM_CCOEFF_NORMED

CV_TM_CCORR

CV_TM_CCORR_NORMED

CV_TM_SQDIFF

CV_TM_SQDIFF_NORMED

CV_WARP_FILL_OUTLIERS

CV_WARP_INVERSE_MAP

CV_XYZ2BGR

CV_XYZ2RGB

CV_YCrCb2BGR

CV_YCrCb2RGB

CV_YUV2BGR

CV_YUV2BGRA_I420

CV_YUV2BGRA_IYUV

CV_YUV2BGRA_NV12

CV_YUV2BGRA_NV21

CV_YUV2BGRA_UYNV

CV_YUV2BGRA_UYVY

CV_YUV2BGRA_Y422

CV_YUV2BGRA_YUNV

CV_YUV2BGRA_YUY2

CV_YUV2BGRA_YUYV

CV_YUV2BGRA_YV12

CV_YUV2BGRA_YVYU

CV_YUV2BGR_I420

CV_YUV2BGR_IYUV

CV_YUV2BGR_NV12

CV_YUV2BGR_NV21

CV_YUV2BGR_UYNV

CV_YUV2BGR_UYVY

CV_YUV2BGR_Y422

CV_YUV2BGR_YUNV

CV_YUV2BGR_YUY2

CV_YUV2BGR_YUYV

CV_YUV2BGR_YV12

CV_YUV2BGR_YVYU

CV_YUV2GRAY_420

CV_YUV2GRAY_I420

CV_YUV2GRAY_IYUV

CV_YUV2GRAY_NV12

CV_YUV2GRAY_NV21

CV_YUV2GRAY_UYNV

CV_YUV2GRAY_UYVY

CV_YUV2GRAY_Y422

CV_YUV2GRAY_YUNV

CV_YUV2GRAY_YUY2

CV_YUV2GRAY_YUYV

CV_YUV2GRAY_YV12

CV_YUV2GRAY_YVYU

CV_YUV2RGB

CV_YUV2RGBA_I420

CV_YUV2RGBA_IYUV

CV_YUV2RGBA_NV12

CV_YUV2RGBA_NV21

CV_YUV2RGBA_UYNV

CV_YUV2RGBA_UYVY

CV_YUV2RGBA_Y422

CV_YUV2RGBA_YUNV

CV_YUV2RGBA_YUY2

CV_YUV2RGBA_YUYV

CV_YUV2RGBA_YV12

CV_YUV2RGBA_YVYU

CV_YUV2RGB_I420

CV_YUV2RGB_IYUV

CV_YUV2RGB_NV12

CV_YUV2RGB_NV21

CV_YUV2RGB_UYNV

CV_YUV2RGB_UYVY

CV_YUV2RGB_Y422

CV_YUV2RGB_YUNV

CV_YUV2RGB_YUY2

CV_YUV2RGB_YUYV

CV_YUV2RGB_YV12

CV_YUV2RGB_YVYU

CV_YUV420p2BGR

CV_YUV420p2BGRA

CV_YUV420p2GRAY

CV_YUV420p2RGB

CV_YUV420p2RGBA

CV_YUV420sp2BGR

CV_YUV420sp2BGRA

CV_YUV420sp2GRAY

CV_YUV420sp2RGB

CV_YUV420sp2RGBA

CV_mRGBA2RGBA

DIST_LABEL_CCOMP

DIST_LABEL_PIXEL

FLOODFILL_FIXED_RANGE

FLOODFILL_MASK_ONLY

GC_BGD

GC_EVAL

GC_FGD

GC_INIT_WITH_MASK

GC_INIT_WITH_RECT

GC_PR_BGD

GC_PR_FGD

GHT_POSITION

GHT_ROTATION

GHT_SCALE

INTER_AREA

INTER_BITS

INTER_BITS2

INTER_CUBIC

INTER_LANCZOS4

INTER_LINEAR

INTER_MAX

INTER_NEAREST

INTER_TAB_SIZE

INTER_TAB_SIZE2

KERNEL_ASYMMETRICAL

KERNEL_GENERAL

KERNEL_INTEGER

KERNEL_SMOOTH

KERNEL_SYMMETRICAL

MORPH_BLACKHAT

MORPH_CLOSE

MORPH_CROSS

MORPH_DILATE

MORPH_ELLIPSE

MORPH_ERODE

MORPH_GRADIENT

MORPH_OPEN

MORPH_RECT

MORPH_TOPHAT

PROJ_SPHERICAL_EQRECT

PROJ_SPHERICAL_ORTHO

RETR_CCOMP

RETR_EXTERNAL

RETR_FLOODFILL

RETR_LIST

RETR_TREE

Subdiv2D_NEXT_AROUND_DST

Subdiv2D_NEXT_AROUND_LEFT

Subdiv2D_NEXT_AROUND_ORG

Subdiv2D_NEXT_AROUND_RIGHT

Subdiv2D_PREV_AROUND_DST

Subdiv2D_PREV_AROUND_LEFT

Subdiv2D_PREV_AROUND_ORG

Subdiv2D_PREV_AROUND_RIGHT

Subdiv2D_PTLOC_ERROR

Subdiv2D_PTLOC_INSIDE

Subdiv2D_PTLOC_ON_EDGE

Subdiv2D_PTLOC_OUTSIDE_RECT

Subdiv2D_PTLOC_VERTEX

THRESH_BINARY

THRESH_BINARY_INV

THRESH_MASK

THRESH_OTSU

THRESH_TOZERO

THRESH_TOZERO_INV

THRESH_TRUNC

TM_CCOEFF

TM_CCOEFF_NORMED

TM_CCORR

TM_CCORR_NORMED

TM_SQDIFF

TM_SQDIFF_NORMED

WARP_INVERSE_MAP

Traits

CLAHE

GeneralizedHough

Functions

adaptive_bilateral_filter
smooths the image using adaptive bilateral filter

adaptive_threshold
applies variable (adaptive) threshold to the image

approx_poly_dp
approximates contour or a curve using Douglas-Peucker algorithm

arc_length
computes the contour perimeter (closed=true) or a curve length

bilateral_filter
smooths the image using bilateral filter

blur
a synonym for normalized box filter

border_interpolate
1D interpolation function: returns coordinate of the "donor" pixel for the specified location p.

bounding_rect
computes the bounding rectangle for a contour

box_filter
smooths the image using the box filter. Each pixel is processed in O(1) time

canny
applies Canny edge detector and produces the edge map.

compare_hist
compares two histograms stored in dense arrays

contour_area
computes the contour area

convert_maps
converts maps for remap from floating-point to fixed-point format or backwards

convex_hull
computes convex hull for a set of 2D points.

convexity_defects
computes the contour convexity defects

copy_make_border
copies 2D array to a larger destination array with extrapolation of the outer part of src using the specified border mode

corner_eigen_vals_and_vecs
computes both eigenvalues and the eigenvectors of 2x2 derivative covariation matrix at each pixel. The output is stored as 6-channel matrix.

corner_harris
computes Harris cornerness criteria at each image pixel

corner_min_eigen_val
computes minimum eigen value of 2x2 derivative covariation matrix at each pixel - the cornerness criteria

create_box_filter
returns box filter engine

create_deriv_filter
returns filter engine for the generalized Sobel operator

create_gaussian_filter
returns the Gaussian filter engine

create_hanning_window

create_linear_filter
returns the non-separable linear filter engine

create_morphology_filter
returns morphological filter engine. Only MORPH_ERODE and MORPH_DILATE are supported.

create_separable_linear_filter
returns the separable linear filter engine

cvt_color
converts image from one color space to another

dilate
dilates the image (applies the local maximum operator)

distance_transform
computes the distance transform map

distance_transform_labels
builds the discrete Voronoi diagram

equalize_hist
normalizes the grayscale image brightness and contrast by normalizing its histogram

erode
erodes the image (applies the local minimum operator)

filter2_d
applies non-separable 2D linear filter to the image

fit_line
fits line to the set of 2D points using M-estimator algorithm

gaussian_blur
smooths the image using Gaussian filter.

get_affine_transform

get_default_new_camera_matrix
returns the default new camera matrix (by default it is the same as cameraMatrix unless centerPricipalPoint=true)

get_deriv_kernels
initializes kernels of the generalized Sobel operator

get_gabor_kernel
returns the Gabor kernel with the specified parameters

get_gaussian_kernel
returns the Gaussian kernel with the specified parameters

get_kernel_type
returns type (one of KERNEL_*) of 1D or 2D kernel specified by its coefficients.

get_perspective_transform

get_rect_sub_pix
extracts rectangle from the image at sub-pixel location

get_rotation_matrix2_d
returns 2x3 affine transformation matrix for the planar rotation.

get_structuring_element
returns structuring element of the specified shape and size

good_features_to_track
finds the strong enough corners where the cornerMinEigenVal() or cornerHarris() report the local maxima

hough_circles
finds circles in the grayscale image using 2+1 gradient Hough transform

hough_lines
finds lines in the black-n-white image using the standard or pyramid Hough transform

hough_lines_p
finds line segments in the black-n-white image using probabilistic Hough transform

hu_moments

init_undistort_rectify_map
initializes maps for cv::remap() to correct lens distortion and optionally rectify the image

init_wide_angle_proj_map
initializes maps for cv::remap() for wide-angle

integral
computes the integral image

integral_sq
computes the integral image and integral for the squared image

integral_tilted
computes the integral image, integral for the squared image and the tilted integral image

intersect_convex_convex
finds intersection of two convex polygons

invert_affine_transform
computes 2x3 affine transformation matrix that is inverse to the specified 2x3 affine transformation.

is_contour_convex
returns true if the contour is convex. Does not support contours with self-intersection

laplacian
applies Laplacian operator to the image

match_shapes
matches two contours using one of the available algorithms

match_template
computes the proximity map for the raster template and the image where the template is searched for

median_blur
smooths the image using median filter.

min_enclosing_circle
computes the minimal enclosing circle for a set of points

moments
computes moments of the rasterized shape or a vector of points

morphology_default_border_value
returns "magic" border value for erosion and dilation. It is automatically transformed to Scalar::all(-DBL_MAX) for dilation.

morphology_ex
applies an advanced morphological operation to the image

phase_correlate

point_polygon_test
checks if the point is inside the contour. Optionally computes the signed distance from the point to the contour boundary

pre_corner_detect
computes another complex cornerness criteria at each pixel

psnr
computes PSNR image/video quality metric

pyr_down
smooths and downsamples the image

pyr_up
upsamples and smoothes the image

remap
warps the image using the precomputed maps. The maps are stored in either floating-point or integer fixed-point format

resize
resizes the image

scharr
applies the vertical or horizontal Scharr operator to the image

sep_filter2_d
applies separable 2D linear filter to the image

sobel
applies generalized Sobel operator to the image

threshold
applies fixed threshold to the image

undistort
corrects lens distortion for the given camera matrix and distortion coefficients

undistort_points
returns points' coordinates after lens distortion correction

warp_affine
warps the image using affine transformation

warp_perspective
warps the image using perspective transformation