Function opencv::imgproc::arrowed_line_def

pub fn arrowed_line_def(
    img: &mut impl ToInputOutputArray,
    pt1: Point,
    pt2: Point,
    color: Scalar
) -> Result<()>

Draws an arrow segment pointing from the first point to the second one.

The function cv::arrowedLine draws an arrow between pt1 and pt2 points in the image. See also #line.

Parameters

img: Image.

pt1: The point the arrow starts from.

pt2: The point the arrow points to.

color: Line color.

thickness: Line thickness.

line_type: Type of the line. See [line_types]

shift: Number of fractional bits in the point coordinates.

tipLength: The length of the arrow tip in relation to the arrow length

Note

This alternative version of arrowed_line function uses the following default values for its arguments:

thickness: 1

line_type: 8

shift: 0

tip_length: 0.1