opencv::imgproc
Function build_pyramid_def

pub fn build_pyramid_def(
    src: &impl ToInputArray,
    dst: &mut impl ToOutputArray,
    maxlevel: i32,
) -> Result<()>

Constructs the Gaussian pyramid for an image.

The function constructs a vector of images and builds the Gaussian pyramid by recursively applying pyrDown to the previously built pyramid layers, starting from dst[0]==src.

Parameters

src: Source image. Check pyrDown for the list of supported types.

dst: Destination vector of maxlevel+1 images of the same type as src. dst[0] will be the same as src. dst[1] is the next pyramid layer, a smoothed and down-sized src, and so on.

maxlevel: 0-based index of the last (the smallest) pyramid layer. It must be non-negative.

borderType: Pixel extrapolation method, see [border_types] ([BORDER_CONSTANT] isn’t supported)

Note

This alternative version of build_pyramid function uses the following default values for its arguments:

border_type: BORDER_DEFAULT