Function opencv::imgproc::connected_components_with_stats

pub fn connected_components_with_stats(
    image: &impl ToInputArray,
    labels: &mut impl ToOutputArray,
    stats: &mut impl ToOutputArray,
    centroids: &mut impl ToOutputArray,
    connectivity: i32,
    ltype: i32,
) -> Result<i32>

computes the connected components labeled image of boolean image and also produces a statistics output for each label

image with 4 or 8 way connectivity - returns N, the total number of labels [0, N-1] where 0 represents the background label. ltype specifies the output label image type, an important consideration based on the total number of labels or alternatively the total number of pixels in the source image. ccltype specifies the connected components labeling algorithm to use, currently Grana’s (BBDT) and Wu’s (SAUF) Wu2009 algorithms are supported, see the [connected_components_algorithms_types] for details. Note that SAUF algorithm forces a row major ordering of labels while BBDT does not. This function uses parallel version of both Grana and Wu’s algorithms (statistics included) if at least one allowed parallel framework is enabled and if the rows of the image are at least twice the number returned by #getNumberOfCPUs.

Parameters

image: the 8-bit single-channel image to be labeled

labels: destination labeled image

stats: statistics output for each label, including the background label. Statistics are accessed via stats(label, COLUMN) where COLUMN is one of #ConnectedComponentsTypes, selecting the statistic. The data type is CV_32S.

centroids: centroid output for each label, including the background label. Centroids are accessed via centroids(label, 0) for x and centroids(label, 1) for y. The data type CV_64F.

connectivity: 8 or 4 for 8-way or 4-way connectivity respectively

ltype: output image label type. Currently CV_32S and CV_16U are supported.

ccltype: connected components algorithm type (see #ConnectedComponentsAlgorithmsTypes).

Overloaded parameters

image: the 8-bit single-channel image to be labeled

labels: destination labeled image

stats: statistics output for each label, including the background label. Statistics are accessed via stats(label, COLUMN) where COLUMN is one of #ConnectedComponentsTypes, selecting the statistic. The data type is CV_32S.

centroids: centroid output for each label, including the background label. Centroids are accessed via centroids(label, 0) for x and centroids(label, 1) for y. The data type CV_64F.

connectivity: 8 or 4 for 8-way or 4-way connectivity respectively

ltype: output image label type. Currently CV_32S and CV_16U are supported.

C++ default parameters

connectivity: 8

ltype: CV_32S