Function opencv::imgproc::create_line_segment_detector_def

pub fn create_line_segment_detector_def() -> Result<Ptr<LineSegmentDetector>>

Creates a smart pointer to a LineSegmentDetector object and initializes it.

The LineSegmentDetector algorithm is defined using the standard values. Only advanced users may want to edit those, as to tailor it for their own application.

Parameters

refine: The way found lines will be refined, see [line_segment_detector_modes]

scale: The scale of the image that will be used to find the lines. Range (0..1].

sigma_scale: Sigma for Gaussian filter. It is computed as sigma = sigma_scale/scale.

quant: Bound to the quantization error on the gradient norm.

ang_th: Gradient angle tolerance in degrees.

log_eps: Detection threshold: -log10(NFA) > log_eps. Used only when advance refinement is chosen.

density_th: Minimal density of aligned region points in the enclosing rectangle.

n_bins: Number of bins in pseudo-ordering of gradient modulus.

Note

This alternative version of create_line_segment_detector function uses the following default values for its arguments:

refine: LSD_REFINE_STD

scale: 0.8

sigma_scale: 0.6

quant: 2.0

ang_th: 22.5

log_eps: 0

density_th: 0.7

n_bins: 1024