Function opencv::imgproc::demosaicing_def

pub fn demosaicing_def(
    src: &impl ToInputArray,
    dst: &mut impl ToOutputArray,
    code: i32,
) -> Result<()>

main function for all demosaicing processes

Parameters

src: input image: 8-bit unsigned or 16-bit unsigned.

dst: output image of the same size and depth as src.

code: Color space conversion code (see the description below).

dstCn: number of channels in the destination image; if the parameter is 0, the number of the channels is derived automatically from src and code.

The function can do the following transformations:

Demosaicing using bilinear interpolation

[color_bayer_bg2_bgr] , [color_bayer_gb2_bgr] , [color_bayer_rg2_bgr] , [color_bayer_gr2_bgr]

[color_bayer_bg2_gray] , [color_bayer_gb2_gray] , [color_bayer_rg2_gray] , [color_bayer_gr2_gray]

Demosaicing using Variable Number of Gradients.

[color_bayer_bg2_bgr_vng] , [color_bayer_gb2_bgr_vng] , [color_bayer_rg2_bgr_vng] , [color_bayer_gr2_bgr_vng]

Edge-Aware Demosaicing.

[color_bayer_bg2_bgr_ea] , [color_bayer_gb2_bgr_ea] , [color_bayer_rg2_bgr_ea] , [color_bayer_gr2_bgr_ea]

Demosaicing with alpha channel

[color_bayer_bg2_bgra] , [color_bayer_gb2_bgra] , [color_bayer_rg2_bgra] , [color_bayer_gr2_bgra]

See also

cvtColor

Note

This alternative version of demosaicing function uses the following default values for its arguments:

dst_cn: 0