Function opencv::imgproc::draw_marker

pub fn draw_marker(
    img: &mut impl ToInputOutputArray,
    position: Point,
    color: Scalar,
    marker_type: i32,
    marker_size: i32,
    thickness: i32,
    line_type: i32,
) -> Result<()>

Draws a marker on a predefined position in an image.

The function cv::drawMarker draws a marker on a given position in the image. For the moment several marker types are supported, see [marker_types] for more information.

Parameters

img: Image.

position: The point where the crosshair is positioned.

color: Line color.

markerType: The specific type of marker you want to use, see [marker_types]

thickness: Line thickness.

line_type: Type of the line, See [line_types]

markerSize: The length of the marker axis [default = 20 pixels]

C++ default parameters

marker_type: MARKER_CROSS

marker_size: 20

thickness: 1

line_type: 8