Function opencv::imgproc::draw_marker_def

pub fn draw_marker_def(
    img: &mut impl ToInputOutputArray,
    position: Point,
    color: Scalar
) -> Result<()>

Draws a marker on a predefined position in an image.

The function cv::drawMarker draws a marker on a given position in the image. For the moment several marker types are supported, see [marker_types] for more information.

Parameters

img: Image.

position: The point where the crosshair is positioned.

color: Line color.

markerType: The specific type of marker you want to use, see [marker_types]

thickness: Line thickness.

line_type: Type of the line, See [line_types]

markerSize: The length of the marker axis [default = 20 pixels]

Note

This alternative version of draw_marker function uses the following default values for its arguments:

marker_type: MARKER_CROSS

marker_size: 20

thickness: 1

line_type: 8