Function opencv::imgproc::find_contours_with_hierarchy_def

pub fn find_contours_with_hierarchy_def(
    image: &impl ToInputArray,
    contours: &mut impl ToOutputArray,
    hierarchy: &mut impl ToOutputArray,
    mode: i32,
    method: i32
) -> Result<()>

Finds contours in a binary image.

The function retrieves contours from the binary image using the algorithm Suzuki85 . The contours are a useful tool for shape analysis and object detection and recognition. See squares.cpp in the OpenCV sample directory.

Note: Since opencv 3.2 source image is not modified by this function.

Parameters

image: Source, an 8-bit single-channel image. Non-zero pixels are treated as 1’s. Zero pixels remain 0’s, so the image is treated as binary . You can use #compare, #inRange, threshold , #adaptiveThreshold, #Canny, and others to create a binary image out of a grayscale or color one. If mode equals to RETR_CCOMP or #RETR_FLOODFILL, the input can also be a 32-bit integer image of labels (CV_32SC1).

contours: Detected contours. Each contour is stored as a vector of points (e.g. std::vector<std::vectorcv::Point >).

hierarchy: Optional output vector (e.g. std::vectorcv::Vec4i), containing information about the image topology. It has as many elements as the number of contours. For each i-th contour contours[i], the elements hierarchy[i][0] , hierarchy[i][1] , hierarchy[i][2] , and hierarchy[i][3] are set to 0-based indices in contours of the next and previous contours at the same hierarchical level, the first child contour and the parent contour, respectively. If for the contour i there are no next, previous, parent, or nested contours, the corresponding elements of hierarchy[i] will be negative.

Note: In Python, hierarchy is nested inside a top level array. Use hierarchy[0][i] to access hierarchical elements of i-th contour.

mode: Contour retrieval mode, see [retrieval_modes]

method: Contour approximation method, see [contour_approximation_modes]

offset: Optional offset by which every contour point is shifted. This is useful if the contours are extracted from the image ROI and then they should be analyzed in the whole image context.

Note

This alternative version of find_contours_with_hierarchy function uses the following default values for its arguments:

offset: Point()