opencv::imgproc::prelude
Trait CLAHETrait

pub trait CLAHETrait: AlgorithmTrait + CLAHETraitConst {
    // Required method
    fn as_raw_mut_CLAHE(&mut self) -> *mut c_void;

    // Provided methods
    fn apply(
        &mut self,
        src: &impl ToInputArray,
        dst: &mut impl ToOutputArray,
    ) -> Result<()> { ... }
    fn set_clip_limit(&mut self, clip_limit: f64) -> Result<()> { ... }
    fn set_tiles_grid_size(&mut self, tile_grid_size: Size) -> Result<()> { ... }
    fn collect_garbage(&mut self) -> Result<()> { ... }
}

Mutable methods for crate::imgproc::CLAHE

Required Methods

fn as_raw_mut_CLAHE(&mut self) -> *mut c_void

Provided Methods

fn apply( &mut self, src: &impl ToInputArray, dst: &mut impl ToOutputArray, ) -> Result<()>

Equalizes the histogram of a grayscale image using Contrast Limited Adaptive Histogram Equalization.

Parameters

src: Source image of type CV_8UC1 or CV_16UC1.

dst: Destination image.

fn set_clip_limit(&mut self, clip_limit: f64) -> Result<()>

Sets threshold for contrast limiting.

Parameters

clipLimit: threshold value.

fn set_tiles_grid_size(&mut self, tile_grid_size: Size) -> Result<()>

Sets size of grid for histogram equalization. Input image will be divided into equally sized rectangular tiles.

Parameters

tileGridSize: defines the number of tiles in row and column.

fn collect_garbage(&mut self) -> Result<()>

Object Safety

This trait is not object safe.