Struct opencv::imgproc::IntelligentScissorsMB

pub struct IntelligentScissorsMB {
    ptr: *mut c_void,
}

Intelligent Scissors image segmentation

This class is used to find the path (contour) between two points which can be used for image segmentation.

Usage example: usage_example_intelligent_scissors

Reference: “Intelligent Scissors for Image Composition” algorithm designed by Eric N. Mortensen and William A. Barrett, Brigham Young University Mortensen95intelligentscissors

Fields
ptr: *mut c_void
Implementations

impl IntelligentScissorsMB

pub fn default() -> Result<IntelligentScissorsMB>