Trait opencv::manual::core::mat::MatTraitConstManual

pub trait MatTraitConstManual: MatTraitConst {

    // Provided methods
    unsafe fn at_unchecked<T: DataType>(&self, i0: i32) -> Result<&T> { ... }
    unsafe fn at_2d_unchecked<T: DataType>(
        &self,
        row: i32,
        col: i32
    ) -> Result<&T> { ... }
    unsafe fn at_pt_unchecked<T: DataType>(&self, pt: Point) -> Result<&T> { ... }
    unsafe fn at_3d_unchecked<T: DataType>(
        &self,
        i0: i32,
        i1: i32,
        i2: i32
    ) -> Result<&T> { ... }
    unsafe fn at_nd_unchecked<T: DataType>(&self, idx: &[i32]) -> Result<&T> { ... }
    fn at_row<T: DataType>(&self, row: i32) -> Result<&[T]> { ... }
    unsafe fn at_row_unchecked<T: DataType>(&self, row: i32) -> Result<&[T]> { ... }
    fn is_allocated(&self) -> bool { ... }
    fn data_bytes(&self) -> Result<&[u8]> { ... }
    fn data_typed<T: DataType>(&self) -> Result<&[T]> { ... }
    unsafe fn data_typed_unchecked<T: DataType>(&self) -> Result<&[T]> { ... }
    fn to_vec_2d<T: DataType>(&self) -> Result<Vec<Vec<T>>> { ... }
    fn iter<T: DataType>(&self) -> Result<MatIter<'_, T>>
       where Self: Sized { ... }
    fn try_into_typed<T: DataType>(self) -> Result<Mat_<T>>
       where Self: Sized,
             Mat_<T>: TryFrom<Self, Error = Error> { ... }

}

Provided Methods

unsafe fn at_unchecked<T: DataType>(&self, i0: i32) -> Result<&T>

Like Mat::at() but performs no bounds or type checks

Safety

Caller must ensure that index is within Mat bounds

unsafe fn at_2d_unchecked<T: DataType>(&self, row: i32, col: i32) -> Result<&T>

Like Mat::at_2d() but performs no bounds or type checks

Safety

Caller must ensure that indices are within Mat bounds

unsafe fn at_pt_unchecked<T: DataType>(&self, pt: Point) -> Result<&T>

Like Mat::at_pt() but performs no bounds or type checks

Safety

Caller must ensure that point is within Mat bounds

unsafe fn at_3d_unchecked<T: DataType>( &self, i0: i32, i1: i32, i2: i32 ) -> Result<&T>

Like Mat::at_3d() but performs no bounds or type checks

Safety

Caller must ensure that indices are within Mat bounds

unsafe fn at_nd_unchecked<T: DataType>(&self, idx: &[i32]) -> Result<&T>

Like Mat::at_nd() but performs no bounds or type checks

Safety

Caller must ensure that indices are within Mat bounds

fn at_row<T: DataType>(&self, row: i32) -> Result<&[T]>

Return a complete read-only row

unsafe fn at_row_unchecked<T: DataType>(&self, row: i32) -> Result<&[T]>

Like Mat::at_row() but performs no bounds or type checks

Safety

Caller must ensure that index is within Mat bounds

fn is_allocated(&self) -> bool

fn data_bytes(&self) -> Result<&[u8]>

Returns underlying data array as byte slice, Mat must be continuous

fn data_typed<T: DataType>(&self) -> Result<&[T]>

unsafe fn data_typed_unchecked<T: DataType>(&self) -> Result<&[T]>

Safety

Caller must ensure that the T type argument corresponds to the data stored in the Mat and Mat is continuous

fn to_vec_2d<T: DataType>(&self) -> Result<Vec<Vec<T>>>

fn iter<T: DataType>(&self) -> Result<MatIter<'_, T>>
where Self: Sized,

Returns an iterator over Mat elements and their positions

fn try_into_typed<T: DataType>(self) -> Result<Mat_<T>>
where Self: Sized, Mat_<T>: TryFrom<Self, Error = Error>,

Object Safety

This trait is not object safe.