Trait opencv::manual::core::matx::MatxTrait

pub trait MatxTrait: Sized {
    type ElemType;

    // Required methods
    fn rows(&self) -> usize;
    fn cols(&self) -> usize;
    fn val(&self) -> &[Self::ElemType];
    fn val_mut(&mut self) -> &mut [Self::ElemType];
    fn all(alpha: Self::ElemType) -> Self;

    // Provided methods
    fn channels(&self) -> usize { ... }
    fn shortdim(&self) -> usize { ... }
    fn zeros() -> Self
       where Self::ElemType: Zero { ... }
    fn ones() -> Self
       where Self::ElemType: One { ... }
    fn get(&self, idx: (usize, usize)) -> Option<&Self::ElemType> { ... }
    unsafe fn get_unchecked(&self, idx: (usize, usize)) -> &Self::ElemType { ... }
    fn get_mut(&mut self, idx: (usize, usize)) -> Option<&mut Self::ElemType> { ... }
    unsafe fn get_unchecked_mut(
        &mut self,
        idx: (usize, usize)
    ) -> &mut Self::ElemType { ... }
    fn eye() -> Self
       where Self::ElemType: One + Zero { ... }

}

Required Associated Types

type ElemType

Required Methods

fn rows(&self) -> usize

fn cols(&self) -> usize

fn val(&self) -> &[Self::ElemType]

fn val_mut(&mut self) -> &mut [Self::ElemType]

fn all(alpha: Self::ElemType) -> Self

Provided Methods

fn channels(&self) -> usize

fn shortdim(&self) -> usize

fn zeros() -> Self
where Self::ElemType: Zero,

fn ones() -> Self
where Self::ElemType: One,

fn get(&self, idx: (usize, usize)) -> Option<&Self::ElemType>

unsafe fn get_unchecked(&self, idx: (usize, usize)) -> &Self::ElemType

Safety

Caller must ensure that the specified idx is within the Matx bounds

fn get_mut(&mut self, idx: (usize, usize)) -> Option<&mut Self::ElemType>

unsafe fn get_unchecked_mut( &mut self, idx: (usize, usize) ) -> &mut Self::ElemType

Safety

Caller must ensure that the specified idx is within the Matx bounds

fn eye() -> Self
where Self::ElemType: One + Zero,

Object Safety

This trait is not object safe.