Trait opencv::manual::core::vector::vector_extern::VectorElement

pub trait VectorElement: for<'a> OpenCVType<'a>
where
    Vector<Self>: VectorExtern<Self>,
{ }

This trait is implemented by any type that can be stored inside Vector.

It is mostly used internally, feasibility of implementing it for your own types hasn’t been considered. Use is mostly for external type constraints.

Object Safety

This trait is not object safe.