Function opencv::opencv::hub::core::copy_to

pub fn copy_to(
    src: &impl ToInputArray,
    dst: &mut impl ToOutputArray,
    mask: &impl ToInputArray,
) -> Result<()>

This is an overloaded member function, provided for convenience (python) Copies the matrix to another one. When the operation mask is specified, if the Mat::create call shown above reallocates the matrix, the newly allocated matrix is initialized with all zeros before copying the data.

Parameters

src: source matrix.

dst: Destination matrix. If it does not have a proper size or type before the operation, it is reallocated.

mask: Operation mask of the same size as *this. Its non-zero elements indicate which matrix elements need to be copied. The mask has to be of type CV_8U and can have 1 or multiple channels.