opencv::opencv::hub::core
Struct Algorithm

pub struct Algorithm {
    ptr: *mut c_void,
}

This is a base class for all more or less complex algorithms in OpenCV

especially for classes of algorithms, for which there can be multiple implementations. The examples are stereo correspondence (for which there are algorithms like block matching, semi-global block matching, graph-cut etc.), background subtraction (which can be done using mixture-of-gaussians models, codebook-based algorithm etc.), optical flow (block matching, Lucas-Kanade, Horn-Schunck etc.).

Here is example of SimpleBlobDetector use in your application via Algorithm interface: Algorithm

Fields
ptr: *mut c_void
Implementations

impl Algorithm

pub fn default() -> Result<Algorithm>