Struct opencv::opencv::hub::core::FileNode

pub struct FileNode {
    ptr: *mut c_void,
}

File Storage Node class.

The node is used to store each and every element of the file storage opened for reading. When XML/YAML file is read, it is first parsed and stored in the memory as a hierarchical collection of nodes. Each node can be a “leaf” that is contain a single number or a string, or be a collection of other nodes. There can be named collections (mappings) where each element has a name and it is accessed by a name, and ordered collections (sequences) where elements do not have names but rather accessed by index. Type of the file node can be determined using FileNode::type method.

Note that file nodes are only used for navigating file storages opened for reading. When a file storage is opened for writing, no data is stored in memory after it is written.

Fields
ptr: *mut c_void
Implementations

impl FileNode

pub fn default() -> Result<FileNode>

The constructors.

These constructors are used to create a default file node, construct it from obsolete structures or from the another file node.

pub fn new( fs: &impl FileStorageTraitConst, block_idx: size_t, ofs: size_t ) -> Result<FileNode>

The constructors.

These constructors are used to create a default file node, construct it from obsolete structures or from the another file node.

Overloaded parameters

Parameters

fs: Pointer to the file storage structure.

blockIdx: Index of the memory block where the file node is stored

ofs: Offset in bytes from the beginning of the serialized storage

@deprecated

pub fn copy(node: &impl FileNodeTraitConst) -> Result<FileNode>

The constructors.

These constructors are used to create a default file node, construct it from obsolete structures or from the another file node.

Overloaded parameters

Parameters

node: File node to be used as initialization for the created file node.

pub fn is_map(flags: i32) -> Result<bool>

pub fn is_seq(flags: i32) -> Result<bool>

pub fn is_collection(flags: i32) -> Result<bool>

pub fn is_empty_collection(flags: i32) -> Result<bool>

pub fn is_flow(flags: i32) -> Result<bool>