Struct opencv::opencv::hub::core::GpuMatND

pub struct GpuMatND {
    ptr: *mut c_void,
}

Fields
ptr: *mut c_void
Implementations

impl GpuMatND

pub fn default() -> Result<GpuMatND>

default constructor

pub fn new(size: GpuMatND_SizeArray, typ: i32) -> Result<GpuMatND>

default constructor

Overloaded parameters

Parameters

size: Array of integers specifying an n-dimensional array shape.

type: Array type. Use CV_8UC1, …, CV_16FC4 to create 1-4 channel matrices, or CV_8UC(n), …, CV_64FC(n) to create multi-channel (up to CV_CN_MAX channels) matrices.

pub unsafe fn new_1( size: GpuMatND_SizeArray, typ: i32, data: *mut c_void, step: GpuMatND_StepArray ) -> Result<GpuMatND>

default constructor

Overloaded parameters

Parameters

size: Array of integers specifying an n-dimensional array shape.

type: Array type. Use CV_8UC1, …, CV_16FC4 to create 1-4 channel matrices, or CV_8UC(n), …, CV_64FC(n) to create multi-channel (up to CV_CN_MAX channels) matrices.

data: Pointer to the user data. Matrix constructors that take data and step parameters do not allocate matrix data. Instead, they just initialize the matrix header that points to the specified data, which means that no data is copied. This operation is very efficient and can be used to process external data using OpenCV functions. The external data is not automatically deallocated, so you should take care of it.

step: Array of _size.size()-1 steps in case of a multi-dimensional array (the last step is always set to the element size). If not specified, the matrix is assumed to be continuous.

C++ default parameters

step: StepArray()

pub unsafe fn new_def( size: GpuMatND_SizeArray, typ: i32, data: *mut c_void ) -> Result<GpuMatND>

@overload

Parameters

size: Array of integers specifying an n-dimensional array shape.

type: Array type. Use CV_8UC1, …, CV_16FC4 to create 1-4 channel matrices, or CV_8UC(n), …, CV_64FC(n) to create multi-channel (up to CV_CN_MAX channels) matrices.

data: Pointer to the user data. Matrix constructors that take data and step parameters do not allocate matrix data. Instead, they just initialize the matrix header that points to the specified data, which means that no data is copied. This operation is very efficient and can be used to process external data using OpenCV functions. The external data is not automatically deallocated, so you should take care of it.

step: Array of _size.size()-1 steps in case of a multi-dimensional array (the last step is always set to the element size). If not specified, the matrix is assumed to be continuous.

Note

This alternative version of [new] function uses the following default values for its arguments:

step: StepArray()

pub fn copy(unnamed: &impl GpuMatNDTraitConst) -> GpuMatND

pub fn copy_mut(unnamed: GpuMatND) -> GpuMatND