Struct opencv::opencv::hub::core::HostMem

pub struct HostMem {
    ptr: *mut c_void,
}

Class with reference counting wrapping special memory type allocation functions from CUDA.

Its interface is also Mat-like but with additional memory type parameters.

PAGE_LOCKED sets a page locked memory type used commonly for fast and asynchronous uploading/downloading data from/to GPU.

SHARED specifies a zero copy memory allocation that enables mapping the host memory to GPU address space, if supported.

WRITE_COMBINED sets the write combined buffer that is not cached by CPU. Such buffers are used to supply GPU with data when GPU only reads it. The advantage is a better CPU cache utilization.

Note: Allocation size of such memory types is usually limited. For more details, see CUDA 2.2 Pinned Memory APIs document or CUDA C Programming Guide.

Fields
ptr: *mut c_void
Implementations

impl HostMem

pub fn new(alloc_type: HostMem_AllocType) -> Result<HostMem>

C++ default parameters

alloc_type: HostMem::AllocType::PAGE_LOCKED

pub fn new_def() -> Result<HostMem>

Note

This alternative version of [new] function uses the following default values for its arguments:

alloc_type: HostMem::AllocType::PAGE_LOCKED

pub fn copy(m: &impl HostMemTraitConst) -> Result<HostMem>

pub fn new_1( rows: i32, cols: i32, typ: i32, alloc_type: HostMem_AllocType ) -> Result<HostMem>

C++ default parameters

alloc_type: HostMem::AllocType::PAGE_LOCKED

pub fn new_def_1(rows: i32, cols: i32, typ: i32) -> Result<HostMem>

Note

This alternative version of [new] function uses the following default values for its arguments:

alloc_type: HostMem::AllocType::PAGE_LOCKED

pub fn new_2( size: Size, typ: i32, alloc_type: HostMem_AllocType ) -> Result<HostMem>

C++ default parameters

alloc_type: HostMem::AllocType::PAGE_LOCKED

pub fn new_def_2(size: Size, typ: i32) -> Result<HostMem>

Note

This alternative version of [new] function uses the following default values for its arguments:

alloc_type: HostMem::AllocType::PAGE_LOCKED

pub fn new_3( arr: &impl ToInputArray, alloc_type: HostMem_AllocType ) -> Result<HostMem>

creates from host memory with coping data

C++ default parameters

alloc_type: HostMem::AllocType::PAGE_LOCKED

pub fn new_def_3(arr: &impl ToInputArray) -> Result<HostMem>

creates from host memory with coping data

Note

This alternative version of [new] function uses the following default values for its arguments:

alloc_type: HostMem::AllocType::PAGE_LOCKED