Struct opencv::opencv::hub::core::Stream

pub struct Stream {
    ptr: *mut c_void,
}

This class encapsulates a queue of asynchronous calls.

Note: Currently, you may face problems if an operation is enqueued twice with different data. Some functions use the constant GPU memory, and next call may update the memory before the previous one has been finished. But calling different operations asynchronously is safe because each operation has its own constant buffer. Memory copy/upload/download/set operations to the buffers you hold are also safe.

Note: The Stream class is not thread-safe. Please use different Stream objects for different CPU threads.

void thread1()
{
   cv::cuda::Stream stream1;
   cv::cuda::func1(..., stream1);
}
 
void thread2()
{
   cv::cuda::Stream stream2;
   cv::cuda::func2(..., stream2);
}

Note: By default all CUDA routines are launched in Stream::Null() object, if the stream is not specified by user. In multi-threading environment the stream objects must be passed explicitly (see previous note).

Fields
ptr: *mut c_void
Implementations

impl Stream

pub fn default() -> Result<Stream>

creates a new asynchronous stream

pub fn new(allocator: &Ptr<GpuMat_Allocator>) -> Result<Stream>

creates a new asynchronous stream with custom allocator

pub fn new_1(cuda_flags: size_t) -> Result<Stream>

creates a new Stream using the cudaFlags argument to determine the behaviors of the stream

Note: The cudaFlags parameter is passed to the underlying api cudaStreamCreateWithFlags() and supports the same parameter values.

   // creates an OpenCV cuda::Stream that manages an asynchronous, non-blocking,
   // non-default CUDA stream
   cv::cuda::Stream cvStream(cudaStreamNonBlocking);

pub fn null() -> Result<Stream>

return Stream object for default CUDA stream