Struct opencv::opencv::hub::core::TickMeter

pub struct TickMeter {
    ptr: *mut c_void,
}

a Class to measure passing time.

The class computes passing time by counting the number of ticks per second. That is, the following code computes the execution time in seconds: TickMeter_total

It is also possible to compute the average time over multiple runs: TickMeter_average

See also

getTickCount, getTickFrequency

Fields
ptr: *mut c_void
Implementations

impl TickMeter

pub fn default() -> Result<TickMeter>

the default constructor