opencv::opencv::hub::core
Trait AsyncArrayTraitConst

pub trait AsyncArrayTraitConst {
    // Required method
    fn as_raw_AsyncArray(&self) -> *const c_void;

    // Provided methods
    fn get(&self, dst: &mut impl ToOutputArray) -> Result<()> { ... }
    fn get_with_timeout(
        &self,
        dst: &mut impl ToOutputArray,
        timeout_ns: i64,
    ) -> Result<bool> { ... }
    fn get_with_timeout_f64(
        &self,
        dst: &mut impl ToOutputArray,
        timeout_ns: f64,
    ) -> Result<bool> { ... }
    fn wait_for(&self, timeout_ns: i64) -> Result<bool> { ... }
    fn wait_for_f64(&self, timeout_ns: f64) -> Result<bool> { ... }
    fn valid(&self) -> bool { ... }
}

Constant methods for core::AsyncArray

Required Methods

fn as_raw_AsyncArray(&self) -> *const c_void

Provided Methods

fn get(&self, dst: &mut impl ToOutputArray) -> Result<()>

Fetch the result.

Parameters

dst:[out] destination array

Waits for result until container has valid result. Throws exception if exception was stored as a result.

Throws exception on invalid container state.

Note: Result or stored exception can be fetched only once.

fn get_with_timeout( &self, dst: &mut impl ToOutputArray, timeout_ns: i64, ) -> Result<bool>

Retrieving the result with timeout

Parameters

dst:[out] destination array

timeoutNs: timeout in nanoseconds, -1 for infinite wait

Returns

true if result is ready, false if the timeout has expired

Note: Result or stored exception can be fetched only once.

fn get_with_timeout_f64( &self, dst: &mut impl ToOutputArray, timeout_ns: f64, ) -> Result<bool>

fn wait_for(&self, timeout_ns: i64) -> Result<bool>

fn wait_for_f64(&self, timeout_ns: f64) -> Result<bool>

fn valid(&self) -> bool

Object Safety

This trait is not object safe.