Trait opencv::opencv::hub::core::KeyPointTrait

pub trait KeyPointTrait: KeyPointTraitConst {
    // Required method
    fn as_raw_mut_KeyPoint(&mut self) -> *mut c_void;

    // Provided methods
    fn set_pt(&mut self, val: Point2f) { ... }
    fn set_size(&mut self, val: f32) { ... }
    fn set_angle(&mut self, val: f32) { ... }
    fn set_response(&mut self, val: f32) { ... }
    fn set_octave(&mut self, val: i32) { ... }
    fn set_class_id(&mut self, val: i32) { ... }
}

Mutable methods for core::KeyPoint

Required Methods

fn as_raw_mut_KeyPoint(&mut self) -> *mut c_void

Provided Methods

fn set_pt(&mut self, val: Point2f)

coordinates of the keypoints

fn set_size(&mut self, val: f32)

diameter of the meaningful keypoint neighborhood

fn set_angle(&mut self, val: f32)

computed orientation of the keypoint (-1 if not applicable); it’s in [0,360) degrees and measured relative to image coordinate system, ie in clockwise.

fn set_response(&mut self, val: f32)

the response by which the most strong keypoints have been selected. Can be used for the further sorting or subsampling

fn set_octave(&mut self, val: i32)

octave (pyramid layer) from which the keypoint has been extracted

fn set_class_id(&mut self, val: i32)

object class (if the keypoints need to be clustered by an object they belong to)