Trait opencv::opencv::hub::core::MatConstIteratorTrait

pub trait MatConstIteratorTrait: MatConstIteratorTraitConst {
    // Required method
    fn as_raw_mut_MatConstIterator(&mut self) -> *mut c_void;

    // Provided methods
    fn set_elem_size(&mut self, val: size_t) { ... }
    fn set(&mut self, it: &impl MatConstIteratorTraitConst) -> Result<()> { ... }
    fn decr(&mut self) -> Result<MatConstIterator> { ... }
    fn incr(&mut self) -> Result<MatConstIterator> { ... }
    fn seek(&mut self, ofs: ptrdiff_t, relative: bool) -> Result<()> { ... }
    fn seek_def(&mut self, ofs: ptrdiff_t) -> Result<()> { ... }
    fn seek_idx(&mut self, _idx: &i32, relative: bool) -> Result<()> { ... }
    fn seek_idx_def(&mut self, _idx: &i32) -> Result<()> { ... }
}

Mutable methods for core::MatConstIterator

Required Methods

fn as_raw_mut_MatConstIterator(&mut self) -> *mut c_void

Provided Methods

fn set_elem_size(&mut self, val: size_t)

fn set(&mut self, it: &impl MatConstIteratorTraitConst) -> Result<()>

copy operator

fn decr(&mut self) -> Result<MatConstIterator>

decrements the iterator

fn incr(&mut self) -> Result<MatConstIterator>

increments the iterator

fn seek(&mut self, ofs: ptrdiff_t, relative: bool) -> Result<()>

C++ default parameters

relative: false

fn seek_def(&mut self, ofs: ptrdiff_t) -> Result<()>

Note

This alternative version of MatConstIteratorTrait::seek function uses the following default values for its arguments:

relative: false

fn seek_idx(&mut self, _idx: &i32, relative: bool) -> Result<()>

C++ default parameters

relative: false

fn seek_idx_def(&mut self, _idx: &i32) -> Result<()>

Note

This alternative version of MatConstIteratorTrait::seek_idx function uses the following default values for its arguments:

relative: false

Object Safety

This trait is not object safe.