Trait opencv::opencv::hub::core::MatOpTraitConst

pub trait MatOpTraitConst {

    // Required method
    fn as_raw_MatOp(&self) -> *const c_void;

    // Provided methods
    fn element_wise(&self, expr: &impl MatExprTraitConst) -> Result<bool> { ... }
    fn assign(
        &self,
        expr: &impl MatExprTraitConst,
        m: &mut impl MatTrait,
        typ: i32
    ) -> Result<()> { ... }
    fn assign_def(
        &self,
        expr: &impl MatExprTraitConst,
        m: &mut impl MatTrait
    ) -> Result<()> { ... }
    fn roi(
        &self,
        expr: &impl MatExprTraitConst,
        row_range: &impl RangeTraitConst,
        col_range: &impl RangeTraitConst,
        res: &mut impl MatExprTrait
    ) -> Result<()> { ... }
    fn diag(
        &self,
        expr: &impl MatExprTraitConst,
        d: i32,
        res: &mut impl MatExprTrait
    ) -> Result<()> { ... }
    fn aug_assign_add(
        &self,
        expr: &impl MatExprTraitConst,
        m: &mut impl MatTrait
    ) -> Result<()> { ... }
    fn aug_assign_subtract(
        &self,
        expr: &impl MatExprTraitConst,
        m: &mut impl MatTrait
    ) -> Result<()> { ... }
    fn aug_assign_multiply(
        &self,
        expr: &impl MatExprTraitConst,
        m: &mut impl MatTrait
    ) -> Result<()> { ... }
    fn aug_assign_divide(
        &self,
        expr: &impl MatExprTraitConst,
        m: &mut impl MatTrait
    ) -> Result<()> { ... }
    fn aug_assign_and(
        &self,
        expr: &impl MatExprTraitConst,
        m: &mut impl MatTrait
    ) -> Result<()> { ... }
    fn aug_assign_or(
        &self,
        expr: &impl MatExprTraitConst,
        m: &mut impl MatTrait
    ) -> Result<()> { ... }
    fn aug_assign_xor(
        &self,
        expr: &impl MatExprTraitConst,
        m: &mut impl MatTrait
    ) -> Result<()> { ... }
    fn add(
        &self,
        expr1: &impl MatExprTraitConst,
        expr2: &impl MatExprTraitConst,
        res: &mut impl MatExprTrait
    ) -> Result<()> { ... }
    fn add_scalar(
        &self,
        expr1: &impl MatExprTraitConst,
        s: Scalar,
        res: &mut impl MatExprTrait
    ) -> Result<()> { ... }
    fn subtract(
        &self,
        expr1: &impl MatExprTraitConst,
        expr2: &impl MatExprTraitConst,
        res: &mut impl MatExprTrait
    ) -> Result<()> { ... }
    fn subtract_scalar(
        &self,
        s: Scalar,
        expr: &impl MatExprTraitConst,
        res: &mut impl MatExprTrait
    ) -> Result<()> { ... }
    fn multiply(
        &self,
        expr1: &impl MatExprTraitConst,
        expr2: &impl MatExprTraitConst,
        res: &mut impl MatExprTrait,
        scale: f64
    ) -> Result<()> { ... }
    fn multiply_def(
        &self,
        expr1: &impl MatExprTraitConst,
        expr2: &impl MatExprTraitConst,
        res: &mut impl MatExprTrait
    ) -> Result<()> { ... }
    fn multiply_f64(
        &self,
        expr1: &impl MatExprTraitConst,
        s: f64,
        res: &mut impl MatExprTrait
    ) -> Result<()> { ... }
    fn divide(
        &self,
        expr1: &impl MatExprTraitConst,
        expr2: &impl MatExprTraitConst,
        res: &mut impl MatExprTrait,
        scale: f64
    ) -> Result<()> { ... }
    fn divide_def(
        &self,
        expr1: &impl MatExprTraitConst,
        expr2: &impl MatExprTraitConst,
        res: &mut impl MatExprTrait
    ) -> Result<()> { ... }
    fn divide_f64(
        &self,
        s: f64,
        expr: &impl MatExprTraitConst,
        res: &mut impl MatExprTrait
    ) -> Result<()> { ... }
    fn abs(
        &self,
        expr: &impl MatExprTraitConst,
        res: &mut impl MatExprTrait
    ) -> Result<()> { ... }
    fn transpose(
        &self,
        expr: &impl MatExprTraitConst,
        res: &mut impl MatExprTrait
    ) -> Result<()> { ... }
    fn matmul(
        &self,
        expr1: &impl MatExprTraitConst,
        expr2: &impl MatExprTraitConst,
        res: &mut impl MatExprTrait
    ) -> Result<()> { ... }
    fn invert(
        &self,
        expr: &impl MatExprTraitConst,
        method: i32,
        res: &mut impl MatExprTrait
    ) -> Result<()> { ... }
    fn size(&self, expr: &impl MatExprTraitConst) -> Result<Size> { ... }
    fn typ(&self, expr: &impl MatExprTraitConst) -> Result<i32> { ... }

}

Constant methods for core::MatOp

Required Methods

fn as_raw_MatOp(&self) -> *const c_void

Provided Methods

fn element_wise(&self, expr: &impl MatExprTraitConst) -> Result<bool>

fn assign( &self, expr: &impl MatExprTraitConst, m: &mut impl MatTrait, typ: i32 ) -> Result<()>

C++ default parameters

typ: -1

fn assign_def( &self, expr: &impl MatExprTraitConst, m: &mut impl MatTrait ) -> Result<()>

Note

This alternative version of MatOpTraitConst::assign function uses the following default values for its arguments:

typ: -1

fn roi( &self, expr: &impl MatExprTraitConst, row_range: &impl RangeTraitConst, col_range: &impl RangeTraitConst, res: &mut impl MatExprTrait ) -> Result<()>

fn diag( &self, expr: &impl MatExprTraitConst, d: i32, res: &mut impl MatExprTrait ) -> Result<()>

fn aug_assign_add( &self, expr: &impl MatExprTraitConst, m: &mut impl MatTrait ) -> Result<()>

fn aug_assign_subtract( &self, expr: &impl MatExprTraitConst, m: &mut impl MatTrait ) -> Result<()>

fn aug_assign_multiply( &self, expr: &impl MatExprTraitConst, m: &mut impl MatTrait ) -> Result<()>

fn aug_assign_divide( &self, expr: &impl MatExprTraitConst, m: &mut impl MatTrait ) -> Result<()>

fn aug_assign_and( &self, expr: &impl MatExprTraitConst, m: &mut impl MatTrait ) -> Result<()>

fn aug_assign_or( &self, expr: &impl MatExprTraitConst, m: &mut impl MatTrait ) -> Result<()>

fn aug_assign_xor( &self, expr: &impl MatExprTraitConst, m: &mut impl MatTrait ) -> Result<()>

fn add( &self, expr1: &impl MatExprTraitConst, expr2: &impl MatExprTraitConst, res: &mut impl MatExprTrait ) -> Result<()>

fn add_scalar( &self, expr1: &impl MatExprTraitConst, s: Scalar, res: &mut impl MatExprTrait ) -> Result<()>

fn subtract( &self, expr1: &impl MatExprTraitConst, expr2: &impl MatExprTraitConst, res: &mut impl MatExprTrait ) -> Result<()>

fn subtract_scalar( &self, s: Scalar, expr: &impl MatExprTraitConst, res: &mut impl MatExprTrait ) -> Result<()>

fn multiply( &self, expr1: &impl MatExprTraitConst, expr2: &impl MatExprTraitConst, res: &mut impl MatExprTrait, scale: f64 ) -> Result<()>

C++ default parameters

scale: 1

fn multiply_def( &self, expr1: &impl MatExprTraitConst, expr2: &impl MatExprTraitConst, res: &mut impl MatExprTrait ) -> Result<()>

Note

This alternative version of MatOpTraitConst::multiply function uses the following default values for its arguments:

scale: 1

fn multiply_f64( &self, expr1: &impl MatExprTraitConst, s: f64, res: &mut impl MatExprTrait ) -> Result<()>

fn divide( &self, expr1: &impl MatExprTraitConst, expr2: &impl MatExprTraitConst, res: &mut impl MatExprTrait, scale: f64 ) -> Result<()>

C++ default parameters

scale: 1

fn divide_def( &self, expr1: &impl MatExprTraitConst, expr2: &impl MatExprTraitConst, res: &mut impl MatExprTrait ) -> Result<()>

Note

This alternative version of MatOpTraitConst::divide function uses the following default values for its arguments:

scale: 1

fn divide_f64( &self, s: f64, expr: &impl MatExprTraitConst, res: &mut impl MatExprTrait ) -> Result<()>

fn abs( &self, expr: &impl MatExprTraitConst, res: &mut impl MatExprTrait ) -> Result<()>

fn transpose( &self, expr: &impl MatExprTraitConst, res: &mut impl MatExprTrait ) -> Result<()>

fn matmul( &self, expr1: &impl MatExprTraitConst, expr2: &impl MatExprTraitConst, res: &mut impl MatExprTrait ) -> Result<()>

fn invert( &self, expr: &impl MatExprTraitConst, method: i32, res: &mut impl MatExprTrait ) -> Result<()>

fn size(&self, expr: &impl MatExprTraitConst) -> Result<Size>

fn typ(&self, expr: &impl MatExprTraitConst) -> Result<i32>

Object Safety

This trait is not object safe.