Function opencv::opencv::hub::dnn::read_net_from_darknet_str_def

pub fn read_net_from_darknet_str_def(
    buffer_cfg: &str,
    len_cfg: size_t
) -> Result<Net>

Reads a network model stored in Darknet model files.

Parameters

bufferCfg: A buffer contains a content of .cfg file with text description of the network architecture.

lenCfg: Number of bytes to read from bufferCfg

bufferModel: A buffer contains a content of .weights file with learned network.

lenModel: Number of bytes to read from bufferModel

Returns

Net object.

Note

This alternative version of read_net_from_darknet_str function uses the following default values for its arguments:

buffer_model: NULL

len_model: 0