Struct opencv::opencv::hub::dnn::KeypointsModel

pub struct KeypointsModel {
    ptr: *mut c_void,
}

This class represents high-level API for keypoints models

KeypointsModel allows to set params for preprocessing input image. KeypointsModel creates net from file with trained weights and config, sets preprocessing input, runs forward pass and returns the x and y coordinates of each detected keypoint

Fields
ptr: *mut c_void
Implementations

impl KeypointsModel

pub fn new(model: &str, config: &str) -> Result<KeypointsModel>

Create keypoints model from network represented in one of the supported formats. An order of @p model and @p config arguments does not matter.

Parameters

model: Binary file contains trained weights.

config: Text file contains network configuration.

C++ default parameters

config: “”

pub fn new_def(model: &str) -> Result<KeypointsModel>

Create keypoints model from network represented in one of the supported formats. An order of @p model and @p config arguments does not matter.

Parameters

model: Binary file contains trained weights.

config: Text file contains network configuration.

Note

This alternative version of [new] function uses the following default values for its arguments:

config: “”

pub fn new_1(network: &Net) -> Result<KeypointsModel>

Create model from deep learning network.

Parameters

network: Net object.