Struct opencv::opencv::hub::dnn::LayerParams

pub struct LayerParams {
    ptr: *mut c_void,
}

This class provides all data needed to initialize layer.

It includes dictionary with scalar params (which can be read by using Dict interface), blob params [blobs] and optional meta information: [name] and [type] of layer instance.

Fields
ptr: *mut c_void
Implementations

impl LayerParams

fn default() -> Self

Creates a default instance of the class by calling the default constructor