opencv::opencv::hub::imgcodecs
Function imreadmulti_def

pub fn imreadmulti_def(filename: &str, mats: &mut Vector<Mat>) -> Result<bool>

Loads a multi-page image from a file.

The function imreadmulti loads a multi-page image from the specified file into a vector of Mat objects.

Parameters

filename: Name of file to be loaded.

mats: A vector of Mat objects holding each page.

flags: Flag that can take values of cv::ImreadModes, default with cv::IMREAD_ANYCOLOR.

See also

cv::imread

Note

This alternative version of imreadmulti function uses the following default values for its arguments:

flags: IMREAD_ANYCOLOR