Function opencv::opencv::hub::imgproc::connected_components_with_stats_def

pub fn connected_components_with_stats_def(
    image: &impl ToInputArray,
    labels: &mut impl ToOutputArray,
    stats: &mut impl ToOutputArray,
    centroids: &mut impl ToOutputArray
) -> Result<i32>

@overload

Parameters

image: the 8-bit single-channel image to be labeled

labels: destination labeled image

stats: statistics output for each label, including the background label. Statistics are accessed via stats(label, COLUMN) where COLUMN is one of #ConnectedComponentsTypes, selecting the statistic. The data type is CV_32S.

centroids: centroid output for each label, including the background label. Centroids are accessed via centroids(label, 0) for x and centroids(label, 1) for y. The data type CV_64F.

connectivity: 8 or 4 for 8-way or 4-way connectivity respectively

ltype: output image label type. Currently CV_32S and CV_16U are supported.

Note

This alternative version of connected_components_with_stats function uses the following default values for its arguments:

connectivity: 8

ltype: CV_32S