Function opencv::opencv::hub::imgproc::ellipse_2_poly

pub fn ellipse_2_poly(
    center: Point,
    axes: Size,
    angle: i32,
    arc_start: i32,
    arc_end: i32,
    delta: i32,
    pts: &mut Vector<Point>
) -> Result<()>

Approximates an elliptic arc with a polyline.

The function ellipse2Poly computes the vertices of a polyline that approximates the specified elliptic arc. It is used by #ellipse. If arcStart is greater than arcEnd, they are swapped.

Parameters

center: Center of the arc.

axes: Half of the size of the ellipse main axes. See ellipse for details.

angle: Rotation angle of the ellipse in degrees. See ellipse for details.

arcStart: Starting angle of the elliptic arc in degrees.

arcEnd: Ending angle of the elliptic arc in degrees.

delta: Angle between the subsequent polyline vertices. It defines the approximation accuracy.

pts: Output vector of polyline vertices.