Function opencv::opencv::hub::imgproc::get_font_scale_from_height_def

pub fn get_font_scale_from_height_def(
    font_face: i32,
    pixel_height: i32
) -> Result<f64>

Calculates the font-specific size to use to achieve a given height in pixels.

Parameters

fontFace: Font to use, see cv::HersheyFonts.

pixelHeight: Pixel height to compute the fontScale for

thickness: Thickness of lines used to render the text.See putText for details.

Returns

The fontSize to use for cv::putText

See also

cv::putText

Note

This alternative version of get_font_scale_from_height function uses the following default values for its arguments:

thickness: 1