Function opencv::opencv::hub::imgproc::good_features_to_track_with_quality_def

pub fn good_features_to_track_with_quality_def(
    image: &impl ToInputArray,
    corners: &mut impl ToOutputArray,
    max_corners: i32,
    quality_level: f64,
    min_distance: f64,
    mask: &impl ToInputArray,
    corners_quality: &mut impl ToOutputArray
) -> Result<()>

Same as above, but returns also quality measure of the detected corners.

Parameters

image: Input 8-bit or floating-point 32-bit, single-channel image.

corners: Output vector of detected corners.

maxCorners: Maximum number of corners to return. If there are more corners than are found, the strongest of them is returned. maxCorners <= 0 implies that no limit on the maximum is set and all detected corners are returned.

qualityLevel: Parameter characterizing the minimal accepted quality of image corners. The parameter value is multiplied by the best corner quality measure, which is the minimal eigenvalue (see corner_min_eigen_val ) or the Harris function response (see corner_harris ). The corners with the quality measure less than the product are rejected. For example, if the best corner has the quality measure = 1500, and the qualityLevel=0.01 , then all the corners with the quality measure less than 15 are rejected.

minDistance: Minimum possible Euclidean distance between the returned corners.

mask: Region of interest. If the image is not empty (it needs to have the type CV_8UC1 and the same size as image ), it specifies the region in which the corners are detected.

cornersQuality: Output vector of quality measure of the detected corners.

blockSize: Size of an average block for computing a derivative covariation matrix over each pixel neighborhood. See cornerEigenValsAndVecs .

gradientSize: Aperture parameter for the Sobel operator used for derivatives computation. See cornerEigenValsAndVecs .

useHarrisDetector: Parameter indicating whether to use a Harris detector (see #cornerHarris) or #cornerMinEigenVal.

k: Free parameter of the Harris detector.

Note

This alternative version of good_features_to_track_with_quality function uses the following default values for its arguments:

block_size: 3

gradient_size: 3

use_harris_detector: false

k: 0.04