Function opencv::opencv::hub::imgproc::put_text

pub fn put_text(
    img: &mut impl ToInputOutputArray,
    text: &str,
    org: Point,
    font_face: i32,
    font_scale: f64,
    color: Scalar,
    thickness: i32,
    line_type: i32,
    bottom_left_origin: bool
) -> Result<()>

Draws a text string.

The function cv::putText renders the specified text string in the image. Symbols that cannot be rendered using the specified font are replaced by question marks. See get_text_size for a text rendering code example.

Parameters

img: Image.

text: Text string to be drawn.

org: Bottom-left corner of the text string in the image.

fontFace: Font type, see #HersheyFonts.

fontScale: Font scale factor that is multiplied by the font-specific base size.

color: Text color.

thickness: Thickness of the lines used to draw a text.

lineType: Line type. See [line_types]

bottomLeftOrigin: When true, the image data origin is at the bottom-left corner. Otherwise, it is at the top-left corner.

C++ default parameters

thickness: 1

line_type: LINE_8

bottom_left_origin: false