Struct opencv::opencv::hub::imgproc::Subdiv2D

pub struct Subdiv2D {
    ptr: *mut c_void,
}

Fields
ptr: *mut c_void
Implementations

impl Subdiv2D

pub fn default() -> Result<Subdiv2D>

creates an empty Subdiv2D object. To create a new empty Delaunay subdivision you need to use the [init_delaunay] function.

pub fn new(rect: Rect) -> Result<Subdiv2D>

creates an empty Subdiv2D object. To create a new empty Delaunay subdivision you need to use the [init_delaunay] function.

Overloaded parameters

Parameters

rect: Rectangle that includes all of the 2D points that are to be added to the subdivision.

The function creates an empty Delaunay subdivision where 2D points can be added using the function insert() . All of the points to be added must be within the specified rectangle, otherwise a runtime error is raised.