Trait opencv::opencv::hub::videoio::VideoWriterTraitConst

pub trait VideoWriterTraitConst {
    // Required method
    fn as_raw_VideoWriter(&self) -> *const c_void;

    // Provided methods
    fn is_opened(&self) -> Result<bool> { ... }
    fn get(&self, prop_id: i32) -> Result<f64> { ... }
    fn get_backend_name(&self) -> Result<String> { ... }
}

Constant methods for crate::videoio::VideoWriter

Required Methods

fn as_raw_VideoWriter(&self) -> *const c_void

Provided Methods

fn is_opened(&self) -> Result<bool>

Returns true if video writer has been successfully initialized.

fn get(&self, prop_id: i32) -> Result<f64>

Returns the specified VideoWriter property

Parameters

propId: Property identifier from cv::VideoWriterProperties (eg. cv::VIDEOWRITER_PROP_QUALITY) or one of [videoio_flags_others]

Returns

Value for the specified property. Value 0 is returned when querying a property that is not supported by the backend used by the VideoWriter instance.

fn get_backend_name(&self) -> Result<String>

Returns used backend API name

Note: Stream should be opened.