Trait opencv::prelude::ArraysTrait

pub trait ArraysTrait: ArraysTraitConst {
    // Required method
    fn as_raw_mut_Arrays(&mut self) -> *mut c_void;

    // Provided methods
    fn set_vertex_array(&mut self, vertex: &impl ToInputArray) -> Result<()> { ... }
    fn reset_vertex_array(&mut self) -> Result<()> { ... }
    fn set_color_array(&mut self, color: &impl ToInputArray) -> Result<()> { ... }
    fn reset_color_array(&mut self) -> Result<()> { ... }
    fn set_normal_array(&mut self, normal: &impl ToInputArray) -> Result<()> { ... }
    fn reset_normal_array(&mut self) -> Result<()> { ... }
    fn set_tex_coord_array(
        &mut self,
        tex_coord: &impl ToInputArray
    ) -> Result<()> { ... }
    fn reset_tex_coord_array(&mut self) -> Result<()> { ... }
    fn release(&mut self) -> Result<()> { ... }
    fn set_auto_release(&mut self, flag: bool) -> Result<()> { ... }
}

Mutable methods for core::Arrays

Required Methods

fn as_raw_mut_Arrays(&mut self) -> *mut c_void

Provided Methods

fn set_vertex_array(&mut self, vertex: &impl ToInputArray) -> Result<()>

Sets an array of vertex coordinates.

Parameters

vertex: array with vertex coordinates, can be both host and device memory.

fn reset_vertex_array(&mut self) -> Result<()>

Resets vertex coordinates.

fn set_color_array(&mut self, color: &impl ToInputArray) -> Result<()>

Sets an array of vertex colors.

Parameters

color: array with vertex colors, can be both host and device memory.

fn reset_color_array(&mut self) -> Result<()>

Resets vertex colors.

fn set_normal_array(&mut self, normal: &impl ToInputArray) -> Result<()>

Sets an array of vertex normals.

Parameters

normal: array with vertex normals, can be both host and device memory.

fn reset_normal_array(&mut self) -> Result<()>

Resets vertex normals.

fn set_tex_coord_array(&mut self, tex_coord: &impl ToInputArray) -> Result<()>

Sets an array of vertex texture coordinates.

Parameters

texCoord: array with vertex texture coordinates, can be both host and device memory.

fn reset_tex_coord_array(&mut self) -> Result<()>

Resets vertex texture coordinates.

fn release(&mut self) -> Result<()>

Releases all inner buffers.

fn set_auto_release(&mut self, flag: bool) -> Result<()>

Sets auto release mode all inner buffers.

Parameters

flag: Auto release mode.

Object Safety

This trait is not object safe.