Trait opencv::prelude::GeneralizedHoughTrait

pub trait GeneralizedHoughTrait: AlgorithmTrait + GeneralizedHoughTraitConst {

    // Required method
    fn as_raw_mut_GeneralizedHough(&mut self) -> *mut c_void;

    // Provided methods
    fn set_template(
        &mut self,
        templ: &impl ToInputArray,
        templ_center: Point
    ) -> Result<()> { ... }
    fn set_template_def(&mut self, templ: &impl ToInputArray) -> Result<()> { ... }
    fn set_template_1(
        &mut self,
        edges: &impl ToInputArray,
        dx: &impl ToInputArray,
        dy: &impl ToInputArray,
        templ_center: Point
    ) -> Result<()> { ... }
    fn set_template_def_1(
        &mut self,
        edges: &impl ToInputArray,
        dx: &impl ToInputArray,
        dy: &impl ToInputArray
    ) -> Result<()> { ... }
    fn detect(
        &mut self,
        image: &impl ToInputArray,
        positions: &mut impl ToOutputArray,
        votes: &mut impl ToOutputArray
    ) -> Result<()> { ... }
    fn detect_def(
        &mut self,
        image: &impl ToInputArray,
        positions: &mut impl ToOutputArray
    ) -> Result<()> { ... }
    fn detect_with_edges(
        &mut self,
        edges: &impl ToInputArray,
        dx: &impl ToInputArray,
        dy: &impl ToInputArray,
        positions: &mut impl ToOutputArray,
        votes: &mut impl ToOutputArray
    ) -> Result<()> { ... }
    fn detect_with_edges_def(
        &mut self,
        edges: &impl ToInputArray,
        dx: &impl ToInputArray,
        dy: &impl ToInputArray,
        positions: &mut impl ToOutputArray
    ) -> Result<()> { ... }
    fn set_canny_low_thresh(&mut self, canny_low_thresh: i32) -> Result<()> { ... }
    fn set_canny_high_thresh(&mut self, canny_high_thresh: i32) -> Result<()> { ... }
    fn set_min_dist(&mut self, min_dist: f64) -> Result<()> { ... }
    fn set_dp(&mut self, dp: f64) -> Result<()> { ... }
    fn set_max_buffer_size(&mut self, max_buffer_size: i32) -> Result<()> { ... }

}

Mutable methods for crate::imgproc::GeneralizedHough

Required Methods

fn as_raw_mut_GeneralizedHough(&mut self) -> *mut c_void

Provided Methods

fn set_template( &mut self, templ: &impl ToInputArray, templ_center: Point ) -> Result<()>

set template to search

C++ default parameters

templ_center: Point(-1,-1)

fn set_template_def(&mut self, templ: &impl ToInputArray) -> Result<()>

set template to search

Note

This alternative version of GeneralizedHoughTrait::set_template function uses the following default values for its arguments:

templ_center: Point(-1,-1)

fn set_template_1( &mut self, edges: &impl ToInputArray, dx: &impl ToInputArray, dy: &impl ToInputArray, templ_center: Point ) -> Result<()>

C++ default parameters

templ_center: Point(-1,-1)

fn set_template_def_1( &mut self, edges: &impl ToInputArray, dx: &impl ToInputArray, dy: &impl ToInputArray ) -> Result<()>

Note

This alternative version of GeneralizedHoughTrait::set_template function uses the following default values for its arguments:

templ_center: Point(-1,-1)

fn detect( &mut self, image: &impl ToInputArray, positions: &mut impl ToOutputArray, votes: &mut impl ToOutputArray ) -> Result<()>

find template on image

C++ default parameters

votes: noArray()

fn detect_def( &mut self, image: &impl ToInputArray, positions: &mut impl ToOutputArray ) -> Result<()>

find template on image

Note

This alternative version of GeneralizedHoughTrait::detect function uses the following default values for its arguments:

votes: noArray()

fn detect_with_edges( &mut self, edges: &impl ToInputArray, dx: &impl ToInputArray, dy: &impl ToInputArray, positions: &mut impl ToOutputArray, votes: &mut impl ToOutputArray ) -> Result<()>

C++ default parameters

votes: noArray()

fn detect_with_edges_def( &mut self, edges: &impl ToInputArray, dx: &impl ToInputArray, dy: &impl ToInputArray, positions: &mut impl ToOutputArray ) -> Result<()>

Note

This alternative version of GeneralizedHoughTrait::detect_with_edges function uses the following default values for its arguments:

votes: noArray()

fn set_canny_low_thresh(&mut self, canny_low_thresh: i32) -> Result<()>

Canny low threshold.

fn set_canny_high_thresh(&mut self, canny_high_thresh: i32) -> Result<()>

Canny high threshold.

fn set_min_dist(&mut self, min_dist: f64) -> Result<()>

Minimum distance between the centers of the detected objects.

fn set_dp(&mut self, dp: f64) -> Result<()>

Inverse ratio of the accumulator resolution to the image resolution.

fn set_max_buffer_size(&mut self, max_buffer_size: i32) -> Result<()>

Maximal size of inner buffers.

Object Safety

This trait is not object safe.