Trait opencv::prelude::IntelligentScissorsMBTraitConst

pub trait IntelligentScissorsMBTraitConst {
    // Required method
    fn as_raw_IntelligentScissorsMB(&self) -> *const c_void;

    // Provided methods
    fn get_contour(
        &self,
        target_pt: Point,
        contour: &mut impl ToOutputArray,
        backward: bool
    ) -> Result<()> { ... }
    fn get_contour_def(
        &self,
        target_pt: Point,
        contour: &mut impl ToOutputArray
    ) -> Result<()> { ... }
}

Constant methods for crate::imgproc::IntelligentScissorsMB

Required Methods

fn as_raw_IntelligentScissorsMB(&self) -> *const c_void

Provided Methods

fn get_contour( &self, target_pt: Point, contour: &mut impl ToOutputArray, backward: bool ) -> Result<()>

Extracts optimal contour for the given target point on the image

Note: buildMap() must be called before this call

Parameters

targetPt: The target point

contour:[out] The list of pixels which contains optimal path between the source and the target points of the image. Type is CV_32SC2 (compatible with std::vector<Point>)

backward: Flag to indicate reverse order of retrived pixels (use “true” value to fetch points from the target to the source point)

C++ default parameters

backward: false

fn get_contour_def( &self, target_pt: Point, contour: &mut impl ToOutputArray ) -> Result<()>

Extracts optimal contour for the given target point on the image

Note: buildMap() must be called before this call

Parameters

targetPt: The target point

contour:[out] The list of pixels which contains optimal path between the source and the target points of the image. Type is CV_32SC2 (compatible with std::vector<Point>)

backward: Flag to indicate reverse order of retrived pixels (use “true” value to fetch points from the target to the source point)

Note

This alternative version of IntelligentScissorsMBTraitConst::get_contour function uses the following default values for its arguments:

backward: false

Object Safety

This trait is not object safe.