Trait opencv::prelude::MatExprTraitConst

pub trait MatExprTraitConst {

    // Required method
    fn as_raw_MatExpr(&self) -> *const c_void;

    // Provided methods
    fn flags(&self) -> i32 { ... }
    fn a(&self) -> Mat { ... }
    fn b(&self) -> Mat { ... }
    fn c(&self) -> Mat { ... }
    fn alpha(&self) -> f64 { ... }
    fn beta(&self) -> f64 { ... }
    fn s(&self) -> Scalar { ... }
    fn to_mat(&self) -> Result<Mat> { ... }
    fn size(&self) -> Result<Size> { ... }
    fn typ(&self) -> Result<i32> { ... }
    fn row(&self, y: i32) -> Result<MatExpr> { ... }
    fn col(&self, x: i32) -> Result<MatExpr> { ... }
    fn diag(&self, d: i32) -> Result<MatExpr> { ... }
    fn diag_def(&self) -> Result<MatExpr> { ... }
    fn rowscols(
        &self,
        row_range: &impl RangeTraitConst,
        col_range: &impl RangeTraitConst
    ) -> Result<MatExpr> { ... }
    fn roi(&self, roi: Rect) -> Result<MatExpr> { ... }
    fn t(&self) -> Result<MatExpr> { ... }
    fn inv(&self, method: i32) -> Result<MatExpr> { ... }
    fn inv_def(&self) -> Result<MatExpr> { ... }
    fn mul_matexpr(
        &self,
        e: &impl MatExprTraitConst,
        scale: f64
    ) -> Result<MatExpr> { ... }
    fn mul_matexpr_def(&self, e: &impl MatExprTraitConst) -> Result<MatExpr> { ... }
    fn mul(&self, m: &impl MatTraitConst, scale: f64) -> Result<MatExpr> { ... }
    fn mul_def(&self, m: &impl MatTraitConst) -> Result<MatExpr> { ... }
    fn cross(&self, m: &impl MatTraitConst) -> Result<Mat> { ... }
    fn dot(&self, m: &impl MatTraitConst) -> Result<f64> { ... }

}

Constant methods for core::MatExpr

Required Methods

fn as_raw_MatExpr(&self) -> *const c_void

Provided Methods

fn flags(&self) -> i32

fn a(&self) -> Mat

fn b(&self) -> Mat

fn c(&self) -> Mat

fn alpha(&self) -> f64

fn beta(&self) -> f64

fn s(&self) -> Scalar

fn to_mat(&self) -> Result<Mat>

fn size(&self) -> Result<Size>

fn typ(&self) -> Result<i32>

fn row(&self, y: i32) -> Result<MatExpr>

fn col(&self, x: i32) -> Result<MatExpr>

fn diag(&self, d: i32) -> Result<MatExpr>

C++ default parameters

d: 0

fn diag_def(&self) -> Result<MatExpr>

Note

This alternative version of MatExprTraitConst::diag function uses the following default values for its arguments:

d: 0

fn rowscols( &self, row_range: &impl RangeTraitConst, col_range: &impl RangeTraitConst ) -> Result<MatExpr>

fn roi(&self, roi: Rect) -> Result<MatExpr>

fn t(&self) -> Result<MatExpr>

fn inv(&self, method: i32) -> Result<MatExpr>

C++ default parameters

method: DECOMP_LU

fn inv_def(&self) -> Result<MatExpr>

Note

This alternative version of MatExprTraitConst::inv function uses the following default values for its arguments:

method: DECOMP_LU

fn mul_matexpr(&self, e: &impl MatExprTraitConst, scale: f64) -> Result<MatExpr>

C++ default parameters

scale: 1

fn mul_matexpr_def(&self, e: &impl MatExprTraitConst) -> Result<MatExpr>

Note

This alternative version of MatExprTraitConst::mul_matexpr function uses the following default values for its arguments:

scale: 1

fn mul(&self, m: &impl MatTraitConst, scale: f64) -> Result<MatExpr>

C++ default parameters

scale: 1

fn mul_def(&self, m: &impl MatTraitConst) -> Result<MatExpr>

Note

This alternative version of MatExprTraitConst::mul function uses the following default values for its arguments:

scale: 1

fn cross(&self, m: &impl MatTraitConst) -> Result<Mat>

fn dot(&self, m: &impl MatTraitConst) -> Result<f64>

Object Safety

This trait is not object safe.