Trait opencv::prelude::ModifyInplace

pub trait ModifyInplace {
    // Required method
    unsafe fn modify_inplace<Res>(
        &mut self,
        f: impl FnOnce(&Self, &mut Self) -> Res,
    ) -> Res;
}

Required Methods

unsafe fn modify_inplace<Res>( &mut self, f: impl FnOnce(&Self, &mut Self) -> Res, ) -> Res

Helper function to call OpenCV functions that allow in-place modification of a Mat or another similar object. By passing a mutable reference to the Mat to this function your closure will get called with the read reference and a write references to the same Mat. This is of course unsafe as it breaks the Rust aliasing rules, but it might be useful for some performance sensitive operations. One example of an OpenCV function that allows such in-place modification is imgproc::threshold.

Safety

Caller must make sure that any functions called inside the closure can act on a Mat in-place.

Object Safety

This trait is not object safe.