Trait opencv::prelude::_InputArrayTraitConst

pub trait _InputArrayTraitConst {

    // Required method
    fn as_raw__InputArray(&self) -> *const c_void;

    // Provided methods
    fn get_mat(&self, idx: i32) -> Result<Mat> { ... }
    fn get_mat_def(&self) -> Result<Mat> { ... }
    fn get_mat_(&self, idx: i32) -> Result<Mat> { ... }
    fn get_mat__def(&self) -> Result<Mat> { ... }
    fn get_umat(&self, idx: i32) -> Result<UMat> { ... }
    fn get_umat_def(&self) -> Result<UMat> { ... }
    fn get_mat_vector(&self, mv: &mut Vector<Mat>) -> Result<()> { ... }
    fn get_umat_vector(&self, umv: &mut Vector<UMat>) -> Result<()> { ... }
    fn get_gpu_mat_vector(&self, gpumv: &mut Vector<GpuMat>) -> Result<()> { ... }
    fn get_gpu_mat(&self) -> Result<GpuMat> { ... }
    fn get_o_gl_buffer(&self) -> Result<Buffer> { ... }
    fn get_flags(&self) -> Result<i32> { ... }
    fn get_obj(&self) -> Result<*mut c_void> { ... }
    fn get_sz(&self) -> Result<Size> { ... }
    fn kind(&self) -> Result<_InputArray_KindFlag> { ... }
    fn dims(&self, i: i32) -> Result<i32> { ... }
    fn dims_def(&self) -> Result<i32> { ... }
    fn cols(&self, i: i32) -> Result<i32> { ... }
    fn cols_def(&self) -> Result<i32> { ... }
    fn rows(&self, i: i32) -> Result<i32> { ... }
    fn rows_def(&self) -> Result<i32> { ... }
    fn size(&self, i: i32) -> Result<Size> { ... }
    fn size_def(&self) -> Result<Size> { ... }
    fn sizend(&self, sz: &mut i32, i: i32) -> Result<i32> { ... }
    fn sizend_def(&self, sz: &mut i32) -> Result<i32> { ... }
    fn same_size(&self, arr: &impl ToInputArray) -> Result<bool> { ... }
    fn total(&self, i: i32) -> Result<size_t> { ... }
    fn total_def(&self) -> Result<size_t> { ... }
    fn typ(&self, i: i32) -> Result<i32> { ... }
    fn typ_def(&self) -> Result<i32> { ... }
    fn depth(&self, i: i32) -> Result<i32> { ... }
    fn depth_def(&self) -> Result<i32> { ... }
    fn channels(&self, i: i32) -> Result<i32> { ... }
    fn channels_def(&self) -> Result<i32> { ... }
    fn is_continuous(&self, i: i32) -> Result<bool> { ... }
    fn is_continuous_def(&self) -> Result<bool> { ... }
    fn is_submatrix(&self, i: i32) -> Result<bool> { ... }
    fn is_submatrix_def(&self) -> Result<bool> { ... }
    fn empty(&self) -> Result<bool> { ... }
    fn copy_to(&self, arr: &mut impl ToOutputArray) -> Result<()> { ... }
    fn copy_to_masked(
        &self,
        arr: &mut impl ToOutputArray,
        mask: &impl ToInputArray,
    ) -> Result<()> { ... }
    fn offset(&self, i: i32) -> Result<size_t> { ... }
    fn offset_def(&self) -> Result<size_t> { ... }
    fn step(&self, i: i32) -> Result<size_t> { ... }
    fn step_def(&self) -> Result<size_t> { ... }
    fn is_mat(&self) -> Result<bool> { ... }
    fn is_umat(&self) -> Result<bool> { ... }
    fn is_mat_vector(&self) -> Result<bool> { ... }
    fn is_umat_vector(&self) -> Result<bool> { ... }
    fn is_matx(&self) -> Result<bool> { ... }
    fn is_vector(&self) -> Result<bool> { ... }
    fn is_gpu_mat(&self) -> Result<bool> { ... }
    fn is_gpu_mat_vector(&self) -> Result<bool> { ... }

}

Constant methods for core::_InputArray

Required Methods

fn as_raw__InputArray(&self) -> *const c_void

Provided Methods

fn get_mat(&self, idx: i32) -> Result<Mat>

C++ default parameters

idx: -1

fn get_mat_def(&self) -> Result<Mat>

Note

This alternative version of _InputArrayTraitConst::get_mat function uses the following default values for its arguments:

idx: -1

fn get_mat_(&self, idx: i32) -> Result<Mat>

C++ default parameters

idx: -1

fn get_mat__def(&self) -> Result<Mat>

Note

This alternative version of InputArrayTraitConst::get_mat function uses the following default values for its arguments:

idx: -1

fn get_umat(&self, idx: i32) -> Result<UMat>

C++ default parameters

idx: -1

fn get_umat_def(&self) -> Result<UMat>

Note

This alternative version of _InputArrayTraitConst::get_umat function uses the following default values for its arguments:

idx: -1

fn get_mat_vector(&self, mv: &mut Vector<Mat>) -> Result<()>

fn get_umat_vector(&self, umv: &mut Vector<UMat>) -> Result<()>

fn get_gpu_mat_vector(&self, gpumv: &mut Vector<GpuMat>) -> Result<()>

fn get_gpu_mat(&self) -> Result<GpuMat>

fn get_o_gl_buffer(&self) -> Result<Buffer>

fn get_flags(&self) -> Result<i32>

fn get_obj(&self) -> Result<*mut c_void>

fn get_sz(&self) -> Result<Size>

fn kind(&self) -> Result<_InputArray_KindFlag>

fn dims(&self, i: i32) -> Result<i32>

C++ default parameters

i: -1

fn dims_def(&self) -> Result<i32>

Note

This alternative version of _InputArrayTraitConst::dims function uses the following default values for its arguments:

i: -1

fn cols(&self, i: i32) -> Result<i32>

C++ default parameters

i: -1

fn cols_def(&self) -> Result<i32>

Note

This alternative version of _InputArrayTraitConst::cols function uses the following default values for its arguments:

i: -1

fn rows(&self, i: i32) -> Result<i32>

C++ default parameters

i: -1

fn rows_def(&self) -> Result<i32>

Note

This alternative version of _InputArrayTraitConst::rows function uses the following default values for its arguments:

i: -1

fn size(&self, i: i32) -> Result<Size>

C++ default parameters

i: -1

fn size_def(&self) -> Result<Size>

Note

This alternative version of _InputArrayTraitConst::size function uses the following default values for its arguments:

i: -1

fn sizend(&self, sz: &mut i32, i: i32) -> Result<i32>

C++ default parameters

i: -1

fn sizend_def(&self, sz: &mut i32) -> Result<i32>

Note

This alternative version of _InputArrayTraitConst::sizend function uses the following default values for its arguments:

i: -1

fn same_size(&self, arr: &impl ToInputArray) -> Result<bool>

fn total(&self, i: i32) -> Result<size_t>

C++ default parameters

i: -1

fn total_def(&self) -> Result<size_t>

Note

This alternative version of _InputArrayTraitConst::total function uses the following default values for its arguments:

i: -1

fn typ(&self, i: i32) -> Result<i32>

C++ default parameters

i: -1

fn typ_def(&self) -> Result<i32>

Note

This alternative version of _InputArrayTraitConst::typ function uses the following default values for its arguments:

i: -1

fn depth(&self, i: i32) -> Result<i32>

C++ default parameters

i: -1

fn depth_def(&self) -> Result<i32>

Note

This alternative version of _InputArrayTraitConst::depth function uses the following default values for its arguments:

i: -1

fn channels(&self, i: i32) -> Result<i32>

C++ default parameters

i: -1

fn channels_def(&self) -> Result<i32>

Note

This alternative version of _InputArrayTraitConst::channels function uses the following default values for its arguments:

i: -1

fn is_continuous(&self, i: i32) -> Result<bool>

C++ default parameters

i: -1

fn is_continuous_def(&self) -> Result<bool>

Note

This alternative version of _InputArrayTraitConst::is_continuous function uses the following default values for its arguments:

i: -1

fn is_submatrix(&self, i: i32) -> Result<bool>

C++ default parameters

i: -1

fn is_submatrix_def(&self) -> Result<bool>

Note

This alternative version of _InputArrayTraitConst::is_submatrix function uses the following default values for its arguments:

i: -1

fn empty(&self) -> Result<bool>

fn copy_to(&self, arr: &mut impl ToOutputArray) -> Result<()>

fn copy_to_masked( &self, arr: &mut impl ToOutputArray, mask: &impl ToInputArray, ) -> Result<()>

fn offset(&self, i: i32) -> Result<size_t>

C++ default parameters

i: -1

fn offset_def(&self) -> Result<size_t>

Note

This alternative version of _InputArrayTraitConst::offset function uses the following default values for its arguments:

i: -1

fn step(&self, i: i32) -> Result<size_t>

C++ default parameters

i: -1

fn step_def(&self) -> Result<size_t>

Note

This alternative version of _InputArrayTraitConst::step function uses the following default values for its arguments:

i: -1

fn is_mat(&self) -> Result<bool>

fn is_umat(&self) -> Result<bool>

fn is_mat_vector(&self) -> Result<bool>

fn is_umat_vector(&self) -> Result<bool>

fn is_matx(&self) -> Result<bool>

fn is_vector(&self) -> Result<bool>

fn is_gpu_mat(&self) -> Result<bool>

fn is_gpu_mat_vector(&self) -> Result<bool>

Object Safety

This trait is not object safe.