Trait opencv::prelude::_OutputArrayTraitConst

pub trait _OutputArrayTraitConst: _InputArrayTraitConst {

    // Required method
    fn as_raw__OutputArray(&self) -> *const c_void;

    // Provided methods
    fn fixed_size(&self) -> Result<bool> { ... }
    fn fixed_type(&self) -> Result<bool> { ... }
    fn needed(&self) -> Result<bool> { ... }
    fn get_mat_ref(&self, i: i32) -> Result<Mat> { ... }
    fn get_mat_ref_def(&self) -> Result<Mat> { ... }
    fn get_umat_ref(&self, i: i32) -> Result<UMat> { ... }
    fn get_umat_ref_def(&self) -> Result<UMat> { ... }
    fn get_gpu_mat_ref(&self) -> Result<GpuMat> { ... }
    fn get_gpu_mat_vec_ref(&self) -> Result<Vector<GpuMat>> { ... }
    fn get_o_gl_buffer_ref(&self) -> Result<Buffer> { ... }
    fn get_host_mem_ref(&self) -> Result<HostMem> { ... }
    fn create_size(
        &self,
        sz: Size,
        typ: i32,
        i: i32,
        allow_transposed: bool,
        fixed_depth_mask: _OutputArray_DepthMask,
    ) -> Result<()> { ... }
    fn create_size_def(&self, sz: Size, typ: i32) -> Result<()> { ... }
    fn create(
        &self,
        rows: i32,
        cols: i32,
        typ: i32,
        i: i32,
        allow_transposed: bool,
        fixed_depth_mask: _OutputArray_DepthMask,
    ) -> Result<()> { ... }
    fn create_def(&self, rows: i32, cols: i32, typ: i32) -> Result<()> { ... }
    fn create_nd(
        &self,
        size: &[i32],
        typ: i32,
        i: i32,
        allow_transposed: bool,
        fixed_depth_mask: _OutputArray_DepthMask,
    ) -> Result<()> { ... }
    fn create_nd_def(&self, size: &[i32], typ: i32) -> Result<()> { ... }
    unsafe fn create_same_size(
        &self,
        arr: &impl ToInputArray,
        mtype: i32,
    ) -> Result<()> { ... }
    fn release(&self) -> Result<()> { ... }
    fn clear(&self) -> Result<()> { ... }
    fn set_to(
        &self,
        value: &impl ToInputArray,
        mask: &impl ToInputArray,
    ) -> Result<()> { ... }
    fn set_to_def(&self, value: &impl ToInputArray) -> Result<()> { ... }
    fn assign_umat(&self, u: &impl UMatTraitConst) -> Result<()> { ... }
    fn assign_mat(&self, m: &impl MatTraitConst) -> Result<()> { ... }
    fn assign_umat_vec(&self, v: &Vector<UMat>) -> Result<()> { ... }
    fn assign_mat_vec(&self, v: &Vector<Mat>) -> Result<()> { ... }
    fn move_umat(&self, u: &mut impl UMatTrait) -> Result<()> { ... }
    fn move_mat(&self, m: &mut impl MatTrait) -> Result<()> { ... }

}

Constant methods for core::_OutputArray

Required Methods

fn as_raw__OutputArray(&self) -> *const c_void

Provided Methods

fn fixed_size(&self) -> Result<bool>

fn fixed_type(&self) -> Result<bool>

fn needed(&self) -> Result<bool>

fn get_mat_ref(&self, i: i32) -> Result<Mat>

C++ default parameters

i: -1

fn get_mat_ref_def(&self) -> Result<Mat>

Note

This alternative version of _OutputArrayTraitConst::get_mat_ref function uses the following default values for its arguments:

i: -1

fn get_umat_ref(&self, i: i32) -> Result<UMat>

C++ default parameters

i: -1

fn get_umat_ref_def(&self) -> Result<UMat>

Note

This alternative version of _OutputArrayTraitConst::get_umat_ref function uses the following default values for its arguments:

i: -1

fn get_gpu_mat_ref(&self) -> Result<GpuMat>

fn get_gpu_mat_vec_ref(&self) -> Result<Vector<GpuMat>>

fn get_o_gl_buffer_ref(&self) -> Result<Buffer>

fn get_host_mem_ref(&self) -> Result<HostMem>

fn create_size( &self, sz: Size, typ: i32, i: i32, allow_transposed: bool, fixed_depth_mask: _OutputArray_DepthMask, ) -> Result<()>

C++ default parameters

i: -1

allow_transposed: false

fixed_depth_mask: static_cast<_OutputArray::DepthMask>(0)

fn create_size_def(&self, sz: Size, typ: i32) -> Result<()>

Note

This alternative version of _OutputArrayTraitConst::create_size function uses the following default values for its arguments:

i: -1

allow_transposed: false

fixed_depth_mask: static_cast<_OutputArray::DepthMask>(0)

fn create( &self, rows: i32, cols: i32, typ: i32, i: i32, allow_transposed: bool, fixed_depth_mask: _OutputArray_DepthMask, ) -> Result<()>

C++ default parameters

i: -1

allow_transposed: false

fixed_depth_mask: static_cast<_OutputArray::DepthMask>(0)

fn create_def(&self, rows: i32, cols: i32, typ: i32) -> Result<()>

Note

This alternative version of _OutputArrayTraitConst::create function uses the following default values for its arguments:

i: -1

allow_transposed: false

fixed_depth_mask: static_cast<_OutputArray::DepthMask>(0)

fn create_nd( &self, size: &[i32], typ: i32, i: i32, allow_transposed: bool, fixed_depth_mask: _OutputArray_DepthMask, ) -> Result<()>

C++ default parameters

i: -1

allow_transposed: false

fixed_depth_mask: static_cast<_OutputArray::DepthMask>(0)

fn create_nd_def(&self, size: &[i32], typ: i32) -> Result<()>

Note

This alternative version of _OutputArrayTraitConst::create_nd function uses the following default values for its arguments:

i: -1

allow_transposed: false

fixed_depth_mask: static_cast<_OutputArray::DepthMask>(0)

unsafe fn create_same_size( &self, arr: &impl ToInputArray, mtype: i32, ) -> Result<()>

fn release(&self) -> Result<()>

fn clear(&self) -> Result<()>

fn set_to( &self, value: &impl ToInputArray, mask: &impl ToInputArray, ) -> Result<()>

C++ default parameters

mask: _InputArray()

fn set_to_def(&self, value: &impl ToInputArray) -> Result<()>

Note

This alternative version of _OutputArrayTraitConst::set_to function uses the following default values for its arguments:

mask: _InputArray()

fn assign_umat(&self, u: &impl UMatTraitConst) -> Result<()>

fn assign_mat(&self, m: &impl MatTraitConst) -> Result<()>

fn assign_umat_vec(&self, v: &Vector<UMat>) -> Result<()>

fn assign_mat_vec(&self, v: &Vector<Mat>) -> Result<()>

fn move_umat(&self, u: &mut impl UMatTrait) -> Result<()>

fn move_mat(&self, m: &mut impl MatTrait) -> Result<()>

Object Safety

This trait is not object safe.