Module opencv::templ

Macros

callback_arg ğŸ”’

extern_container_arg ğŸ”’

input_array_arg ğŸ”’

input_output_array_arg ğŸ”’

output_array_arg ğŸ”’

return_receive ğŸ”’

return_send ğŸ”’

string_arg_output_receive ğŸ”’

string_arg_output_send ğŸ”’

string_array_arg ğŸ”’

string_array_arg_mut ğŸ”’

userdata_arg ğŸ”’

Functions

ocvrs_create_byte_string ğŸ”’ âš

The return type of this function goes into receive_byte_string

ocvrs_create_string ğŸ”’ âš

The return type of this function goes into receive_string

receive_stringâš

Used for both regular String and byte string (Vec<u8>)