Module opencv::video

Structs

BackgroundSubtractorGMG

BackgroundSubtractorMOG

BackgroundSubtractorMOG2

KalmanFilter

Constants

CV_LKFLOW_GET_MIN_EIGENVALS

CV_LKFLOW_INITIAL_GUESSES

CV_LKFLOW_PYR_A_READY

CV_LKFLOW_PYR_B_READY

OPTFLOW_FARNEBACK_GAUSSIAN

OPTFLOW_LK_GET_MIN_EIGENVALS

OPTFLOW_USE_INITIAL_FLOW

Traits

BackgroundSubtractor

DenseOpticalFlow

Functions

calc_global_orientation
computes the global orientation of the selected motion history image part

calc_motion_gradient
computes the motion gradient orientation image from the motion history image

estimate_rigid_transform
estimates the best-fit Euqcidean, similarity, affine or perspective transformation

init_module_video

new
computes dense optical flow using Simple Flow algorithm

new_sigmas

segment_motion