Struct opencv::video::BackgroundSubtractorMOG2

pub struct BackgroundSubtractorMOG2 {
    pub ptr: *mut c_void,
}

Fields

ptr

Methods

impl BackgroundSubtractorMOG2

fn as_raw_BackgroundSubtractorMOG2(&self) -> *mut c_void

impl BackgroundSubtractorMOG2

fn default() -> Result<BackgroundSubtractorMOG2, String>

the default constructor

fn new(history: i32, varThreshold: f32, bShadowDetection: bool) -> Result<BackgroundSubtractorMOG2, String>

the full constructor that takes the length of the history, the number of gaussian mixtures, the background ratio parameter and the noise strength

default value for arguments: - bShadowDetection: default true

fn get_background_image(&self, backgroundImage: &Mat) -> Result<(), String>

computes a background image which are the mean of all background gaussians

fn initialize(&mut self, frameSize: Size, frameType: i32) -> Result<(), String>

re-initiaization method