Struct opencv::video::KalmanFilter

pub struct KalmanFilter {
    pub ptr: *mut c_void,
}

Fields

ptr

Methods

impl KalmanFilter

fn as_raw_KalmanFilter(&self) -> *mut c_void

impl KalmanFilter

fn default() -> Result<KalmanFilter, String>

the default constructor

fn new(dynamParams: i32, measureParams: i32, controlParams: i32, _type: i32) -> Result<KalmanFilter, String>

the full constructor taking the dimensionality of the state, of the measurement and of the control vector

default value for arguments: - controlParams: default 0 - _type: default CV_32F

fn init(&mut self, dynamParams: i32, measureParams: i32, controlParams: i32, _type: i32) -> Result<(), String>

re-initializes Kalman filter. The previous content is destroyed.

default value for arguments: - controlParams: default 0 - _type: default CV_32F

fn predict(&mut self, control: &Mat) -> Result<Mat, String>

computes predicted state

default value for arguments: - control: default Mat()

fn correct(&mut self, measurement: &Mat) -> Result<Mat, String>

updates the predicted state from the measurement