Struct opencv::core::Mat_

pub struct Mat_<T> {
    inner: Mat,
    _type: PhantomData<T>,
}

docs.opencv.org

This struct is freely convertible into and from Mat using into and try_from methods. You might want to convert Mat to Mat_ before calling typed methods (like at, data_typed) when more performance is required because this way you will skip the data type checks.

Fields
inner: Mat_type: PhantomData<T>
Implementations

impl<T: DataType> Mat_<T>

pub fn into_untyped(self) -> Mat

pub fn as_untyped(&self) -> &Mat

pub fn as_raw_Mat_(&self) -> *const c_void

pub fn as_raw_mut_Mat_(&mut self) -> *mut c_void

pub fn at(&self, i0: i32) -> Result<&T>

See Mat::at

pub fn at_2d(&self, row: i32, col: i32) -> Result<&T>

See Mat::at_2d

pub fn at_3d(&self, i0: i32, i1: i32, i2: i32) -> Result<&T>

See Mat::at_3d

pub fn at_nd(&self, idx: &[i32]) -> Result<&T>

See Mat::at_nd

pub fn at_pt(&self, pt: Point) -> Result<&T>

See Mat::at_pt

pub fn at_row(&self, row: i32) -> Result<&[T]>

See [Mat::at_row]

pub fn at_mut(&mut self, i0: i32) -> Result<&mut T>

See Mat::at_mut

pub fn at_2d_mut(&mut self, row: i32, col: i32) -> Result<&mut T>

See Mat::at_2d_mut

pub fn at_3d_mut(&mut self, i0: i32, i1: i32, i2: i32) -> Result<&mut T>

See Mat::at_3d_mut

pub fn at_nd_mut(&mut self, idx: &[i32]) -> Result<&mut T>

See Mat::at_nd_mut

pub fn at_pt_mut(&mut self, pt: Point) -> Result<&mut T>

See Mat::at_pt_mut

pub fn at_row_mut(&mut self, row: i32) -> Result<&mut [T]>

See [Mat::at_row_mut]

pub fn data_typed(&self) -> Result<&[T]>

See [Mat::data_typed]

pub fn data_typed_mut(&mut self) -> Result<&mut [T]>

See [Mat::data_typed_mut]