Struct opencv::core::Vector

pub struct Vector<T>
where
    Self: VectorExtern<T>,
{
    ptr: *mut c_void,
    _d: PhantomData<T>,
}

Wrapper for C++ std::vector

Fields
ptr: *mut c_void_d: PhantomData<T>
Implementations

impl<T> Vector<T>
where Self: VectorExtern<T>,

pub fn new() -> Self

Create a new Vector

pub fn with_capacity(capacity: size_t) -> Self

Create a Vector with pre-defined capacity

pub fn from_iter<'a>( s: impl IntoIterator<Item = <T as OpenCVType<'a>>::Arg> ) -> Self
where T: for<'t> OpenCVType<'t>,

Create a Vector from iterator

pub fn from_slice(s: &[T]) -> Self
where Self: VectorExternCopyNonBool<T>,

Create a Vector from slice, the element type needs to be Copy (and not bool)

pub fn from_elem<'a>(elem: <T as OpenCVType<'a>>::Arg, n: size_t) -> Self
where T: for<'t> OpenCVType<'t>, <T as OpenCVType<'a>>::Arg: Clone,

pub fn len(&self) -> size_t

Return Vector length

pub fn is_empty(&self) -> bool

Return true if Vector is empty

pub fn capacity(&self) -> size_t

Return Vector current capacity

pub fn shrink_to_fit(&mut self)

Free extra capacity

pub fn reserve(&mut self, additional: size_t)

Reserve capacity for additional new elements

pub fn clear(&mut self)

Remove all elements

pub fn remove(&mut self, index: size_t) -> Result<()>

Remove the element at the specified index

pub fn swap(&mut self, index1: size_t, index2: size_t) -> Result<()>

Swap 2 elements in the Vector

pub fn push(&mut self, val: <T as OpenCVType<'_>>::Arg)
where T: for<'t> OpenCVType<'t>,

Add new element

pub(crate) fn push_owned(&mut self, val: T)
where T: OpenCVIntoExternContainer,

pub fn insert( &mut self, index: size_t, val: <T as OpenCVType<'_>>::Arg ) -> Result<()>
where T: for<'t> OpenCVType<'t>,

Insert a new element at the specified index

pub fn set( &mut self, index: size_t, val: <T as OpenCVType<'_>>::Arg ) -> Result<()>
where T: for<'t> OpenCVType<'t>,

Set element at the specified index

pub unsafe fn set_unchecked( &mut self, index: size_t, val: <T as OpenCVType<'_>>::Arg )
where T: for<'t> OpenCVType<'t>,

Same as set() but without bounds checking

Safety

Caller must ensure that the specified index is within the Vector bounds

pub fn get(&self, index: size_t) -> Result<T>
where T: OpenCVFromExtern,

Get element at the specified index

pub unsafe fn get_unchecked(&self, index: size_t) -> T
where T: OpenCVFromExtern,

Same as get() but without bounds checking

Safety

Caller must ensure that the specified index is within the Vector bounds

pub fn iter(&self) -> VectorRefIterator<'_, T>

pub fn as_slice(&self) -> &[T]
where Self: VectorExternCopyNonBool<T>,

Return slice to the elements of the array.

This method is only available for OpenCV types that are Copy, except for bool because bool is handled in a special way on the C++ side.

pub fn as_mut_slice(&mut self) -> &mut [T]
where Self: VectorExternCopyNonBool<T>,

Return mutable slice to the elements of the array.

This method is only available for OpenCV types that are Copy, except for bool because bool is handled in a special way on the C++ side.

impl<'b, T: Boxed> Vector<BoxedRef<'b, T>>
where Self: VectorExtern<BoxedRef<'b, T>>, Vector<T>: VectorExtern<T>,

pub(crate) fn as_non_ref_vec(&self) -> &Vector<T>

Transmutes a &Vector<BoxedRef<T>> into a &Vector<T>. This is safe as BoxedRef is a transparent wrapper around T, but it breaks the lifetime guards imposed by BoxedRef, so this is a crate private function.

impl Vector<DMatch>

pub fn as_raw_VectorOfDMatch( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfDMatch( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<GpuMat>

pub fn as_raw_VectorOfGpuMat( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfGpuMat( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<KeyPoint>

pub fn as_raw_VectorOfKeyPoint( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfKeyPoint( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<Mat>

pub fn as_raw_VectorOfMat( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfMat( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<PlatformInfo>

pub fn as_raw_VectorOfPlatformInfo( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfPlatformInfo( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<Point>

pub fn as_raw_VectorOfPoint( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfPoint( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<Point2d>

pub fn as_raw_VectorOfPoint2d( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfPoint2d( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<Point2f>

pub fn as_raw_VectorOfPoint2f( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfPoint2f( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<Range>

pub fn as_raw_VectorOfRange( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfRange( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<Rect>

pub fn as_raw_VectorOfRect( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfRect( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<Rect2d>

pub fn as_raw_VectorOfRect2d( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfRect2d( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<RotatedRect>

pub fn as_raw_VectorOfRotatedRect( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfRotatedRect( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<String>

pub fn as_raw_VectorOfString( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfString( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<UMat>

pub fn as_raw_VectorOfUMat( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfUMat( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<Vec2d>

pub fn as_raw_VectorOfVec2d( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfVec2d( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<Vec2f>

pub fn as_raw_VectorOfVec2f( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfVec2f( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<Vec3d>

pub fn as_raw_VectorOfVec3d( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfVec3d( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<Vec3f>

pub fn as_raw_VectorOfVec3f( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfVec3f( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<Vec4f>

pub fn as_raw_VectorOfVec4f( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfVec4f( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<Vec4i>

pub fn as_raw_VectorOfVec4i( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfVec4i( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<Vec6f>

pub fn as_raw_VectorOfVec6f( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfVec6f( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<Vector<Mat>>

pub fn as_raw_VectorOfVectorOfMat( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfVectorOfMat( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<Vector<Point>>

pub fn as_raw_VectorOfVectorOfPoint( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfVectorOfPoint( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<Vector<Point2f>>

pub fn as_raw_VectorOfVectorOfPoint2f( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfVectorOfPoint2f( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<Vector<Range>>

pub fn as_raw_VectorOfVectorOfRange( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfVectorOfRange( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<Vector<f32>>

pub fn as_raw_VectorOfVectorOff32( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfVectorOff32( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<Vector<i32>>

pub fn as_raw_VectorOfVectorOfi32( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfVectorOfi32( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<bool>

pub fn as_raw_VectorOfbool( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfbool( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<c_char>

pub fn as_raw_VectorOfc_char( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfc_char( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<f32>

pub fn as_raw_VectorOff32( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOff32( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<f64>

pub fn as_raw_VectorOff64( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOff64( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<i32>

pub fn as_raw_VectorOfi32( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfi32( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<i8>

pub fn as_raw_VectorOfi8( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfi8( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<size_t>

pub fn as_raw_VectorOfsize_t( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfsize_t( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<u8>

pub fn as_raw_VectorOfu8( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfu8( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<MatShape>

pub fn as_raw_VectorOfMatShape( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfMatShape( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<Ptr<BackendNode>>

pub fn as_raw_VectorOfPtrOfBackendNode( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfPtrOfBackendNode( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<Ptr<BackendWrapper>>

pub fn as_raw_VectorOfPtrOfBackendWrapper( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfPtrOfBackendWrapper( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<Ptr<Layer>>

pub fn as_raw_VectorOfPtrOfLayer( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfPtrOfLayer( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<Target>

pub fn as_raw_VectorOfTarget( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfTarget( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<Tuple<(Backend, Target)>>

pub fn as_raw_VectorOfTupleOfBackend_Target( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfTupleOfBackend_Target( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<Vector<MatShape>>

pub fn as_raw_VectorOfVectorOfMatShape( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfVectorOfMatShape( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<VideoCapture>

pub fn as_raw_VectorOfVideoCapture( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfVideoCapture( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut

impl Vector<VideoCaptureAPIs>

pub fn as_raw_VectorOfVideoCaptureAPIs( &self ) -> <Self as OpenCVTypeExternContainer>::ExternSend

pub fn as_raw_mut_VectorOfVideoCaptureAPIs( &mut self ) -> <Self as OpenCVTypeExternContainer>::ExternSendMut