Trait opencv::hub_prelude::AlgorithmTraitConst

pub trait AlgorithmTraitConst {
    // Required method
    fn as_raw_Algorithm(&self) -> *const c_void;

    // Provided methods
    fn write(&self, fs: &mut impl FileStorageTrait) -> Result<()> { ... }
    fn write_with_name(&self, fs: &Ptr<FileStorage>, name: &str) -> Result<()> { ... }
    fn write_with_name_def(&self, fs: &Ptr<FileStorage>) -> Result<()> { ... }
    fn empty(&self) -> Result<bool> { ... }
    fn save(&self, filename: &str) -> Result<()> { ... }
    fn get_default_name(&self) -> Result<String> { ... }
}

Constant methods for core::Algorithm

Required Methods

fn as_raw_Algorithm(&self) -> *const c_void

Provided Methods

fn write(&self, fs: &mut impl FileStorageTrait) -> Result<()>

Stores algorithm parameters in a file storage

fn write_with_name(&self, fs: &Ptr<FileStorage>, name: &str) -> Result<()>

simplified API for language bindings Stores algorithm parameters in a file storage

Overloaded parameters

C++ default parameters

name: String()

fn write_with_name_def(&self, fs: &Ptr<FileStorage>) -> Result<()>

simplified API for language bindings @overload

Note

This alternative version of AlgorithmTraitConst::write_with_name function uses the following default values for its arguments:

name: String()

fn empty(&self) -> Result<bool>

Returns true if the Algorithm is empty (e.g. in the very beginning or after unsuccessful read

fn save(&self, filename: &str) -> Result<()>

Saves the algorithm to a file. In order to make this method work, the derived class must implement Algorithm::write(FileStorage& fs).

fn get_default_name(&self) -> Result<String>

Returns the algorithm string identifier. This string is used as top level xml/yml node tag when the object is saved to a file or string.

Object Safety

This trait is not object safe.