Struct opencv::opencv::hub::core::Device

pub struct Device {
    ptr: *mut c_void,
}

Fields
ptr: *mut c_void
Implementations

impl Device

pub fn default() -> Device

pub unsafe fn new(d: *mut c_void) -> Result<Device>

pub fn copy(d: &impl DeviceTraitConst) -> Result<Device>

pub fn copy_mut(d: Device) -> Device

pub fn get_default() -> Result<Device>

pub unsafe fn from_handle(d: *mut c_void) -> Result<Device>

Parameters

d: OpenCL handle (cl_device_id). clRetainDevice() is called on success.

Note: Ownership of the passed device is passed to OpenCV on success. The caller should additionally call clRetainDevice on it if it intends to continue using the device.