Trait opencv::opencv::hub::core::GpuMatTrait

pub trait GpuMatTrait: GpuMatTraitConst {

    // Required method
    fn as_raw_mut_GpuMat(&mut self) -> *mut c_void;

    // Provided methods
    fn set_flags(&mut self, val: i32) { ... }
    fn set_rows(&mut self, val: i32) { ... }
    fn set_cols(&mut self, val: i32) { ... }
    fn set_step(&mut self, val: size_t) { ... }
    fn data_mut(&mut self) -> *mut u8 { ... }
    unsafe fn set_data(&mut self, val: *const u8) { ... }
    fn refcount_mut(&mut self) -> *mut i32 { ... }
    unsafe fn set_refcount(&mut self, val: *const i32) { ... }
    fn datastart_mut(&mut self) -> *mut u8 { ... }
    unsafe fn set_datastart(&mut self, val: *const u8) { ... }
    fn allocator(&mut self) -> AbstractRefMut<'_, GpuMat_Allocator> { ... }
    unsafe fn set_allocator(&mut self, val: &mut impl GpuMat_AllocatorTrait) { ... }
    fn set(&mut self, m: &GpuMat) -> Result<()> { ... }
    fn create(&mut self, rows: i32, cols: i32, typ: i32) -> Result<()> { ... }
    fn create_1(&mut self, size: Size, typ: i32) -> Result<()> { ... }
    fn release(&mut self) -> Result<()> { ... }
    fn swap(&mut self, mat: &mut GpuMat) -> Result<()> { ... }
    fn upload(&mut self, arr: &impl ToInputArray) -> Result<()> { ... }
    fn upload_async(
        &mut self,
        arr: &impl ToInputArray,
        stream: &mut Stream
    ) -> Result<()> { ... }
    fn set_to(&mut self, s: Scalar) -> Result<GpuMat> { ... }
    fn set_to_1(&mut self, s: Scalar, stream: &mut Stream) -> Result<GpuMat> { ... }
    fn set_to_2(
        &mut self,
        s: Scalar,
        mask: &impl ToInputArray
    ) -> Result<GpuMat> { ... }
    fn set_to_3(
        &mut self,
        s: Scalar,
        mask: &impl ToInputArray,
        stream: &mut Stream
    ) -> Result<GpuMat> { ... }
    fn ptr_mut(&mut self, y: i32) -> Result<*mut u8> { ... }
    fn ptr_mut_def(&mut self) -> Result<*mut u8> { ... }
    fn adjust_roi(
        &mut self,
        dtop: i32,
        dbottom: i32,
        dleft: i32,
        dright: i32
    ) -> Result<GpuMat> { ... }
    fn update_continuity_flag(&mut self) -> Result<()> { ... }

}

Mutable methods for core::GpuMat

Required Methods

fn as_raw_mut_GpuMat(&mut self) -> *mut c_void

Provided Methods

fn set_flags(&mut self, val: i32)

! includes several bit-fields:

the magic signature

continuity flag

depth

number of channels

fn set_rows(&mut self, val: i32)

the number of rows and columns

fn set_cols(&mut self, val: i32)

the number of rows and columns

fn set_step(&mut self, val: size_t)

a distance between successive rows in bytes; includes the gap if any

fn data_mut(&mut self) -> *mut u8

pointer to the data

unsafe fn set_data(&mut self, val: *const u8)

pointer to the data

fn refcount_mut(&mut self) -> *mut i32

pointer to the reference counter; when GpuMat points to user-allocated data, the pointer is NULL

unsafe fn set_refcount(&mut self, val: *const i32)

pointer to the reference counter; when GpuMat points to user-allocated data, the pointer is NULL

fn datastart_mut(&mut self) -> *mut u8

helper fields used in locateROI and adjustROI

unsafe fn set_datastart(&mut self, val: *const u8)

helper fields used in locateROI and adjustROI

fn allocator(&mut self) -> AbstractRefMut<'_, GpuMat_Allocator>

allocator

unsafe fn set_allocator(&mut self, val: &mut impl GpuMat_AllocatorTrait)

allocator

fn set(&mut self, m: &GpuMat) -> Result<()>

assignment operators

fn create(&mut self, rows: i32, cols: i32, typ: i32) -> Result<()>

allocates new GpuMat data unless the GpuMat already has specified size and type

fn create_1(&mut self, size: Size, typ: i32) -> Result<()>

fn release(&mut self) -> Result<()>

decreases reference counter, deallocate the data when reference counter reaches 0

fn swap(&mut self, mat: &mut GpuMat) -> Result<()>

swaps with other smart pointer

fn upload(&mut self, arr: &impl ToInputArray) -> Result<()>

Performs data upload to GpuMat (Blocking call)

This function copies data from host memory to device memory. As being a blocking call, it is guaranteed that the copy operation is finished when this function returns.

fn upload_async( &mut self, arr: &impl ToInputArray, stream: &mut Stream ) -> Result<()>

Performs data upload to GpuMat (Non-Blocking call)

This function copies data from host memory to device memory. As being a non-blocking call, this function may return even if the copy operation is not finished.

The copy operation may be overlapped with operations in other non-default streams if \p stream is not the default stream and \p dst is HostMem allocated with HostMem::PAGE_LOCKED option.

fn set_to(&mut self, s: Scalar) -> Result<GpuMat>

sets some of the GpuMat elements to s (Blocking call)

fn set_to_1(&mut self, s: Scalar, stream: &mut Stream) -> Result<GpuMat>

sets some of the GpuMat elements to s (Non-Blocking call)

fn set_to_2(&mut self, s: Scalar, mask: &impl ToInputArray) -> Result<GpuMat>

sets some of the GpuMat elements to s, according to the mask (Blocking call)

fn set_to_3( &mut self, s: Scalar, mask: &impl ToInputArray, stream: &mut Stream ) -> Result<GpuMat>

sets some of the GpuMat elements to s, according to the mask (Non-Blocking call)

fn ptr_mut(&mut self, y: i32) -> Result<*mut u8>

returns pointer to y-th row

C++ default parameters

y: 0

fn ptr_mut_def(&mut self) -> Result<*mut u8>

returns pointer to y-th row

Note

This alternative version of GpuMatTrait::ptr_mut function uses the following default values for its arguments:

y: 0

fn adjust_roi( &mut self, dtop: i32, dbottom: i32, dleft: i32, dright: i32 ) -> Result<GpuMat>

moves/resizes the current GpuMat ROI inside the parent GpuMat

fn update_continuity_flag(&mut self) -> Result<()>

internal use method: updates the continuity flag

Object Safety

This trait is not object safe.