Trait opencv::opencv::hub::core::MatOpTraitConst

pub trait MatOpTraitConst {

    // Required method
    fn as_raw_MatOp(&self) -> *const c_void;

    // Provided methods
    fn element_wise(&self, expr: &MatExpr) -> Result<bool> { ... }
    fn assign(&self, expr: &MatExpr, m: &mut Mat, typ: i32) -> Result<()> { ... }
    fn assign_def(&self, expr: &MatExpr, m: &mut Mat) -> Result<()> { ... }
    fn roi(
        &self,
        expr: &MatExpr,
        row_range: &Range,
        col_range: &Range,
        res: &mut MatExpr
    ) -> Result<()> { ... }
    fn diag(&self, expr: &MatExpr, d: i32, res: &mut MatExpr) -> Result<()> { ... }
    fn aug_assign_add(&self, expr: &MatExpr, m: &mut Mat) -> Result<()> { ... }
    fn aug_assign_subtract(&self, expr: &MatExpr, m: &mut Mat) -> Result<()> { ... }
    fn aug_assign_multiply(&self, expr: &MatExpr, m: &mut Mat) -> Result<()> { ... }
    fn aug_assign_divide(&self, expr: &MatExpr, m: &mut Mat) -> Result<()> { ... }
    fn aug_assign_and(&self, expr: &MatExpr, m: &mut Mat) -> Result<()> { ... }
    fn aug_assign_or(&self, expr: &MatExpr, m: &mut Mat) -> Result<()> { ... }
    fn aug_assign_xor(&self, expr: &MatExpr, m: &mut Mat) -> Result<()> { ... }
    fn add(
        &self,
        expr1: &MatExpr,
        expr2: &MatExpr,
        res: &mut MatExpr
    ) -> Result<()> { ... }
    fn add_scalar(
        &self,
        expr1: &MatExpr,
        s: Scalar,
        res: &mut MatExpr
    ) -> Result<()> { ... }
    fn subtract(
        &self,
        expr1: &MatExpr,
        expr2: &MatExpr,
        res: &mut MatExpr
    ) -> Result<()> { ... }
    fn subtract_scalar(
        &self,
        s: Scalar,
        expr: &MatExpr,
        res: &mut MatExpr
    ) -> Result<()> { ... }
    fn multiply(
        &self,
        expr1: &MatExpr,
        expr2: &MatExpr,
        res: &mut MatExpr,
        scale: f64
    ) -> Result<()> { ... }
    fn multiply_def(
        &self,
        expr1: &MatExpr,
        expr2: &MatExpr,
        res: &mut MatExpr
    ) -> Result<()> { ... }
    fn multiply_f64(
        &self,
        expr1: &MatExpr,
        s: f64,
        res: &mut MatExpr
    ) -> Result<()> { ... }
    fn divide(
        &self,
        expr1: &MatExpr,
        expr2: &MatExpr,
        res: &mut MatExpr,
        scale: f64
    ) -> Result<()> { ... }
    fn divide_def(
        &self,
        expr1: &MatExpr,
        expr2: &MatExpr,
        res: &mut MatExpr
    ) -> Result<()> { ... }
    fn divide_f64(
        &self,
        s: f64,
        expr: &MatExpr,
        res: &mut MatExpr
    ) -> Result<()> { ... }
    fn abs(&self, expr: &MatExpr, res: &mut MatExpr) -> Result<()> { ... }
    fn transpose(&self, expr: &MatExpr, res: &mut MatExpr) -> Result<()> { ... }
    fn matmul(
        &self,
        expr1: &MatExpr,
        expr2: &MatExpr,
        res: &mut MatExpr
    ) -> Result<()> { ... }
    fn invert(
        &self,
        expr: &MatExpr,
        method: i32,
        res: &mut MatExpr
    ) -> Result<()> { ... }
    fn size(&self, expr: &MatExpr) -> Result<Size> { ... }
    fn typ(&self, expr: &MatExpr) -> Result<i32> { ... }

}

Constant methods for core::MatOp

Required Methods

fn as_raw_MatOp(&self) -> *const c_void

Provided Methods

fn element_wise(&self, expr: &MatExpr) -> Result<bool>

fn assign(&self, expr: &MatExpr, m: &mut Mat, typ: i32) -> Result<()>

C++ default parameters

typ: -1

fn assign_def(&self, expr: &MatExpr, m: &mut Mat) -> Result<()>

Note

This alternative version of MatOpTraitConst::assign function uses the following default values for its arguments:

typ: -1

fn roi( &self, expr: &MatExpr, row_range: &Range, col_range: &Range, res: &mut MatExpr ) -> Result<()>

fn diag(&self, expr: &MatExpr, d: i32, res: &mut MatExpr) -> Result<()>

fn aug_assign_add(&self, expr: &MatExpr, m: &mut Mat) -> Result<()>

fn aug_assign_subtract(&self, expr: &MatExpr, m: &mut Mat) -> Result<()>

fn aug_assign_multiply(&self, expr: &MatExpr, m: &mut Mat) -> Result<()>

fn aug_assign_divide(&self, expr: &MatExpr, m: &mut Mat) -> Result<()>

fn aug_assign_and(&self, expr: &MatExpr, m: &mut Mat) -> Result<()>

fn aug_assign_or(&self, expr: &MatExpr, m: &mut Mat) -> Result<()>

fn aug_assign_xor(&self, expr: &MatExpr, m: &mut Mat) -> Result<()>

fn add(&self, expr1: &MatExpr, expr2: &MatExpr, res: &mut MatExpr) -> Result<()>

fn add_scalar( &self, expr1: &MatExpr, s: Scalar, res: &mut MatExpr ) -> Result<()>

fn subtract( &self, expr1: &MatExpr, expr2: &MatExpr, res: &mut MatExpr ) -> Result<()>

fn subtract_scalar( &self, s: Scalar, expr: &MatExpr, res: &mut MatExpr ) -> Result<()>

fn multiply( &self, expr1: &MatExpr, expr2: &MatExpr, res: &mut MatExpr, scale: f64 ) -> Result<()>

C++ default parameters

scale: 1

fn multiply_def( &self, expr1: &MatExpr, expr2: &MatExpr, res: &mut MatExpr ) -> Result<()>

Note

This alternative version of MatOpTraitConst::multiply function uses the following default values for its arguments:

scale: 1

fn multiply_f64(&self, expr1: &MatExpr, s: f64, res: &mut MatExpr) -> Result<()>

fn divide( &self, expr1: &MatExpr, expr2: &MatExpr, res: &mut MatExpr, scale: f64 ) -> Result<()>

C++ default parameters

scale: 1

fn divide_def( &self, expr1: &MatExpr, expr2: &MatExpr, res: &mut MatExpr ) -> Result<()>

Note

This alternative version of MatOpTraitConst::divide function uses the following default values for its arguments:

scale: 1

fn divide_f64(&self, s: f64, expr: &MatExpr, res: &mut MatExpr) -> Result<()>

fn abs(&self, expr: &MatExpr, res: &mut MatExpr) -> Result<()>

fn transpose(&self, expr: &MatExpr, res: &mut MatExpr) -> Result<()>

fn matmul( &self, expr1: &MatExpr, expr2: &MatExpr, res: &mut MatExpr ) -> Result<()>

fn invert(&self, expr: &MatExpr, method: i32, res: &mut MatExpr) -> Result<()>

fn size(&self, expr: &MatExpr) -> Result<Size>

fn typ(&self, expr: &MatExpr) -> Result<i32>