Struct opencv::opencv::hub::imgproc::LineIterator

pub struct LineIterator {
    ptr: *mut c_void,
}

Line iterator

The class is used to iterate over all the pixels on the raster line segment connecting two specified points.

The class LineIterator is used to get each pixel of a raster line. It can be treated as versatile implementation of the Bresenham algorithm where you can stop at each pixel and do some extra processing, for example, grab pixel values along the line or draw a line with an effect (for example, with XOR operation).

The number of pixels along the line is stored in LineIterator::count. The method LineIterator::pos returns the current position in the image:

// grabs pixels along the line (pt1, pt2)
// from 8-bit 3-channel image to the buffer
LineIterator it(img, pt1, pt2, 8);
LineIterator it2 = it;
vector<Vec3b> buf(it.count);
 
for(int i = 0; i < it.count; i++, ++it)
   buf[i] = *(const Vec3b*)*it;
 
// alternative way of iterating through the line
for(int i = 0; i < it2.count; i++, ++it2)
{
   Vec3b val = img.at<Vec3b>(it2.pos());
   CV_Assert(buf[i] == val);
}

Fields
ptr: *mut c_void
Implementations

impl LineIterator

pub fn new( img: &impl MatTraitConst, pt1: Point, pt2: Point, connectivity: i32, left_to_right: bool ) -> Result<LineIterator>

initializes the iterator

creates iterators for the line connecting pt1 and pt2 the line will be clipped on the image boundaries the line is 8-connected or 4-connected If leftToRight=true, then the iteration is always done from the left-most point to the right most, not to depend on the ordering of pt1 and pt2 parameters;

C++ default parameters

connectivity: 8

left_to_right: false

pub fn new_def( img: &impl MatTraitConst, pt1: Point, pt2: Point ) -> Result<LineIterator>

initializes the iterator

creates iterators for the line connecting pt1 and pt2 the line will be clipped on the image boundaries the line is 8-connected or 4-connected If leftToRight=true, then the iteration is always done from the left-most point to the right most, not to depend on the ordering of pt1 and pt2 parameters;

Note

This alternative version of [new] function uses the following default values for its arguments:

connectivity: 8

left_to_right: false

pub fn new_1( pt1: Point, pt2: Point, connectivity: i32, left_to_right: bool ) -> Result<LineIterator>

C++ default parameters

connectivity: 8

left_to_right: false

pub fn new_def_1(pt1: Point, pt2: Point) -> Result<LineIterator>

Note

This alternative version of [new] function uses the following default values for its arguments:

connectivity: 8

left_to_right: false

pub fn new_2( bounding_area_size: Size, pt1: Point, pt2: Point, connectivity: i32, left_to_right: bool ) -> Result<LineIterator>

C++ default parameters

connectivity: 8

left_to_right: false

pub fn new_def_2( bounding_area_size: Size, pt1: Point, pt2: Point ) -> Result<LineIterator>

Note

This alternative version of [new] function uses the following default values for its arguments:

connectivity: 8

left_to_right: false

pub fn new_3( bounding_area_rect: Rect, pt1: Point, pt2: Point, connectivity: i32, left_to_right: bool ) -> Result<LineIterator>

C++ default parameters

connectivity: 8

left_to_right: false

pub fn new_def_3( bounding_area_rect: Rect, pt1: Point, pt2: Point ) -> Result<LineIterator>

Note

This alternative version of [new] function uses the following default values for its arguments:

connectivity: 8

left_to_right: false